% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testLocalVsGlobal.R
\name{aggregateLocalVsGlobal}
\alias{aggregateLocalVsGlobal}
\title{Calculate a protein-level significance from peptide level differential association tests}
\usage{
aggregateLocalVsGlobal(tests)
}
\arguments{
\item{tests}{data.table, containing peptide level p-values.
Produced by \code{testLocalVsGlobal}.
A featureVals table can be produced with \code{extractFeatureVals}.}
}
\value{
A data.table containing the differential testing results for every protein.
}
\description{
Calculate a protein-level significance from peptide level differential association tests
}
