% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDifferentialExperssion.R
\name{aggregatePeptideTests}
\alias{aggregatePeptideTests}
\title{Calculate a protein-level significance from peptide level differential expression tests}
\usage{
aggregatePeptideTests(tests)
}
\arguments{
\item{tests}{data.table, containing peptide level p-values.
Produced by \code{testDifferentialExpression} with option \code{level}="peptide".
A featureVals table can be produced with \code{extractFeatureVals}.}
}
\value{
A data.table containing the differential testing results for every protein.
}
\description{
Calculate a protein-level significance from peptide level differential expression tests
}

