% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDifferentialExperssion.R
\name{aggregateProteinTests}
\alias{aggregateProteinTests}
\title{Calculate a complex-level significance from protein level differential expression tests}
\usage{
aggregateProteinTests(tests)
}
\arguments{
\item{tests}{data.table, containing protein level p-values.
Produced by \code{testDifferentialExpression} with option \code{level}="protein".
A featureVals table can be produced with \code{extractFeatureVals}.}
}
\value{
A data.table containing the differential testing results for every complex.
}
\description{
Calculate a complex-level significance from protein level differential expression tests
}
