% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFractions.R
\name{annotateFractionIdCounts}
\alias{annotateFractionIdCounts}
\title{Calculates and annotates the number of ids in every fraction}
\usage{
annotateFractionIdCounts(traces)
}
\description{
Calculates and annotates the number of ids in every fraction
}

