% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFractions.R
\name{annotateFractionIntensity}
\alias{annotateFractionIntensity}
\title{Calculates and annotates the total intensity in every fraction}
\usage{
annotateFractionIntensity(traces)
}
\description{
Calculates and annotates the total intensity in every fraction
}
