% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFractions.R
\name{annotateFractionMissingValues}
\alias{annotateFractionMissingValues}
\title{Calculates and annotates the number of missing values in every fraction}
\usage{
annotateFractionMissingValues(traces, bound_left = 2, bound_right = 2,
  consider_borders = T)
}
\description{
Calculates and annotates the number of missing values in every fraction
}

