% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFractions.R
\name{annotateFractions}
\alias{annotateFractions}
\title{Add summary statistics to the fractionAnnotation object}
\usage{
annotateFractions(traces)
}
\arguments{
\item{traces}{Object of class traces.}
}
\value{
Object of class traces with annotation of fractions.
}
\description{
Add summary statistics to the fractionAnnotation object
}
\examples{
inputTraces <- examplePeptideTraces
exampleTracesAnnotatedFractions <- annotateFractions(inputTraces)
exampleTracesAnnotatedFractions$fraction_annotation
}

