% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteogenomics.R
\name{annotateGenomicCoordinates}
\alias{annotateGenomicCoordinates}
\alias{annotateGenomicCoordinates.tracesList}
\alias{annotateGenomicCoordinates.traces}
\title{Annotate the genomic position of peptides in traces}
\usage{
annotateGenomicCoordinates(
  traces,
  db = system.file("extdata/EnsDb.Hsapiens.v86.sqlite", package = "EnsDb.Hsapiens.v86"),
  proteinIdCol = "LeadingEnsemblProtein",
  verbose = F
)

\method{annotateGenomicCoordinates}{tracesList}(
  tracesList,
  db = system.file("extdata/EnsDb.Hsapiens.v86.sqlite", package = "EnsDb.Hsapiens.v86"),
  proteinIdCol = "LeadingEnsemblProtein",
  verbose = F
)

\method{annotateGenomicCoordinates}{traces}(
  traces,
  db = system.file("extdata/EnsDb.Hsapiens.v86.sqlite", package = "EnsDb.Hsapiens.v86"),
  proteinIdCol = "LeadingEnsemblProtein",
  verbose = F
)
}
\arguments{
\item{traces}{Object of class traces with annotated relative peptide positions
(see =annotateRelativePepPos=).}

\item{db}{An EnsDb object containing the CDS mappings of the proteins in question.}

\item{proteinIdCol}{character, the name of the column in =trace_annotation= containing
the ids of the proteins the relative peptide positions have been calculated for.}

\item{verbose}{boolean whether to print warning messages.}
}
\value{
Object of class traces or tracesList with an additional item 'genomic_coord'
containing a GRangesList with one entry for each peptide containing the genomic coordinates.
}
\description{
uses simple grep in the protein that was found in the mapping table.
}
\section{Methods (by class)}{
\itemize{
\item \code{tracesList}: Annotate the relative Position of all peptides in a tracesList object.

\item \code{traces}: Annotate the relative Position of all peptides in a single traces object.
}}

