% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteogenomics.R
\name{annotateLeadingIsoform}
\alias{annotateLeadingIsoform}
\alias{annotateLeadingIsoform.tracesList}
\alias{annotateLeadingIsoform.traces}
\title{Find the leading Isoform for each peptide}
\usage{
annotateLeadingIsoform(
  traces,
  isoform_col = "isoform_id",
  output_col = "LeadingIsoform"
)

\method{annotateLeadingIsoform}{tracesList}(
  tracesList,
  isoform_col = "isoform_id",
  output_col = "LeadingIsoform"
)

\method{annotateLeadingIsoform}{traces}(
  traces,
  isoform_col = "isoform_id",
  output_col = "LeadingIsoform"
)
}
\arguments{
\item{traces}{Object of class tracesList}

\item{isoform_col}{Character, name of the column in the =trace_annotation= table
containing the annotation of all found isoforms for that peptide. Note:
This function assumes isoform annotation in the format:
'3/ISOFORM1/ISOFORM2/ISOFORM3'}

\item{ouput_col}{character, name of the column that contains the leading isoform}
}
\value{
A traces object with an additional column in =trace_annotation= containing
the leading isoform.
}
\description{
Find the leading Isoform for each peptide
}
\details{
This function goes through every gene and applies a maximum parsimony
approach to find the leading Isoform for each peptide. For ever gene
(defined either by the 'gene_id' or the 'protein_id' column)
and counts the number of occurences for every isoform. Peptides are then annotated
with the fewest possible number of isoforms that are sufficient to explain all peptides.
In case of ties one isoform is randomly selected.
}
\section{Methods (by class)}{
\itemize{
\item \code{tracesList}: Annotate the Leading Isoform in a tracesList object

\item \code{traces}: Annotate the Leading Isoform in a single traces object
}}

