% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMassDistribution.R
\name{annotateMassDistribution}
\alias{annotateMassDistribution}
\alias{annotateMassDistribution.traces}
\alias{annotateMassDistribution.tracesList}
\title{annotateMassDistribution}
\usage{
annotateMassDistribution(traces)

\method{annotateMassDistribution}{traces}(traces)

\method{annotateMassDistribution}{tracesList}(tracesList)
}
\arguments{
\item{traces}{An object of type traces. These should be the raw, unprocessed traces.}
}
\description{
Estimates fraction of assembled vs. monomeric mass per id.
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Estimates fraction of assembled vs. monomeric mass per id.

\item \code{tracesList}: Estimates fraction of assembled vs. monomeric mass per id.
}}

