% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateOutlierFractions.R
\name{annotateOutlierFractions}
\alias{annotateOutlierFractions}
\alias{annotateOutlierFractions.traces}
\alias{annotateOutlierFractions.tracesList}
\title{Annotate outlier fractions}
\usage{
annotateOutlierFractions(traces, excludeBoarder = TRUE, plot = TRUE,
  PDF = FALSE, name = "OutlierFractions")

\method{annotateOutlierFractions}{traces}(traces, excludeBoarder = TRUE,
  plot = TRUE, PDF = FALSE, name = "OutlierFractions", ...)

\method{annotateOutlierFractions}{tracesList}(traces, excludeBoarder = TRUE,
  plot = TRUE, PDF = FALSE, name = "OutlierFractions", ...)
}
\arguments{
\item{traces}{Object of class traces.}

\item{excludeBoarder}{If boarder fractions should be excluded as outliers, default is TRUE.}
}
\value{
Object of class traces with annotation of fractions and outliers.
}
\description{
Annotate outlier fractions
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Detect outlier fractions in single traces object

\item \code{tracesList}: Detect outlier fractions in multiple traces objects
}}

\examples{
inputTraces <- examplePeptideTraces
exampleTracesAnnotatedOutliers <- annotateOutlierFractions(inputTraces)
exampleTracesAnnotatedOutliers$fraction_annotation
}
