% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeRedundantPeptides.R
\name{annotatePeptideSequences}
\alias{annotatePeptideSequences}
\alias{annotatePeptideSequences.traces}
\alias{annotatePeptideSequences.tracesList}
\title{Annotate peptides with sequence information}
\usage{
annotatePeptideSequences(traces, fasta_file)

\method{annotatePeptideSequences}{traces}(traces, fasta_file)

\method{annotatePeptideSequences}{tracesList}(traces, fasta_file)
}
\arguments{
\item{traces}{An object of type traces, trace_type must be peptide.}

\item{fasta_file}{path to a fasta file.}
}
\value{
An object of type traces, trace_type is peptide.
}
\description{
Annotate peptides with sequence information.
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Annotate peptides with sequence information

\item \code{tracesList}: Annotate peptides with sequence information
}}

