% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{annotateProteoformsAcrossConditions}
\alias{annotateProteoformsAcrossConditions}
\title{Annotate tracesList of multiple conditions with proteoforms that were
assigned for a combined traces object.}
\usage{
annotateProteoformsAcrossConditions(tracesList, combinedTraces)
}
\arguments{
\item{tracesList}{Object of class tracesList.}

\item{combinedTraces}{Object of class traces with annotated proteoforms.}
}
\value{
Object of class tracesList witha nnotated proteoforms.
}
\description{
Annotate tracesList of multiple conditions with proteoforms that were
assigned for a combined traces object.
}
