% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteogenomics.R
\name{annotateRelativePepPos}
\alias{annotateRelativePepPos}
\alias{annotateRelativePepPos.tracesList}
\alias{annotateRelativePepPos.traces}
\title{Align peptides to reference proteins to get the relative position}
\usage{
annotateRelativePepPos(
  traces,
  mappingTable,
  multimatch = c("first", "skip"),
  verbose = F
)

\method{annotateRelativePepPos}{tracesList}(
  tracesList,
  mappingTable,
  multimatch = c("first", "skip"),
  verbose = F
)

\method{annotateRelativePepPos}{traces}(
  traces,
  mappingTable,
  multimatch = c("first", "skip"),
  verbose = F
)
}
\arguments{
\item{traces}{Object of class traces with annotated leading isoform (see =annotateLeadingIsoform=)}

\item{mappingTable}{data.table containing all sequences that were in the proteomics search space.
Must have the following columns: 'sequence', 'header'
'header' must be of the format: 'sp|IsoformId|etc'. The function will extract the isoform id
and match it to the leading isoform.}

\item{multimatch}{character, Skip or choose the first sequence if multiple id's match.}

\item{verbose}{boolean, whether to be verbose.}
}
\value{
Object of class traces with annotated relative positions of each peptide.
Two columns are added to =trace_annotation=:
'PeptidePositionStart' the position of the first amino acid in the protein.
'PeptidePositionEnd' the position of the last amino acid in the protein.
}
\description{
Align peptides to reference proteins to get the relative position
}
\section{Methods (by class)}{
\itemize{
\item \code{tracesList}: Annotate the relative Position of all peptides in a tracesList object.

\item \code{traces}: Annotate the relative Position of all peptides in a single traces object.
}}

