% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateTraces.R
\name{annotateTraces}
\alias{annotateTraces}
\title{Import peptide profiles from an OpenSWATH experiment.}
\usage{
annotateTraces(traces, annotation_table)
}
\arguments{
\item{traces.obj}{An object of type \code{traces.obj}.}

\item{annotation.table}{path to tab-separated .txt file for annotation of the traces}
}
\value{
traces.obj An object of type \code{traces.obj}.
}
\description{
Import peptide profiles from an OpenSWATH experiment.
}

