% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{annotateTracesWithProteoformClusters}
\alias{annotateTracesWithProteoformClusters}
\alias{annotateTracesWithProteoformClusters.traces}
\alias{annotateTracesWithProteoformClusters.tracesList}
\title{Annotate traces with cluster information}
\usage{
annotateTracesWithProteoformClusters(
  traces,
  cluster_annotation,
  FDR_cutoff = 0.05
)

\method{annotateTracesWithProteoformClusters}{traces}(
  traces,
  cluster_annotation,
  FDR_cutoff = 0.05
)

\method{annotateTracesWithProteoformClusters}{tracesList}(
  traces,
  cluster_annotation,
  FDR_cutoff = 0.05
)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{cluster_annotation}{data.table generated by cutClusters.}

\item{FDR_cutoff}{FDR cutoff for proteoform significance. Numeric between 0 and 1.
Default is 0.05 (5%).}
}
\value{
Traces with annotated proteoform clusters.
}
\description{
Annotate traces with cluster information
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Annotate traces with cluster information

\item \code{tracesList}: Annotate traces with cluster information
}}

