% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypotheses.R
\name{appendBinaryInteractions}
\alias{appendBinaryInteractions}
\title{appendBinaryInteractions (deprecated)}
\usage{
appendBinaryInteractions(complex, hypothesis, binary_interactions)
}
\arguments{
\item{complex}{character string with complex_id}

\item{hypothesis}{data.table in the format of complex hypotheses}

\item{binary_interactions}{with binary interactions between "a" and "b"}
}
\value{
data.table with binary interactions between "a" and "b" appended by current complex interactions
}
\description{
Append binary interaction table with additional complexes
}
