% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQvalues.R
\name{appendSecondaryComplexFeatures}
\alias{appendSecondaryComplexFeatures}
\title{Append secondary features to scored filtered primary feature table}
\usage{
appendSecondaryComplexFeatures(scoredPrimaryFeatures, allFeatures,
  peakCorr_cutoff = 0.5)
}
\arguments{
\item{scoredPrimaryFeatures}{data.table output from scoreFeatures function}

\item{allFeatures}{data.table, output from findComplexFeatures or findProteinFeatures function including primary and secondary features}

\item{peakCorr_cutoff}{numeric, minimum required peak orrelation for a secondary feature. Between 0 and 1, default = 0.05}
}
\value{
data.table with scored and filtered features
}
\description{
Append secondary features to scored filtered primary feature table
}

