% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQvalues.R
\name{calculateCoelutionScore}
\alias{calculateCoelutionScore}
\title{Calculate coelution score for all detected features.}
\usage{
calculateCoelutionScore(features)
}
\arguments{
\item{features}{data.table with complex or protein features}
}
\value{
data.table with complex or protein features including an extra coelution_score column
}
\description{
Calculate coelution score for all detected features.
}
