% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFeatureCorrelation.R
\name{calculateFeatureCorrelation}
\alias{calculateFeatureCorrelation}
\title{A helper function to calculate correlation within peak boundaries after Feature
detection with sliding window.}
\usage{
calculateFeatureCorrelation(trace.mat, features.obj, toTable = FALSE)
}
\arguments{
\item{toTable}{wether the output should be in table format, or in the original \code{complexFeatures} format.}

\item{traces.obj}{An object of type \code{traces.obj}.}

\item{features}{An object of type \code{complexFeatures} as generated by the
findComplexFeatures function.}
}
\value{
An List containing a features table. Every Feature now has an added peak_corr and area value.
}
\description{
A helper function to calculate correlation within peak boundaries after Feature
detection with sliding window.
}
\details{
The mean correlation of all detected subunits and total peak area is calculated.
         This is a helper function for FindComplexFeatures.
}
