% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{calculateGeneCorrMatrices}
\alias{calculateGeneCorrMatrices}
\title{Calculate pairwise correlations of peptides within all genes}
\usage{
calculateGeneCorrMatrices(
  traces,
  method = c("pearson", "kendall", "spearman"),
  logtransform = FALSE
)
}
\arguments{
\item{traces}{Object of class traces.}

\item{method}{Which correlation metric to use.}

\item{logtransform}{Logical, whether to log transform intensities}
}
\value{
List of matrices with pairwise correllations per protein
}
\description{
Calculate pairwise correlations of peptides within all genes
}
