% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{calculateMinCorr.traces}
\alias{calculateMinCorr.traces}
\title{#' @describeIn calculateMinCorr Calculate minimum correlation of each peptide
#' to all of its sibling peptides
#' @export}
\usage{
\method{calculateMinCorr}{traces}(
  traces,
  FeatureFinding = NULL,
  plot = FALSE,
  PDF = FALSE,
  name = "minCorrHist",
  ...
)
}
\description{
#' @describeIn calculateMinCorr Calculate minimum correlation of each peptide
#' to all of its sibling peptides
#' @export
}
