% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypotheses.R
\name{calculatePathlength}
\alias{calculatePathlength}
\title{Calculate pathlength from binary interactions}
\usage{
calculatePathlength(binaryInteractionTable)
}
\arguments{
\item{binaryInteractionTable}{data.table with binary interactions between "a" and "b" as returned from generateBinaryNetwork()}
}
\value{
data.table with three columns containing all possible combinations of proteins:
\itemize{
  \item x,y: Protein ids of the input proteins (all possible combinations)
  \item dist: distance (shortest path between interactors)
        as determined by the input binary interaction table
}
}
\description{
Calculate shortest paths between all members of a binary interaction network.
This function creates a table that contains all possible combinations of 2 from
all input proteins (binomial coefficient (length(proteins) over 2)).
}
\examples{
## Load example data
complexHypotheses <- exampleComplexHypotheses

## Generate the binary network
binaryInteractions <- generateBinaryNetwork(complex_hypotheses = complexHypotheses)

## Calculate the path lengths
shortestPaths <- calculatePathlength(binaryInteractions)

}

