% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{calculateProteoformScore}
\alias{calculateProteoformScore}
\title{Calculates a score for proteoforms based on the difference of within cluster distances
and between cluster distances.}
\usage{
calculateProteoformScore(traces, summary_fun = "mean")
}
\arguments{
\item{traces}{Object of class traces with `geneCorrMatrices` and clustered peptides.}

\item{summary_fun}{the function to use for distance aggregation. Default: mean}
}
\value{
traces object with proteoform_score column in trace_annotation.
}
\description{
Calculates a score for proteoforms based on the difference of within cluster distances
and between cluster distances.
}
