% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQvalues.R
\name{calculateQvalue}
\alias{calculateQvalue}
\title{Calculate q-values from coelution score for all detected features.}
\usage{
calculateQvalue(
  features,
  lambda = 0.5,
  plot = TRUE,
  PDF = FALSE,
  name = "q_value_stats"
)
}
\arguments{
\item{features}{data.table with complex or protein features}

\item{lambda}{numeric value between 0 and 1, default = 0.5}

\item{plot}{logical, default = TRUE}

\item{PDF}{logical, defalt = FALSE}

\item{name}{character strimg specifying pdf name, default = "q_value_stats"}
}
\value{
An object of type \code{collapsedComplexFeatures} that is a list
       containing the following:
       \itemize{
         \item \code{feature} data.table containing complex feature candidates in the following format:
          \itemize{
          \item \code{subgroup} The protein_ids of the feature separated by semi-colons.
          \item \code{left_sw} The left boundary of the sliding-window feature.
          \item \code{right_sw} The right boundary of the sliding-window feature.
          \item \code{score} The intra-sliding-window-feature correlation.
          }
       }
}
\description{
Calculate q-values from coelution score for all detected features.
}
