% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRepPepCorr.R
\name{calculateRepPepCorr}
\alias{calculateRepPepCorr}
\title{Calculates the correlation of every petide between replicates}
\usage{
calculateRepPepCorr(traces, design_matrix, compare_within = NULL, add = T,
  plot = F, PDF = F, name = "ReplicatePeptideCorrelationDensity")
}
\arguments{
\item{traces}{Object of class \code{tracesList}.}

\item{design_matrix}{data.table, describing the \code{tracesList} object.}

\item{compare_within}{Character string, name of the column defining the conditions.}

\item{add}{Logical,add repPepCorr to \code{tracesList} object or return a table.}
}
\value{
Object of class traces with additional repPepCorr column or data.table
contaioning repPepCors.
}
\description{
Average pairwise pearson correlation of the intensity vectors between
replicates.
}
