% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinSequenceCoverage.R
\name{calculateSequenceCoverage}
\alias{calculateSequenceCoverage}
\title{Calculate the peptide Sequence coverage in every protein by realignment to the protein sequence}
\usage{
calculateSequenceCoverage(pepTraces, proteinFasta)
}
\arguments{
\item{pepTraces}{peptideTraces object to calculate the sequence coverage from}

\item{proteinFasta}{Path to a fasta file with Protein Sequences and headers that contain
the Protein names in the protein_id column of your peptideTraces object. Usually this will be the
fasta file used for library generation in the proteomics search.}
}
\value{
A \code{data.table} containing measured proteins, the length of their AA sequence and the fraction
of the sequence covered by peptides
}
\description{
Calculate the peptide Sequence coverage in every protein by realignment to the protein sequence
}

