% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSibPepCorr.R
\name{calculateSibPepCorr}
\alias{calculateSibPepCorr}
\title{calculateSibPepCorr}
\usage{
calculateSibPepCorr(traces, plot = TRUE, PDF = FALSE)
}
\arguments{
\item{traces}{An object of type \code{traces.obj}.}

\item{plot}{logical TRUE or FALSE}

\item{PDF}{logical TRUE or FALSE}
}
\value{
An object of type \code{traces.obj}.
}
\description{
Calculate sibling peptide correlation in traces.object of type peptide.
}

