% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateMW.R
\name{calibrateMW}
\alias{calibrateMW}
\title{calibrate Molecular Weight}
\usage{
calibrateMW(calibration_table, plot = TRUE, PDF = FALSE)
}
\arguments{
\item{calibration_table}{A table with protein standards, file or R data.table.
Columns:
\item std_weights_kDa = Numeric, the MW of the standard proteins.
\item std_elu_fractions = Numeric, the fraction numbers where these standard proteins eluted.
For an example see \code{exampleCalibrationTable}.}

\item{plot}{Logical, whether to plot calibration curve. Defaults to \code{TRUE}.}

\item{PDF}{Logical, whether to produce a PDF file in the working directory.
Defaults to \code{FALSE}.}
}
\value{
List of two functions: MWtoFraction and FractionToMW.
}
\description{
Establish connectivity between fraction number and apparent
Molecular Weight(MW) (in kDa). Many fractionation based PCP expeiments will have a set
of standard proteins with known MW spiked in the samples. These reference
proteins can be used in this function to calibrate the corresponding MW of every fraction.
While this is not neccesary for downstream analysis, it is recommended if standard proteins
are available, since it facilitates the process of drawing biological conclusions.
}
\details{
The function uses a standard \code{lm()} of the form \code{log2(MW) ~ fraction}.
Therefore, the fractionation technique is required to have a log-linear relationship between
fraction and molecular weight (True for e.g. standard SEC methods).
The estimated slope and intersect of the model are then used for the returned calibration functions.
}
\examples{

## Load example data
standardProteinTable <- exampleCalibrationTable
## Regress the standard proteins against their elution fractions
calibrationFunctions <- calibrateMW(standardProteinTable,
                                       plot = TRUE,
                                       PDF = FALSE)
# The provided plot can be used for quality assessment. The points represent the
# standard proteins and should agree with the model (solid line))

## The molecular weight (in kDa) of an arbitrary fraction can now be calculated.
calibrationFunctions$FractionToMW(14)

## Arbitrary MWs (in kDa) can also be converted to fraction numbers
calibrationFunctions$MWtoFraction(3020) # i.e. 3020 kDa

}
