% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseHypothesisWideFeatures.R
\name{callapseByUniqueFeatureGroups}
\alias{callapseByUniqueFeatureGroups}
\title{Collapse features by unique feature groups}
\usage{
callapseByUniqueFeatureGroups(feature_table, rm_decoys = TRUE)
}
\arguments{
\item{feature_table}{data.table containing filtered complex feature results with an extra column with a unique_feature_identifier.}

\item{rm_decoys}{Logical, whether to remove decoys before collapsing, default = \code{TRUE}.}
}
\value{
data.table containing collapsed complex features.
}
\description{
Collapse features with the same unique_feature_identifier as determined by \code{\link{getUniqueFeatureGroups}}.
}
\examples{
## run unique feature annotation
complexFeaturesUnique <- getUniqueFeatureGroups(exampleComplexFeatures,
                                                rt_height=2,
                                                distance_cutoff=1,
                                                plot=TRUE)
## Collapse across unique features
complexFeaturesCollapsed <- callapseByUniqueFeatureGroups(complexFeaturesUnique,
                                                          rm_decoys=TRUE)
## Inspect output
complexFeaturesCollapsed

}

