% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{clusterPeptides}
\alias{clusterPeptides}
\alias{clusterPeptides.traces}
\alias{clusterPeptides.tracesList}
\title{Cluster peptides of gene to unique proteoforms}
\usage{
clusterPeptides(
  traces,
  method = c("complete", "single", "average"),
  plot = FALSE,
  PDF = FALSE,
  name = "hclust",
  ...
)

\method{clusterPeptides}{traces}(
  traces,
  method = c("complete", "single", "average"),
  plot = FALSE,
  PDF = FALSE,
  name = "hclust",
  ...
)

\method{clusterPeptides}{tracesList}(
  traces,
  method = c("complete", "single", "average"),
  plot = FALSE,
  PDF = FALSE,
  name = "hclust",
  ...
)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{method}{Character string defining method for clustering
Default is "complete".}

\item{plot}{logical,wether to print SibPepCorr density plot to R console.
Deafult is \code{FALSE}.}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file.
Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if
'\code{PDF=TRUE}.PDF file is saved under name.pdf. Default is "hclust".}
}
\value{
List including hclust objects for each protein.

Object of class traces with proteoform annotation
}
\description{
Cluster peptides of gene to unique proteoforms
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Cluster peptides of gene to unique proteoforms

\item \code{tracesList}: Cluster peptides of gene to unique proteoforms
}}

