% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{clustPepMultiCond}
\alias{clustPepMultiCond}
\title{Cluster peptides of gene to unique proteoforms across conditions}
\usage{
clustPepMultiCond(tracesList, tracesListRaw = NULL, method = c("complete",
  "single"), clusterH = 0.5, clusterN = NULL, index = "silhouette",
  plot = FALSE, PDF = FALSE, name = "hclustMultiCond", ...)
}
\arguments{
\item{tracesList}{Object of class tracesList.}

\item{tracesListRaw}{Object of class tracesList. Default is NULL.}

\item{method}{Character string defining method for clustering
Default is "complete".}

\item{clusterH}{Numeric Cluster hight for cutree. Default is 0.5.}

\item{clusterN}{Integer Number of clusters. Default is NULL.}

\item{index}{Character string for method of cluster number estimation
Default is "silhouette".}

\item{plot}{logical,wether to print SibPepCorr density plot to R console.
Deafult is \code{FALSE}.}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file.
Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if
'\code{PDF=TRUE}.PDF file is saved under name.pdf.
Default is "hclustMultiCond".}
}
\value{
Object of class tracesList with proteoform annotation
across conditions
}
\description{
Cluster peptides of gene to unique proteoforms across conditions
}

