% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseComplexFearures.R
\name{collapseComplexFeatures}
\alias{collapseComplexFeatures}
\title{Collapse detected complex fetures by common apex and subunits.}
\usage{
collapseComplexFeatures(complexFeature = complexFeaturesPP, rt_height,
  collapse_method)
}
\arguments{
\item{complexFeature}{Object of class ComplexFeaturePP}

\item{rt_height}{Height at which to cur apex clustering tree.}
}
\value{
An object of type \code{collapsedComplexFeatures} that is a list
       containing the following:
       \itemize{
         \item \code{feature} data.table containing complex feature candidates in the following format:
          \itemize{
          \item \code{subgroup} The protein_ids of the feature separated by semi-colons.
          \item \code{left_sw} The left boundary of the sliding-window feature.
          \item \code{right_sw} The right boundary of the sliding-window feature.
          \item \code{score} The intra-sliding-window-feature correlation.
          }
       }
}
\description{
Collapse detected complex fetures by common apex and subunits.
}
