% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseHypothesisWideFeatures.R
\name{collapseHypothesisWideFeatures}
\alias{collapseHypothesisWideFeatures}
\title{collapseHypothesisWideFeatures}
\usage{
collapseHypothesisWideFeatures(complexFeature = complexFeatures,
  rt_height = 2, distance_cutoff = 1)
}
\arguments{
\item{complexFeature}{data.table containing filtered complex feature results.}

\item{rt_height}{numeric rt difference for clustering along the sec dimension, default = 2}

\item{distance_cutoff}{numeric distnce of detected subunit overlap (0=identical, 1=subcomplexes, 2=no shared subunits), default=1}
}
\value{
data.table containing collapsed complex feature results
}
\description{
collapseHypothesisWideFeatures.
}
