% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{combineTracesMutiCond}
\alias{combineTracesMutiCond}
\title{Combine tracesList across conditions to one traces object with all
the traces of all conditions added as additional fractions for
peptide clustering across conditions.}
\usage{
combineTracesMutiCond(tracesList)
}
\arguments{
\item{tracesList}{Object of class tracesList.}
}
\value{
Object of class traces
}
\description{
Combine tracesList across conditions to one traces object with all
the traces of all conditions added as additional fractions for
peptide clustering across conditions.
}

