% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypothesesHelpers.R
\name{complexClustering}
\alias{complexClustering}
\title{complexClustering}
\usage{
complexClustering(complexFeatures, dist_mat, clust_method = "single")
}
\arguments{
\item{dist_mat}{distance matrix from getDistanceMatrix}

\item{clust_method}{character string specifying method for hclust, default is "single"}

\item{complexFeature}{data.table containing filtered complex feature results.}
}
\value{
cluster object
}
\description{
complexClustering.
}

