% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{computeSilhouettes}
\alias{computeSilhouettes}
\title{Compute the silhouette score of every peptide}
\usage{
computeSilhouettes(tree, dist, clusters = NULL, h = 0.3)
}
\arguments{
\item{tree}{hclust object}

\item{dist}{distances that the tree is based on}

\item{clusters}{vector of cluster assignments (as produced by cutree)}

\item{h}{int, height at which to cut the tree if clusters are not provided}
}
\description{
Compute the silhouette score of every peptide
}
