% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSECexplorer.R
\name{convertIDs}
\alias{convertIDs}
\title{Convert any input ids to uniprotkb accession ids.}
\usage{
convertIDs(ids = ids, type = "UNIPROTKB", traces.obj = traces.obj)
}
\arguments{
\item{ids}{A character vector containing names or identifiers.}

\item{type}{A character string specifying the type of ids.}

\item{traces.obj}{An object of type \code{traces.obj}.}
}
\value{
A data.table containing the following information:
       \itemize{
          \item \code{not_defined} A character vector containing all input ids that are not defined in the current version of the package.
          \item \code{no_ms_signal} A chaacter vector of all input ids that do not have an MS signal in the current SEC experiment.
          \item \code{protein_ids}  A chaacter vector of all uniprotkb accession ids used for further analysis.
          \item \code{mapping_table} A data.table containing the following information:
          \itemize{
           \item \code{UNIPROTKB} Uniprotkb accession ids.
           \item \code{name} The combined name for the uniprotkb accession ids. This is \code{UNIPROTKB} | uniprot name.
           \item \code{input type} If the input type was not \code{UNIPROTKB} this contains the ids of the input type.
          }
       }
}
\description{
Convert any input ids to uniprotkb accession ids.
}
