% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSECToMW.R
\name{convertSECToMW}
\alias{convertSECToMW}
\title{Convert a fraction number to an estimated molecular weight.}
\usage{
convertSECToMW(sec)
}
\arguments{
\item{sec}{The fraction number.}
}
\value{
A molecular weight estimate.
}
\description{
Convert a fraction number to an estimated molecular weight.
}

