% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corumComplexHypothesesRedundant}
\alias{corumComplexHypothesesRedundant}
\title{Complex hypotheses from CORUM DB (redundant)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8923 rows and 3 columns.
}
\usage{
corumComplexHypothesesRedundant
}
\description{
Each row in this data.table corresponds to an association between a CORUM
complex id/name and a protein subunit.
The structure of this data.table is as follows:
\itemize{
 \item \code{complex_id} The numeric id of a CORUM complex.
 \item \code{complex_name} A character string of the complex name.
 \item \code{protein_id} The Uniprot identifier of the protein.
}
}
\keyword{datasets}
