% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelpers.R
\name{createGGplotColMap}
\alias{createGGplotColMap}
\title{Create a reproducible color map for a set of ids by ordering
them alphabetically}
\usage{
createGGplotColMap(ids)
}
\description{
Create a reproducible color map for a set of ids by ordering
them alphabetically
}
