% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGroupsFromNetworkTable.R
\name{createGroupsFromNetworkTable}
\alias{createGroupsFromNetworkTable}
\title{createGroupsFromNetworkTable create protein groups around network nodes 
with all proteins connected by 1 edge (first degree interactors)}
\usage{
createGroupsFromNetworkTable(interaction_table, n.cores = detectCores() - 1,
  id_column_1 = "UniprotA", id_column_2 = "UniprotB")
}
\arguments{
\item{interaction_table}{Interaction network in edge table format (data.frame)}

\item{id_column_1}{Name of the column containing protein identifier #1}

\item{id_column_2}{Name of the column containing protein identifier #2}
}
\value{
A data.frame containing group name (seed node name) and its first line interactors
e.g. for use as "complex hypotheses" for
}
\description{
createGroupsFromNetworkTable create protein groups around network nodes 
with all proteins connected by 1 edge (first degree interactors)
}
