% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{cutClusters}
\alias{cutClusters}
\title{Cut clusters from clusterPeptides at a specified hight cutoff.}
\usage{
cutClusters(clusterList, clusterH = 0.5)
}
\arguments{
\item{clusterList}{List of hclust objects generated by clusterPeptides.}

\item{clusterH}{Cluster hight cutoff.}
}
\value{
data.table with information of assigned clusters for each protein and peptide.
}
\description{
Cut clusters from clusterPeptides at a specified hight cutoff.
}
