% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{cutClustersInN}
\alias{cutClustersInN}
\title{Cut clusters from clusterPeptides into N clusters.}
\usage{
cutClustersInN(clusterList, clusterN = 2, min_peptides_per_cluster = 2)
}
\arguments{
\item{clusterList}{List of hclust objects generated by clusterPeptides.}

\item{clusterN}{Number of clusters. Default is 2.}

\item{min_peptides_per_cluster}{Minimum number of peptides per cluster. Default is 2.}
}
\value{
data.table with information of assigned clusters for each protein and peptide.
}
\description{
Cut clusters from clusterPeptides into N clusters.
}
