% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{cutClustersInNreal}
\alias{cutClustersInNreal}
\title{Cut clusters from clusterPeptides into N clusters with more than 1 peptide.}
\usage{
cutClustersInNreal(traces, clusterN = 2, min_peptides_per_cluster = 2)
}
\arguments{
\item{traces}{Object of class traces with a member `peptideClustering` as produced by `clusterPeptides`.}

\item{clusterN}{How many clusters to cut the tree into. Default: 2.}

\item{min_peptides_per_cluster}{How many peptides per cluster.}
}
\value{
traces object with cluster annotations in trace_annotation
}
\description{
Cut clusters from clusterPeptides into N clusters with more than 1 peptide.
}
\details{
Cuts the hierarchical tree so that there are N clusters with at least the specified number of peptides.
If there are single peptides above the cutoff height they will get the label 100.
}
