% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypothesesHelpers.R
\name{.collapseWideHypothesis}
\alias{.collapseWideHypothesis}
\title{Collapse redundant hypotheses}
\usage{
.collapseWideHypothesis(hypothesis, redundancy_cutoff = 1)
}
\arguments{
\item{hypothesis}{data.table with complex hypotheses}

\item{redundancy_cutoff}{numeric maximum overlap distance between two hypotheses (0=identical,1=subset,between 1 and 2=some shared subunits, 2=no shared subunits), default=1}
}
\value{
data.table in the format of complex hypotheses
}
\description{
Remove redundancy in existing complex hypotheses, internal function
}
