% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{.tracesListTest}
\alias{.tracesListTest}
\title{Test if an object is of class tracesList.}
\usage{
.tracesListTest(tracesList, type = NULL, additionalItems = NULL)
}
\arguments{
\item{type}{Character string specifying whether a specific type of traces is required.}

\item{additionalItems}{Character string specifying additional entries that are required in the list.
The two options are "peptide" or "protein". Default is \code{NULL},
meaning that no specific type is required.}

\item{traces}{Object of class tracesList.}
}
\description{
Test if an object is of class tracesList.
}
