% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{.tracesTest}
\alias{.tracesTest}
\title{Test if an object is of class traces.}
\usage{
.tracesTest(traces, type = NULL)
}
\arguments{
\item{traces}{Object of class traces.}

\item{type}{Character string specifying whether a specific type of traces is required.
The two options are "peptide" or "protein". Default is \code{NULL},
meaning that no specific type is required.}
}
\description{
Test if an object is of class traces.
}
