% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEstimation.R
\name{estimate_error_decoy}
\alias{estimate_error_decoy}
\title{Estimate statistics based on decoy model.}
\usage{
estimate_error_decoy(complex_hypothesis, detected_complex_features)
}
\arguments{
\item{complex_hypothesis}{data.table containing complex hypotheses.}

\item{detected_complex_features}{data.table containing filtered complex feature results.}
}
\value{
List with stats
}
\description{
Estimate statistics based on decoy model.
}
