% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEstimation.R
\name{estimate_error_manual}
\alias{estimate_error_manual}
\title{Estimate statistics based on manual annotation.}
\usage{
estimate_error_manual(detected_features, manual_features)
}
\arguments{
\item{detected_features}{data.table containing filtered feature results.}

\item{manual_features}{data.table containing manual feature results.}
}
\value{
List with stats
}
\description{
Estimate statistics based on manual annotation.
}
