% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEstimation.R
\name{estimate_errors}
\alias{estimate_errors}
\title{Estimate statistics based on manual annotation.}
\usage{
estimate_errors(table)
}
\arguments{
\item{table}{data.table with complex features mapped to manual annotation.}
}
\value{
List with stats
}
\description{
Estimate statistics based on manual annotation.
}
