% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateComplexFeatureStoichiometry.R
\name{estimateComplexFeatureStoichiometry}
\alias{estimateComplexFeatureStoichiometry}
\title{A helper function to extend a list of complex features with additional
stoichiometry estimation.}
\usage{
estimateComplexFeatureStoichiometry(traces.obj, complexFeaturesPP)
}
\arguments{
\item{traces.obj}{An object of type \code{traces.obj}.}

\item{complexFeaturesPP}{An object of type \code{complexFeaturesPP} that is a list
containing the following: 
\itemize{
  \item \code{feature} data.table containing complex feature candidates in the following format:
   \itemize{
   \item \code{subgroup} The protein_ids of the feature separated by semi-colons.
   \item \code{left_sw} The left boundary of the sliding-window feature.
   \item \code{right_sw} The right boundary of the sliding-window feature.
   \item \code{score} The intra-sliding-window-feature correlation.
   \item \code{n_subunits} The number of protein_ids in the sliding-window feature.
   \item \code{apax} The apex of the selected peak by the peak-picker.
   \item \code{left_pp} The left boundary of the selected peak by the peak-picker.
   \item \code{right_pp} The right boundary of the selected peak by the peak-picker.
   \item \code{area} The area (entire complex) of the selected peak by the peak-picker.
   }
}}
}
\value{
An object of type \code{complexFeatureStoichiometries} that is a list
       containing the following: 
       \itemize{
         \item \code{feature} data.table containing complex feature candidates in the following format:
          \itemize{
          \item \code{left_sw} The left boundary of the sliding-window feature.
          \item \code{right_sw} The right boundary of the sliding-window feature.
          \item \code{score} The intra-sliding-window-feature correlation.
          \item \code{n_subunits} The number of protein_ids in the sliding-window feature.
          \item \code{apax} The apex of the selected peak by the peak-picker.
          \item \code{left_pp} The left boundary of the selected peak by the peak-picker.
          \item \code{right_pp} The right boundary of the selected peak by the peak-picker.
          \item \code{area} The area (entire complex) of the selected peak by the peak-picker.
          \item \code{id} The protein_ids of the feature separated by semi-colons.
          \item \code{total_intensity} The intensity of all protein_ids of the feature separated by semi-colons.
          \item \code{intensity_ratio} The intensity ratio of all protein_ids of the feature separated by semi-colons.
          \item \code{stoichiometry} The rounded \code{intensity_ratio} of all protein_ids of the feature separated by semi-colons.
          }
       }
}
\description{
A helper function to extend a list of complex features with additional
stoichiometry estimation.
}

