% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeaturesSW.R
\name{estimateComplexMass}
\alias{estimateComplexMass}
\title{A helper function to compute an estimate of the mass of a complex and
its stoichiometry.}
\usage{
estimateComplexMass(traces, left.boundary, right.boundary,
  subunit.total.intensities, subunit.mws, subunit.abundances)
}
\arguments{
\item{traces}{A long list style data.table holding the intensity
observations of each subunit. This data.table should have the format:
\itemize{
 \item \code{protein_id} The subunit identifier
 \item \code{fraction} Where the observation was made
 \item \code{intensity} The observed intensity
}}

\item{left.boundary}{The left feature boundary}

\item{right.boundary}{The right feature boundary}

\item{subunit.total.intensities}{A numeric vector with the total trace
intensities across all fractions.}

\item{subunit.mws}{A numeric vector with molecular weights of the subunits.}

\item{subunit.abundances}{A numeric vector with total abundances of the
subunits.}
}
\description{
A helper function to compute an estimate of the mass of a complex and
its stoichiometry.
}

