% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R, R/featureFDRestimation.R
\name{estimateDecoyFDR}
\alias{estimateDecoyFDR}
\alias{estimateDecoyFDR}
\title{Estimate decoy based featureFinding FDR.}
\usage{
estimateDecoyFDR(features, FFT = 1)

estimateDecoyFDR(features, FFT = 1)
}
\arguments{
\item{features}{data.table containing filtered complex feature results.}

\item{features}{data.table containing filtered complex feature results.}
}
\value{
List with stats

List with stats
}
\description{
Estimate FDR statistics based on a decoy model after feature finding.

Estimate FDR statistics based on a decoy model after feature finding.
}
\examples{

#-----------------------------
## Complex Features
#-----------------------------

## Load example data
exampleHypotheses <- generateComplexDecoys(target_hypotheses = exampleComplexHypotheses,
                                           dist_info = calculatePathlength(exampleComplexHypotheses),
                                           min_distance = 2, append =  TRUE)
## Perform the complex Feature search
complexFeatures <- findComplexFeatures(traces = exampleProteinTraces,
                                       complex_hypothesis = exampleHypotheses)
## estimate the FDR
estimateDecoyFDR(pf_td, grid_search_list = T, FFT = 1)

#-----------------------------
## Protein Features
#-----------------------------

## Load example data
pepTraces <- examplePeptideTraces

## Generate decoy traces
PepTracesDecoys <- generateRandomPepTraces(examplePeptideTraces,
                                           append = TRUE)
## Perform the complex Feature search
proteinFeatures <- findProteinFeatures(traces = PepTracesDecoys,
                                       parallelized = TRUE,
                                       n_cores = 4)
## estimate the FDR
estimateDecoyFDR(complexFeatures, FFT = 1)


#-----------------------------
## Complex Features
#-----------------------------

## Load example data
complexFeatures <- exampleComplexFeatures
## estimate the FDR
estimateDecoyFDR(complexFeatures, FFT = 1)

#-----------------------------
## Protein Features
#-----------------------------

## Load example data
proteinFeatures <-exampleProteinFeatures

## estimate the FDR
estimateDecoyFDR(proteinFeatures, FFT = 1)

}
