% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinSequenceCoverage.R
\name{estimateGlobalSeqenceCoverage}
\alias{estimateGlobalSeqenceCoverage}
\title{Estimate the total coverage of all measured proteins, and of all proteins in your fasta database}
\usage{
estimateGlobalSeqenceCoverage(sequence_coverage, proteinFasta)
}
\arguments{
\item{sequence_coverage}{A \code{data.frame} of sequence coverages per protein, as output by 
\code{calculateSequenceCoverage}.}

\item{proteinFasta}{Path to a fasta file with Protein Sequences and headers that contain
the Protein names in the protein_id column of your peptideTraces object. Usually this will be the
fasta file used for library generation in the proteomics search.}
}
\value{
A list containding the fraction of all AAs covered by peptides of proteins that were measured
and of all proteins in the supplied sequence database
}
\description{
Estimate the total coverage of all measured proteins, and of all proteins in your fasta database
}
