% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R,
%   R/getBestFeatureParameters.R
\name{estimateGridSearchDecoyFDR}
\alias{estimateGridSearchDecoyFDR}
\alias{estimateGridSearchDecoyFDR}
\title{Perform complex feature grid search FDR}
\usage{
estimateGridSearchDecoyFDR(complex_features_list)

estimateGridSearchDecoyFDR(complex_features_list)
}
\arguments{
\item{complex_features}{data.table containing filtered complex feature results.}

\item{complex_features}{data.table containing filtered complex feature results.}
}
\value{
List with stats

List with stats
}
\description{
Perform complex feature grid search.

Perform complex feature grid search.
}
