% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{estimateProteoformPval}
\alias{estimateProteoformPval}
\alias{estimateProteoformPval.traces}
\alias{estimateProteoformPval.tracesList}
\title{Estimate p-values of whether a gene is likely to have >1 proteoforms}
\usage{
estimateProteoformPval(traces, prot_names = NULL, adj.method = "fdr",
  plot = FALSE, PDF = FALSE, name = "SplicePval", ...)

\method{estimateProteoformPval}{traces}(traces, prot_names = NULL,
  adj.method = "fdr", plot = FALSE, PDF = FALSE, name = "SplicePval",
  ...)

\method{estimateProteoformPval}{tracesList}(tracesList, prot_names = NULL,
  adj.method = "fdr", plot = FALSE, PDF = FALSE, name = "SplicePval",
  ...)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{prot_names}{Vactor with protein names to use for fitting. Default NULL.}

\item{adj.method}{Character string which p-value adjustment to use.
Default "fdr".}

\item{plot}{logical,wether to print SibPepCorr density plot to R console.
Deafult is \code{FALSE}.}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file.
Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if
'\code{PDF=TRUE}.PDF file is saved under name.pdf. Default is "GammaFitted".}
}
\value{
Fitted gamma distribution
}
\description{
Estimate p-values of whether a gene is likely to have >1 proteoforms
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Estimate p-values of whether a
gene is likely to have >1 proteoforms

\item \code{tracesList}: Estimate p-values of whether a
gene is likely to have >1 proteoforms
}}

