% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{estimateProteoformPvalCens}
\alias{estimateProteoformPvalCens}
\alias{estimateProteoformPvalCens.traces}
\title{Fit a censored distribution to the minimum correlation of each gene to estimate
a pvalue of isoforms.}
\usage{
estimateProteoformPvalCens(
  traces,
  distr = "beta",
  prot_names = NULL,
  adj.method = "fdr",
  plot = FALSE,
  PDF = FALSE,
  name = "SplicePval",
  design_matrix = NULL,
  min_sig = 2,
  FDR_cutoff = 0.05,
  ...
)

\method{estimateProteoformPvalCens}{traces}(
  traces,
  distribution = c("beta", "gamma"),
  prot_names = NULL,
  adj.method = "fdr",
  split_value = 0.5,
  plot = FALSE,
  PDF = FALSE,
  name = "SplicePval",
  design_matrix = NULL,
  min_sig = 2,
  FDR_cutoff = 0.05,
  ...
)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{prot_names}{Vactor with protein names to use for fitting. Default NULL.}

\item{adj.method}{Character string which p-value adjustment to use.
Default "fdr".}

\item{plot}{logical,wether to print SibPepCorr density plot to R console.
Deafult is \code{FALSE}.}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file.
Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if
'\code{PDF=TRUE}.PDF file is saved under name.pdf. Default is "DistrFitted".}

\item{distribution}{String, which distribution to use.}

\item{split_value}{Numeric between 0 and 1, where to censor distribution.
Default is 0.5.}
}
\value{
Fitted appropriate distribution. Default is gamma distribution.
}
\description{
Fit a censored distribution to the minimum correlation of each gene to estimate
a pvalue of isoforms.
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Estimate p-values of whether a
gene is likely to have >1 proteoforms
}}

