% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateExonSupport.R
\name{evaluateExonLocation}
\alias{evaluateExonLocation}
\title{Evaluate if proteoforms agree with genomic information of exons}
\usage{
evaluateExonLocation(traces, adj.method = "fdr", optional_filter = F)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{adj.method}{Character string which p-value adjustment to use.
Default "fdr".}
}
\value{
Traces with exon evidence p-value per gene
}
\description{
Evaluate if proteoforms agree with genomic information of exons
}
