% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformLocationMapping.R
\name{evaluateProteoformLocation}
\alias{evaluateProteoformLocation}
\alias{evaluateProteoformLocation.traces}
\alias{evaluateProteoformLocation.tracesList}
\title{Evaluate proteoform sequence proximity}
\usage{
evaluateProteoformLocation(
  traces,
  adj.method = "fdr",
  name = "NormalizedSD",
  minPepPerProtein = 6,
  minPepPerProteoform = 3,
  ...
)

\method{evaluateProteoformLocation}{traces}(
  traces,
  adj.method = "fdr",
  name = "NormalizedSD",
  minPepPerProtein = 6,
  minPepPerProteoform = 3
)

\method{evaluateProteoformLocation}{tracesList}(
  traces,
  adj.method = "fdr",
  name = "NormalizedSD",
  minPepPerProtein = 6,
  minPepPerProteoform = 3
)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}
}
\value{
Traces with annotated sequence proximity stats.
}
\description{
Evaluate proteoform sequence proximity
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Evaluate proteoform sequence proximity

\item \code{tracesList}: Evaluate proteoform sequence proximity
}}

