% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleCalibrationTable}
\alias{exampleCalibrationTable}
\title{A sample data.table that holds information on the connectivity between SEC fraction number and apparent
molecular weight (in kDa).}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 6 rows and 2 columns.}
\usage{
exampleCalibrationTable
}
\description{
Each row in this data.table corresponds to an association between SEC fraction number and apparent
molecular weight (in kDa).
The structure of this data.table is as follows:
\itemize{
 \item \code{std_weights_kDa} The molecular weights of the standard proteins.
 \item \code{std_elu_fractions} The fraction numbers where these standard proteins eluted.
}
}
\keyword{datasets}
