% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleComplexFeatures}
\alias{exampleComplexFeatures}
\title{Examplary output of the \code{findComplexFeatures} function}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 35 rows and 27 columns.}
\usage{
exampleComplexFeatures
}
\description{
Each row in this data.table corresponds to a detected complex co-elution signal. Multiple features per 
complex hypotheses can be detected.
}
\examples{
 ## complex feature finding result generation
 calibration <- calibrateMW(exampleCalibrationTable)
 complexFeatures <- findComplexFeatures(traces = exampleProteinTraces, 
                                        complex_hypothesis = exampleComplexHypotheses) 
 all.equal(complexFeatures,exampleComplexFeatures)
}
\keyword{datasets}
