% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleOpenSWATHinput}
\alias{exampleOpenSWATHinput}
\title{A sample data.table representing the output format of the OpenSWATH workflow.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 20834 rows and 66 columns.}
\usage{
exampleOpenSWATHinput
}
\description{
A sample data.table representing the output format of the OpenSWATH workflow.
}
\keyword{datasets}
