% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplePCPdataLong}
\alias{examplePCPdataLong}
\title{Examplary peptide level quantitative PCP-MS dataset in long format.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 44542 rows and 4 columns.
}
\usage{
examplePCPdataLong
}
\description{
A data.table containing quantitative peptide level information across multiple
fractionations that were analysed in separate MS runs.
The structure of this data.table is as follows:
\itemize{
 \item \code{protein_id} Character string with protein ids (here Uniprot ids).
 \item \code{peptide_id} Character string with peptide ids (peptide sequence).
 \item \code{filename} Character string with filenames of the separate MS runs.
 \item \code{intensity} Intensity values or other quantitative information.
}
}
\keyword{datasets}
