% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplePCPdataWide}
\alias{examplePCPdataWide}
\title{Examplary peptide level quantitative PCP-MS dataset in wide format}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 1889 rows and 83 columns.}
\usage{
examplePCPdataWide
}
\description{
A data.table containing quantitative peptide level information across multiple
fractionations that were analysed in separate MS runs.
The structure of this data.table is as follows:
\itemize{
 \item \code{protein_id} Character string with protein ids (here Uniprot ids).
 \item \code{peptide_id} Character string with peptide ids (peptide sequence).
 \item \code{<filename1>} Intensity values or other quantitative information of <filename1>.
 \item \code{<filename2>} Intensity values or other quantitative information of <filename2>.
 \item \code{...}
 \item \code{<filenameX>} Intensity values or other quantitative information of <filenameX>.
}
}
\keyword{datasets}
