% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplePeptideTraces}
\alias{examplePeptideTraces}
\title{Examplary peptide level traces object before any filtering
#' 
Peptide level traces object as generated by annotating a raw peptide level tarces object with 
Uniprot annotations and a molecular weight calibration.}
\format{An object of class \code{traces} of length 4.}
\usage{
examplePeptideTraces
}
\description{
Examplary peptide level traces object before any filtering
#' 
Peptide level traces object as generated by annotating a raw peptide level tarces object with 
Uniprot annotations and a molecular weight calibration.
}
\examples{
## peptide traces annotation
rawPeptideTraces <- examplePeptideTracesUnannotated
calibration <- calibrateMW(exampleCalibrationTable)
peptideTracesUniprot <- annotateTraces(rawPeptideTraces,
                                    trace_annotation = exampleTraceAnnotation)
peptideTracesUniprotCalibration <- annotateMolecularWeight(peptideTracesUniprot,
                                                               calibration = calibration)
all.equal(peptideTracesUniprotCalibration,examplePeptideTraces)
}
\keyword{datasets}
