% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplePeptideTracesFiltered}
\alias{examplePeptideTracesFiltered}
\title{A sample peptide level traces object after consecutive id and sibling peptide correlation filter.}
\format{An object of class \code{traces} of length 4.}
\usage{
examplePeptideTracesFiltered
}
\description{
A sample peptide level traces object after consecutive id and sibling peptide correlation filter.
}
\keyword{datasets}
