% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplePeptideTracesUnannotated}
\alias{examplePeptideTracesUnannotated}
\title{Examplary raw peptide level traces object before any annootation}
\format{
An object of class \code{traces} of length 4.
}
\usage{
examplePeptideTracesUnannotated
}
\description{
Peptide level traces object as generated directly from the import of a PCP-MS dataset \code{importPCPdata}
such as \code{examplePCPdataLong} or \code{examplePCPdataWide}.
}
\examples{
## raw peptide traces object generation from examplary PCP-MS data
rawPeptideTraces <- importPCPdata(input_data = examplePCPdataLong,
                                  fraction_annotation = exampleFractionAnnotation)
all.equal(rawPeptideTraces,examplePeptideTracesUnannotated)
}
\keyword{datasets}
