% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleProteinFeatures}
\alias{exampleProteinFeatures}
\title{Examplary output of the \code{findProteinFeatures} function}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 461 rows and 16 columns.
}
\usage{
exampleProteinFeatures
}
\description{
Each row in this data.table corresponds to a detected protein co-elution signal. Multiple features per 
protein can be detected.
}
\examples{
## Protein feature finding result generation
calibration <- calibrateMW(exampleCalibrationTable)
proteinFeatures <- findProteinFeatures(traces = examplePeptideTracesFiltered) 
all.equal(proteinFeatures,exampleProteinFeatures)
}
\keyword{datasets}
