% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleProteinTraces}
\alias{exampleProteinTraces}
\title{A sample protein level traces object.}
\format{An object of class \code{traces} of length 4.}
\usage{
exampleProteinTraces
}
\description{
A sample protein level traces object.
}
\keyword{datasets}
