% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatureVals.R
\name{extractFeatureVals}
\alias{extractFeatureVals}
\alias{extractFeatureVals.traces}
\alias{extractFeatureVals.tracesList}
\title{Extract intensity, correlation etc. within feature boundaries}
\usage{
extractFeatureVals(traces, features, perturb_cutoff = "5\%",
  verbose = TRUE, extract = "subunits_detected", imputeZero = FALSE,
  fill = FALSE, minValues = NULL, ...)

\method{extractFeatureVals}{traces}(traces, features,
  perturb_cutoff = "5\%", verbose = TRUE,
  extract = "subunits_detected", imputeZero = FALSE,
  minValues = NULL, ...)

\method{extractFeatureVals}{tracesList}(traces, features,
  perturb_cutoff = "5\%", verbose = TRUE, design_matrix = NULL,
  extract = "subunits_detected", imputeZero = FALSE, fill = FALSE,
  minValues = NULL, ...)
}
\arguments{
\item{traces}{object of class traces}

\item{features}{Protein/complex feature table returned by \code{findProteinFeatures} or
\code{findComplexFeatures}.}

\item{perturb_cutoff}{Character string or numeric, the quantile of values that noise is sampled from.
Noise needs to be imputed to calculate the correlations.}

\item{verbose}{Logical, whether to print messages to console.}

\item{extract}{Character string, name ot the feature table column containing the subunits to extract.}

\item{imputeZero}{Logical, whether zero values of a fraction within a feature should be imputed, default = FALSE.}

\item{minValues}{data.table with minimum values for each peptide, default=NULL.}
}
\value{
Long format table containing extracted feature values.
}
\description{
Goes through protein/complex features and extracts metrics from the traces object
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Extract values from single traces object.

\item \code{tracesList}: Extract values from single traces object.
}}

