% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIdsFromFastaHeader.R
\name{extractIdsFromFastaHeader}
\alias{extractIdsFromFastaHeader}
\title{extractIdsFromFastaHeader}
\usage{
extractIdsFromFastaHeader(fasta_ids, id_type = "accession",
  keep.prefix = TRUE)
}
\arguments{
\item{fasta_ids}{character vector of fasta protein ids}

\item{id_type}{"accession" or "name"}

\item{keep.prefix}{whether IDs shall stay merged to the prefix 1/ or DECOY_1/}
}
\description{
sp|P11021|GRP78_HUMAN --> P11021
}

