% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByRepPepCorr.R
\name{filterByRepPepCorr}
\alias{filterByRepPepCorr}
\title{Filter by sibling peptide correlation}
\usage{
filterByRepPepCorr(
  traces,
  design_matrix,
  condition_column = NULL,
  stepwise_filter = TRUE,
  repPepCorr_cutoff = 0.5
)
}
\arguments{
\item{traces}{object of class tracesList.}

\item{design_matrix}{data.table, design matrix describing the architecture of the tracesList object.}

\item{condition_column}{Character string, The rowname of the design matrix containing
the experimental conditions.}

\item{stepwise_filter}{Logical, wether to iteratively discard replicates until the defined cutoff
is met. E.g. if the avg. RepPepCorr is below the cutoff, discard the worst replicate
and check again. If the cutoff can't be met with 2 replicates the peptide is discarded.}

\item{repPepCorr_cutoff}{Numeric, the cutoff below which traces are removed.}
}
\value{
tracesList object containing the integrated traces
}
\description{
Filter peptides in traces object based on
Replicate Peptide Correlation (RepPepCorr). This function can be called if 
one has replicate measurements of one experimental Condition. Trustworthy peptides should
show robust elution behaviour over these replicates.
}
