% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterConsecutiveIdStretches.R
\name{filterConsecutiveIdStretches}
\alias{filterConsecutiveIdStretches}
\title{filterConsecutiveIdStretches filters protein profiling traces based on
minimal length of consecutive ID stretches.}
\usage{
filterConsecutiveIdStretches(Traces, min_stretch_length = 3,
  remove_empty = TRUE)
}
\arguments{
\item{remove_empty}{Logical whether the entries with rowSum == 0 after
filtering are removed, defaults to TRUE.}

\item{traces}{object of class traces containing the elution profiles that
are to be filtered.}

\item{min_stretch_len}{The minimal length a stretch of contiguous
identifications has to have in order not to be removed.}
}
\value{
An object of class traces containing the filtered
    chromatograms.
}
\description{
filterConsecutiveIdStretches filters protein profiling traces based on
minimal length of consecutive ID stretches.
}

