% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureMethods.R
\name{filterFeatures}
\alias{filterFeatures}
\title{Filter co-elution feature table}
\usage{
filterFeatures(feature_table, complex_ids = NULL, protein_ids = NULL,
  min_feature_completeness = NULL, min_hypothesis_completeness = NULL,
  min_subunits = NULL, min_peak_corr = NULL,
  min_monomer_distance_factor = 2)
}
\arguments{
\item{feature_table}{data.table as reported by \code{\link[SECprofiler]{findComplexFeatures}}
or \code{\link[SECprofiler]{findProteinFeatures}}.}

\item{complex_ids}{character vector containing all desired \code{complex_id} values.
Only applicable to complex feature tables. Defaults to \code{NULL}.}

\item{protein_ids}{character vector containing all desired \code{protein_id} values.
Only applicable to protein feature tables. Defaults to \code{NULL}.}

\item{min_feature_completeness}{Numeric between 0 and 1, specifying the required completeness
of a feature reltive to the tested hypothesis (keeps all features if at least one is bigger than the cutoff).}

\item{min_hypothesis_completeness}{Numeric between 0 and 1, specifying the required completeness
of the most complete feature reltive to the tested hypothesis (keeps all features if at least one feature is bigger than the cutoff).}

\item{min_subunits}{Integer specifying minimum number of subunits in a co-elution feature.}

\item{min_peak_corr}{Numeric value betwee 0 and 1 specifying minimum peak correlation, defaults to \code{NULL}.}

\item{min_monomer_distance_factor}{Numeric value specifying a factor to multiply the largest monomer molecular weight, defaults to 2.
This filters out features that have their apex at a smaller molecular weight than the resulting min_monomer_distance_factor*max(monomer_mw) value.
Using this filtering option can, for example, remove complex features that are likely spontaneous co-elutions of the subunits monomers.}
}
\value{
The same feture table as teh input, but filtered according to the provided parameters.
}
\description{
Filter co-elution feature table according to desired criteria.
}
\examples{
## Load example complex feature finding results:
complexFeatures <- exampleComplexFeatures
## Run summary function:
summarizeFeatures(complexFeatures)
## Filter complex features by a peak correlation of 0.5, a minimum hypothesis
## completeness of 0.5 and a minimum distance to the monomers by a factor of 2:
filteredComplexFeatures <- filterFeatures(complexFeatures,
                                     min_peak_corr=0.5,
                                     min_hypothesis_completeness=0.5,
                                     min_monomer_distance_factor=2)
## Run summary function on filtered data:
summarizeFeatures(filteredComplexFeatures)
}
