% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestFeatureParameters.R
\name{filterGridSearchResults}
\alias{filterGridSearchResults}
\title{Perform complex feature grid search filter}
\usage{
filterGridSearchResults(
  grid_search_results,
  peak_corr_cutoffs = c(0.5, 0.75, 0.9),
  feature_completeness_cutoffs = c(0, 0.5, 1),
  hypothesis_completeness_cutoffs = c(0.5, 1),
  n_subunits_cutoffs = c(2, 3, 4),
  monomer_distance_cutoffs = c(1, 2),
  remove_decoys = FALSE
)
}
\arguments{
\item{grid_search_results}{List containing result tables from a feature finding grid search.}

\item{peak_corr_cutoffs}{Numeric, vector of within peak correlation_cutoff values to test.
Between -1 and 1. Default is c(0.5,0.75,0.9).}

\item{feature_completeness_cutoffs}{Numeric, vector of feature completeness cutoffs to test.
Between 0 and 1. Default is c(0,0.5,1).}

\item{hypothesis_completeness_cutoffs}{Numeric, vector of hypothesis completeness cutoffs to test.
Between 0 and 1. Default is c(0.5,1).}

\item{n_subunits_cutoffs}{Positive integer vector of minimum number of subunits per hypothesis cutoffs to test. Default is c(2,3,4).}

\item{monomer_distance_cutoffs}{Positive numeric, factor of allowed distance to monomer weight to test. Default is c(1,2).}

\item{remove_decoys}{Logical, whether to remove the decoys from the result.
Default=\code{FALSE}.}
}
\value{
List of search result tables for every possible parameter combination.
The result tables contain additional columns specifying the parameters.

List with stats
}
\description{
Perform quality filtering of the feature finding result with different
cutoffs (grid serch). All possible combinations of the specified parameters are tested.
}
\examples{
#------------------------
## Complex level
#------------------------

## Load example data into list to simulate grid search results
complexFeaturesGrid <- list(exampleComplexFeatures)

## Perform the filter grid search
complexFeaturesGridFiltered <- filterGridSearchResults(complexFeaturesGrid,
                                                       peak_corr_cutoffs = c(0.5,0.75,0.9),
                                                       feature_completeness_cutoffs = c(0,0.5,1),
                                                       hypothesis_completeness_cutoffs = c(0.5,1),
                                                       n_subunits_cutoffs = c(2,3,4),
                                                       monomer_distance_cutoffs = c(1,2),
                                                       remove_decoys=FALSE)
#------------------------
## Protein level
#------------------------

## Load example data into list to simulate grid search results
proteinFeaturesGrid <- list(exampleProteinFeatures)

## Perform the filter grid search
 proteinFeaturesGridFiltered <- filterGridSearchResults(proteinFeaturesGrid,
                                                        peak_corr_cutoffs = c(0.5,0.75,0.9),
                                                        feature_completeness_cutoffs = c(0,0.5,1),
                                                        hypothesis_completeness_cutoffs = c(0.5,1),
                                                        n_subunits_cutoffs =c(2,3,4),
                                                        monomer_distance_cutoffs = c(0,1),
                                                        remove_decoys=FALSE)

}
