% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterIdStretches.R
\name{filterIdStretches}
\alias{filterIdStretches}
\title{Detect subgroups within a window. This is a helper function and should 
be called by `findComplexFeatures`.}
\usage{
filterIdStretches(sec.profiles, stretch.len)
}
\arguments{
\item{sec.profiles}{An object of class SECprofiles that is to be filtered.}

\item{stretch.len}{The minimal length a stretch of contiguous identifications
has to have in order not to be removed.}
}
\value{
An object of class SECprofiles that represents the filtered
        chromatograms.
}
\description{
Detect subgroups within a window. This is a helper function and should 
be called by `findComplexFeatures`.
}

