% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeatures.R
\name{findComplexFeatures}
\alias{findComplexFeatures}
\title{Complex feature detection}
\usage{
findComplexFeatures(traces, complex_hypothesis, corr_cutoff = 0.95,
  window_size = 15, parallelized = FALSE, n_cores = 1,
  collapse_method = "apex_only", perturb_cutoff = "5\%", rt_height = 5,
  smoothing_length = 11)
}
\arguments{
\item{traces}{An object of class traces.}

\item{complex_hypothesis}{data.table containing complex hypotheses.
Should have the following columns:
\itemize{
\item complex_id: character strings, a unique id for every complex
\item complex_name: character strings, a unique name for every complex
\item protein_id: character strings, the protein id, e.g. Uniprot id
}}

\item{corr_cutoff}{Numeric, the correlation value for chromatograms above which
peptides are considered to be coeluting, default=0.95.}

\item{window_size}{Integer, size of the window in fractions, default=15}

\item{parallelized}{Logical, if the computation should be done in parallel, default=\code{FALSE}.}

\item{n_cores}{Integer, the number of cores to use for parallel processing
(only applies if parallelized is \code{TRUE}), default=1.}

\item{collapse_method}{Character, method for collapsing multiple features into one feature:
\itemize{
\item "apex_only": collapses by apex
\item "apex_network": collapses by apex and connected network cluster}
Default="apex_only"}

\item{perturb_cutoff}{Numeric, the quantile to use in estimating the perturbation level, default="5%".
Intensity values that are zero are replaced with random values that are
below the specified quantile of the input values. Alternatively a
cutoff value can be specified as an upper limit for perturbation values.
This is nescessary for correlation calculation.}

\item{rt_height}{Numeric, RT cutoff for collapsing features. Defaults to 5.}

\item{smoothing_length}{Numeric, smoothing length of Savitzky-Golay filter. Defaults to 11.}
}
\value{
A data.table containing protein complex features.
}
\description{
Run the sliding window algorithm to find complex features.
}
\examples{
## Load example data
proteinTraces <- exampleProteinTraces
complexHypotheses <- exampleComplexHypotheses

## Perform co-elution signal detection
complexFeatures <- findComplexFeatures(traces=proteinTraces,
                                       complex_hypothesis=complexHypotheses)

## Inspect complex features
summarizeComplexFeatures(complexFeatures)

}
