% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeaturesPP.R
\name{findComplexFeaturesPP}
\alias{findComplexFeaturesPP}
\title{A helper function to perform peak picking on the complex features detected by
the sliding-window approach.}
\usage{
findComplexFeaturesPP(traces.obj, complexFeaturesSW, smoothing_length = 11,
  rt_height = 5)
}
\arguments{
\item{traces.obj}{An object of type \code{traces.obj}.}

\item{complexFeaturesSW}{An object of type \code{complexFeaturesSW} that is a list
containing the following:
\itemize{
  \item \code{feature} data.table containing complex feature candidates in the following format:
   \itemize{
   \item \code{subgroup} The protein_ids of the feature separated by semi-colons.
   \item \code{left_sw} The left boundary of the sliding-window feature.
   \item \code{right_sw} The right boundary of the sliding-window feature.
   \item \code{score} The intra-sliding-window-feature correlation.
   }
  \item \code{window.size} The window.size used when running this
      function.
  \item \code{corr.cutoff} The corr.cutoff used when running this
      function.
}}

\item{smoothing_length}{Integer length of savitzgy-golay filter.}
}
\value{
An object of type \code{complexFeaturesPP} that is a list
       containing the following:
       \itemize{
         \item \code{feature} data.table containing complex feature candidates in the following format:
          \itemize{
          \item \code{subgroup} The protein_ids of the feature separated by semi-colons.
          \item \code{left_sw} The left boundary of the sliding-window feature.
          \item \code{right_sw} The right boundary of the sliding-window feature.
          \item \code{score} The intra-sliding-window-feature correlation.
          \item \code{n_subunits} The number of protein_ids in the sliding-window feature.
          \item \code{apax} The apex of the selected peak by the peak-picker.
          \item \code{left_pp} The left boundary of the selected peak by the peak-picker.
          \item \code{right_pp} The right boundary of the selected peak by the peak-picker.
          \item \code{area} The area (entire complex) of the selected peak by the peak-picker.
          }
       }
}
\description{
A helper function to perform peak picking on the complex features detected by
the sliding-window approach.
}
