% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeaturesSW.R
\name{findComplexFeaturesSW}
\alias{findComplexFeaturesSW}
\title{Detect subgroups of proteins within a matrix of protein intensity traces by
sliding a window across the SEC dimension. Within each window proteins
with traces that correlate well are clustered together.}
\usage{
findComplexFeaturesSW(traces.obj, corr.cutoff = 0.95, window.size = 15,
  with.plot = F, noise.quantile = 0.2, min.sec = 1)
}
\arguments{
\item{corr.cutoff}{The correlation value for chromatograms above which
proteins are considered to be coeluting.}

\item{window.size}{Size of the window. Numeric.}

\item{noise.quantile}{The quantile to use in estimating the noise level.
Intensity values that are zero are imputed with random noise
according to the noise estimation.}

\item{min.sec}{The lowest SEC number in the sample.}

\item{trace.mat}{A numeric matrix where rows correspond to the different
traces.}

\item{protein.names}{A vector with protein identifiers. This vector has to
have the same length as the number of rows in `trace.mat`.}

\item{protein.mw.conc}{A data.table that stores the molecular weight and
estimate of the absolute abundance for each subunit.
\itemize{
 \item \code{protein_id}
 \item \code{protein_mw}
 \item \code{protein_concentration}
}}
}
\value{
An instance of class `complexFeaturesSW`. This is a list with the
        following entries:
        \itemize{
         \item \code{features} A datatable of features. Each feature
             can span several SEC fractions.
         \item \code{window.size} The window.size used when running this
             function.
         \item \code{corr.cutoff} The corr.cutoff used when running this
             function
        }
}
\description{
Detect subgroups of proteins within a matrix of protein intensity traces by
sliding a window across the SEC dimension. Within each window proteins
with traces that correlate well are clustered together.
}
\examples{
# NOT RUN:
# protein.ids <- corum.complex.protein.assoc[complex_id == 181, protein_id]
# traces <- subset(protein.traces[protein_id \%in\% protein.ids],
#                  select=-protein_id)
# sw.res <- findComplexFeaturesSW(traces, protein.ids, protein.mw.conc)
}

