% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeaturesSWBulk.R
\name{findComplexFeaturesSWBulk}
\alias{findComplexFeaturesSWBulk}
\title{Run the sliding window algorithm for a number of protein complexes.}
\usage{
findComplexFeaturesSWBulk(trace.mat, protein.names, protein.mw.conc,
  corr.cutoff, window.size, complex.protein.assoc, parallelized = FALSE,
  n.cores = parallel::detectCores())
}
\arguments{
\item{trace.mat}{A numeric matrix where rows correspond to the different
traces.}

\item{protein.names}{A vector with protein identifiers. This vector has to
have the same length as the number of rows in `trace.mat`.}

\item{protein.mw.conc}{A data.table that stores the molecular weight and
estimate of the absolute abundance for each subunit.
\itemize{
 \item \code{protein_id}
 \item \code{protein_mw}
 \item \code{protein_concentration}
}}

\item{corr.cutoff}{The correlation value for chromatograms above which
proteins are considered to be coeluting.}

\item{window.size}{Size of the window. Numeric.}

\item{complex.protein.assoc}{A data.table that encodes the input complexes
and their subunit compositions. The data.table should have the
format:
\itemize{
 \item \code{complex_id} A character vector that uniquely identifies
       a complex.
 \item \code{protein_id} A character vector that uniquely identifies
       a protein and links it to a trace contained in
       \code{protein.traces}.
}}

\item{parallelized}{If the computation should be done in parallel.
Optional.}

\item{n.cores}{The number of cores to use for parallel processing. Optional.}

\item{protein.mw.conc}{A data.table that stores the molecular weight and
estimate of the absolute abundance for each subunit.
\itemize{
 \item \code{protein_id}
 \item \code{protein_mw}
 \item \code{protein_concentration}
}}
}
\value{
A list containing various results.
        \itemize{
         \item \code{sw.results} A list of results of the function
               \code{findComplexFeatures}. One for each query complex.
         \item \code{complex.stats} A data.table specifying for each query
               complex how many of its subunits were detected.
         \item \code{input.complexes} All query complexes.
         \item \code{input.complexes} A character vector of all query
                complexes.
         \item \code{corr.cutoff} The correlation cutoff used.
         \item \code{window.size} The window size used.
         \item \code{complex.protein.assoc} The protein-complex associations
               used as the original input argument.
        }
}
\description{
Run the sliding window algorithm for a number of protein complexes.
}
\examples{
# NOT RUN:
# protein.ids <- corum.complex.protein.assoc[complex_id == 181, protein_id]
# traces <- subset(protein.traces[protein_id \%in\% protein.ids],
#                  select=-protein_id)
# sample.complexes <- c(181, 191)
# complex.protein.assoc <-
#     corum.complex.protein.assoc[complex_id \%in\% sample.complexes]
# findComplexFeaturesSWBulk(traces,
#                           protein.ids,
#                           protein.mw.conc[protein_id \%in\% protein.ids],
#                           complex.protein.assoc)
}

