% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeaturesSWBulk.R
\name{findComplexFeaturesTargeted}
\alias{findComplexFeaturesTargeted}
\title{Run the sliding window algorithm for a number of protein complexes.}
\usage{
findComplexFeaturesTargeted(protein.traces, protein.mw.conc,
  complex.protein.assoc, corr.cutoff = 0.99, window.size = 15,
  parallelized = FALSE, n.cores = parallel::detectCores())
}
\arguments{
\item{protein.traces}{A wide format data.table where the first column
is named \code{protein_id} and all other columns correspond to
intensity values measured in SEC fractions.}

\item{protein.mw.conc}{A data.table that stores the molecular weight and
estimate of the absolute abundance for each subunit.
\itemize{
 \item \code{protein_id}
 \item \code{protein_mw}
 \item \code{protein_concentration}
}}

\item{complex.protein.assoc}{A data.table that encodes the input complexes
and their subunit compositions. The data.table should have the
format:
\itemize{
 \item \code{complex_id} A character vector that uniquely identifies
       a complex.
 \item \code{protein_id} A character vector that uniquely identifies
       a protein and links it to a trace contained in
       \code{protein.traces}.
}}

\item{corr.cutoff}{The correlation cutoff to use in the sliding window
algorithm. Optional.}

\item{window.size}{The window size to use in the sliding window algorithm.
Optional.}

\item{parallelized}{If the computation should be done in parallel.
Optional.}

\item{n.cores}{The number of cores to use for parallel processing. Optional.}
}
\value{
A list containing various results.
        \itemize{
         \item \code{sw.results} A list of results of the function
               \code{findComplexFeatures}. One for each query complex.
         \item \code{complex.stats} A data.table specifying for each query
               complex how many of its subunits were detected.
         \item \code{input.complexes} All query complexes.
         \item \code{input.complexes} A character vector of all query
                complexes.
         \item \code{corr.cutoff} The correlation cutoff used.
         \item \code{window.size} The window size used.
         \item \code{complex.protein.assoc} The protein-complex associations
               used as the original input argument.
        }
}
\description{
Run the sliding window algorithm for a number of protein complexes.
}
\examples{
# NOT RUN:
# sample.complexes <-
#     sample(unique(corum.complex.protein.assoc$complex_id), 10)
# complex.protein.assoc <-
#     corum.complex.protein.assoc[complex_id \%in\% sample.complexes]
# findComplexFeaturesSWBulk(protein.traces, protein.mw.conc,
#                           complex.protein.assoc)
}

