% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeaturesSW.R
\name{findComplexFeaturesWithinWindow}
\alias{findComplexFeaturesWithinWindow}
\title{Detect subgroups within a window. This is a helper function and should 
be called by `findComplexFeaturesSW`.}
\usage{
findComplexFeaturesWithinWindow(tracemat, corr.cutoff, protein.names,
  with.plot = F, sec = NULL)
}
\arguments{
\item{tracemat}{A matrix of intensity values where chromatograms of
individual proteins are given by the rows.}

\item{corr.cutoff}{The correlation value for chromatograms above which
proteins are considered to be coeluting.}

\item{protein.names}{A vector with protein identifiers. This vector has to
have the same length as the number of rows in `tracemat`.}
}
\value{
A list containing lists that describe the found subclusters.
        An individual list has the structure:
        \itemize{
         \item \code{subunits}: character vector
         \item \code{mean.corr}: intra cluster correlation
        }
}
\description{
Detect subgroups within a window. This is a helper function and should 
be called by `findComplexFeaturesSW`.
}

