% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeaturesSW.R
\name{findFeatureBoundaries}
\alias{findFeatureBoundaries}
\title{Helper function to find boundaries of complex features.}
\usage{
findFeatureBoundaries(m, subgroup.names, groups.dt)
}
\arguments{
\item{m}{Logical matrix where an entry m[i, j] indicates if subgroup i was
detected at SEC fraction j.}
}
\description{
Helper function to find boundaries of complex features.
}

