% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFeaturePeaks.R
\name{findFeaturePeaks}
\alias{findFeaturePeaks}
\title{A helper function to extend a list of complex features with additional
information.}
\usage{
findFeaturePeaks(features, trace.mat, protein.names, protein.mw.conc)
}
\arguments{
\item{features}{A data.table of complex feature candidates with the
following format:
\itemize{
 \item \code{subgroup} A semicolon-separated list of protein
                       identifiers.
 \item \code{left_sw} The left boundary of the feature.
 \item \code{right_sw} The right boundary of the feature.
 \item \code{score} The intra-feature correlation.
}}

\item{trace.mat}{A matrix where rows correspond to protein traces.
This is the matrix that was used to find complex features.}

\item{protein.names}{A character vector specifying the protein identifiers
belonging to the rows in \code{trace.mat}.}

\item{protein.mw.conc}{A data.table that stores the molecular weight and
estimate of the absolute abundance for each subunit.
\itemize{
 \item \code{protein_id}
 \item \code{protein_mw}
 \item \code{protein_concentration}
}}
}
\value{
The same data.table as the input argument extended with the
        following columns:
        \itemize{
         \item \code{n_subunits} The number of subunits in the feature.
         \item \code{stoichiometry} The intensity-based stoichiometry.
         \item \code{mw_estimated} The estimated molecular weight.
         \item \code{mw_apparent} The apparent mw.
         \item \code{mw_delta} The delta mw.
        }
}
\description{
A helper function to extend a list of complex features with additional
information.
}
