% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteogenomics.R
\name{findLeadingIso}
\alias{findLeadingIso}
\title{Convert a presence matrix of isoform annotations to a vector of leading Isoforms}
\usage{
findLeadingIso(presenceMat)
}
\arguments{
\item{presenceMat}{numeric matrix with peptides as rows and isoforms as columns,
where a one indicates that that peptide is part of an isoform and a 0 otherwhise.}
}
\value{
character vector with length = nrow(presenceMat), with the leading
Isoform of every peptide.
}
\description{
Convert a presence matrix of isoform annotations to a vector of leading Isoforms
}
\details{
Applies a maximum parsimony approach, where the least number of isoforms
are used to explain a peptide.
}
