% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissingValues.R
\name{findMissingValues}
\alias{findMissingValues}
\alias{findMissingValues.traces}
\alias{findMissingValues.tracesList}
\title{Find the position of missing values in a traces object}
\usage{
findMissingValues(traces, bound_left = 2, bound_right = 2,
  consider_borders = TRUE)

\method{findMissingValues}{traces}(traces, bound_left = 2, bound_right = 2,
  consider_borders = TRUE)

\method{findMissingValues}{tracesList}(tracesList, bound_left = 2,
  bound_right = 2, consider_borders = TRUE)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{bound_left}{Numeric integer, the minimum number of non-zero values to the
left of a missing value to be replaced with \code{NA}.}

\item{bound_right}{Numeric integer, the minimum number of non-zero values to the
right of a missing value to be replaced with \code{NA}.}

\item{consider_borders}{Logical, whether to find missin values in the first/last fractions
(e.g. \code{0-1-1 -> NA-1-1} and \code{1-0-1-1 -> 1-NA-1-1} if the leftmost value is the
first fraction of the traces object).}
}
\value{
Object of class traces. Missing values are replaced with \code{NA}.
}
\description{
Replaces 0 values in a traces object with \code{NA} according to
a specified rule. NA values can then be imputed.
}
\details{
The algorithm identifies 0 values on the dataset as missing values if they fulfill
the rule: \code{(1)*bound_left - 0 - (1)*bound_right} i.e. a zero value has to have at least
a specified number of non-zero neighbors to be classified as \code{NA}.
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Find missing values one traces object

\item \code{tracesList}: Find missing values in all traces within a tracesList
}}
\examples{
tracesMiss <- findMissingValues(examplePeptideTraces)
View(tracesMiss$traces)
}

