% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findProteinFeatures.R
\name{findProteinFeatures}
\alias{findProteinFeatures}
\title{Run the sliding window algorithm to find protein features.}
\usage{
findProteinFeatures(pepTraces, MWSECcalibrationFunctions, corr.cutoff,
  window.size, parallelized, n.cores, collapse_method, perturb.cutoff,
  rt_height = 5, smoothing_length = 11)
}
\arguments{
\item{pepTraces}{An object of type \code{traces.obj}.}

\item{corr.cutoff}{The correlation value for chromatograms above which
peptides are considered to be coeluting.}

\item{window.size}{Size of the window. Numeric.}

\item{parallelized}{If the computation should be done in parallel.
Optional.}

\item{n.cores}{The number of cores to use for parallel processing. Optional.}

\item{perturb.cutoff}{The quantile to use in estimating the perturbation level.
Intensity values that are zero are replaced with random values that are
below the specified quantile of the input values. Alternatively a
cutoff value can be specified as an upper limit for perturbation values.
This is nescessary for correlation calculation.}
}
\value{
A data.table containing several columns TODO.
}
\description{
Run the sliding window algorithm to find protein features.
}

