% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{fitDistCensored}
\alias{fitDistCensored}
\title{Fit appropriate distributon to minimum correlation of each peptide
to all of its sibling peptides}
\usage{
fitDistCensored(
  x,
  split_value = 0.5,
  start = NULL,
  distribution = c("beta", "gamma", "f"),
  plot = FALSE,
  PDF = FALSE,
  name = "DistrFitted",
  ...
)
}
\arguments{
\item{x}{Vector with values to fit distribution to.}

\item{split_value}{Numeric between 0 and 1. Default is 0.5.}

\item{distribution}{String, which distribution to use.}

\item{plot}{logical,wether to print SibPepCorr density plot to R console.
Deafult is \code{FALSE}.}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file.
Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if
'\code{PDF=TRUE}.PDF file is saved under name.pdf. Default is "DistrFitted".}
}
\value{
Fitted appropriate distribution. Default is gamma distribution.
}
\description{
Fit appropriate distributon to minimum correlation of each peptide
to all of its sibling peptides
}
