% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{fitGammaDist}
\alias{fitGammaDist}
\title{Fit gamma distributon to minimum correlation of each peptide
to all of its sibling peptides}
\usage{
fitGammaDist(traces, prot_names = NULL, distr = "gamma",
  split_value = 0.2, plot = FALSE, PDF = FALSE, name = "GammaFitted")
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{prot_names}{Vactor with protein names to use for fitting. Default NULL.}

\item{distr}{Character string which distribution to fit. Default "gamma".}

\item{split_value}{Numeric between 0 and 1. Default is 0.2.}

\item{plot}{logical,wether to print SibPepCorr density plot to R console.
Deafult is \code{FALSE}.}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file.
Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if
'\code{PDF=TRUE}.PDF file is saved under name.pdf. Default is "GammaFitted".}
}
\value{
Fitted gamma distribution
}
\description{
Fit gamma distributon to minimum correlation of each peptide
to all of its sibling peptides
}

