% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypotheses.R
\name{generateBinaryNetwork}
\alias{generateBinaryNetwork}
\title{Generate binary network from concrete complex hypotheses (e.g. CORUM).}
\usage{
generateBinaryNetwork(complex_hypothesis)
}
\arguments{
\item{complex_hypothesis}{data.table in the format of complex hypotheses}
}
\value{
data.table with binary interactions between "a" and "b"
}
\description{
Generate binary network from concrete complex hypotheses (e.g. CORUM).
}

