% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNetworkTable.R
\name{generateNetworkTable}
\alias{generateNetworkTable}
\title{From protein level Traces, generate a Network table with pairwise profile
correlation score}
\usage{
generateNetworkTable(Traces, correlation_cutoff = 0.9, write_csv = TRUE)
}
\arguments{
\item{Traces}{A Traces object containing wide format elution profiles.}

\item{correlation_cutoff}{Minimum pearson correlation value to report an
edge between two proteins.}

\item{write_file}{Logical (TRUE/FALSE) whether an output file named
[systime]_NetworkTable.csv shall be written to the working folder.}
}
\value{
A data.table containing columns UniProt_A UniProt_B and
    pearson_correlation, readable e.g. by Cytoscape.
}
\description{
From protein level Traces, generate a Network table with pairwise profile
correlation score
}

