% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinFeatureGridSearch.R
\name{generateRandomPepTraces}
\alias{generateRandomPepTraces}
\title{generateRandomPepTraces}
\usage{
generateRandomPepTraces(traces)
}
\arguments{
\item{traces}{traces bject of type peptide}
}
\value{
traces object of type peptide with random peptides per protein
}
\description{
generateRandomPepTraces.
}
