% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{get_tree_height}
\alias{get_tree_height}
\title{Compute the height of the tree without the outlier peptides}
\usage{
get_tree_height(tree, clusters)
}
\arguments{
\item{tree}{hclust object}

\item{clusters}{vector of cluster assignments (as produced by cutree)}
}
\description{
Compute the height of the tree without the outlier peptides
}
