% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotatedTraces.R
\name{getAnnotatedFragmentTraces}
\alias{getAnnotatedFragmentTraces}
\title{Get a data.table of intensities with two columns \code{analyte_id} and
\code{fragment_id}.}
\usage{
getAnnotatedFragmentTraces(trace.obj)
}
\arguments{
\item{trace.obj}{An object of type \code{FragmentTraces}.}
}
\value{
A data.table of annotated traces.
}
\description{
Get a data.table of intensities with two columns \code{analyte_id} and
\code{fragment_id}.
}

