% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotatedTraces.R
\name{getAnnotatedTraces}
\alias{getAnnotatedTraces}
\title{Get a data.table of annotated Traces}
\usage{
getAnnotatedTraces(traces.obj)
}
\arguments{
\item{trace.obj}{An object of type \code{Traces}.}
}
\value{
A data.table of annotated traces.
}
\description{
Get a data.table of annotated Traces
}
