% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexMethods.R
\name{getBestComplexFeature}
\alias{getBestComplexFeature}
\title{Select best complex feature for each hypothesis}
\usage{
getBestComplexFeature(res)
}
\arguments{
\item{A}{data.table with potentially muliple features for each query complex.}
}
\value{
A data.table with the "best" detected complex feature of each query complex.
}
\description{
Reformat result object to a data.table containing only the best subcomplex feature per complex query.
}
