% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureMethods.R
\name{getBestFeatures}
\alias{getBestFeatures}
\title{Select best feature for each complex or protein id.}
\usage{
getBestFeatures(feature_table)
}
\arguments{
\item{feature_table}{data.table as reported by \code{\link[SECprofiler]{findComplexFeatures}}
or \code{\link[SECprofiler]{findProteinFeatures}}.}
}
\value{
data.table Filtered feature table with only one best feature per complex or protein id.
}
\description{
Filter feature table to only contain one best feature per complex or protein id.
"Best" is defined as the feature with most subunits, highest peak correlation and largest signal area.
}
\examples{
## Load example complex feature finding results:
complexFeatures <- exampleComplexFeatures
## Run summary function:
summarizeFeatures(complexFeatures)
## Filter complex features to only contain one best feature per complex id:
bestComplexFeatures <- getBestFeatures(complexFeatures)
## Run summary function on filtered data:
summarizeFeatures(bestComplexFeatures)
}

