% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R,
%   R/getBestFeatureParameters.R
\name{getBestParameterData}
\alias{getBestParameterData}
\alias{getBestParameterData}
\title{getBestParameterData}
\usage{
getBestParameterData(grid_data, FDR = 0.1)

getBestParameterData(grid_data, FDR = 0.1)
}
\arguments{
\item{grid_data}{list with all filtered grid search results}

\item{FDR}{numeric maximum FDR tat should be considered, default = 0.1}

\item{grid_data}{list with all filtered grid search results}

\item{FDR}{numeric maximum FDR tat should be considered, default = 0.1}
}
\value{
data.table with complex features for best parameter set

data.table with complex features for best parameter set
}
\description{
getBestParameterData.

getBestParameterData.
}
