% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestFeatureParameters.R,
%   R/getBestParameterStats_constraintFDR.R
\name{getBestParameterStats_constraintFDR}
\alias{getBestParameterStats_constraintFDR}
\alias{getBestParameterStats_constraintFDR}
\title{getBestParameterStats_constraintFDR}
\usage{
getBestParameterStats_constraintFDR(complex_feature_grid_filtered,
  complex_hypotheses, n_subsets = 3, FDR_cutoff = 0.1)

getBestParameterStats_constraintFDR(complex_feature_grid_filtered,
  complex_hypotheses, n_subsets = 3, FDR_cutoff = 0.1)
}
\arguments{
\item{complex_feature_grid_filtered}{list with all complex features from grid search}

\item{complex_hypotheses}{data.table with complex hypotheses including decoys}

\item{n_subsets}{numeric number of subsets used for cross-calculation of FDR, default = 3}

\item{FDR_cutoff}{numeric maximum FDR that should be considered, default = 0.1}

\item{complex_feature_grid_filtered}{list with all complex features from grid search}

\item{complex_hypotheses}{data.table with complex hypotheses including decoys}

\item{n_subsets}{numeric number of subsets used for cross-calculation of FDR, default = 3}

\item{FDR_cutoff}{numeric maximum FDR that should be considered, default = 0.1}
}
\value{
data.table one row with best stats across grid search

data.table one row with best stats across grid search
}
\description{
getBestParameterStats_constraintFDR. Do x-fold cross-FDR estimation.
getBestParameterStats_constraintFDR(complex_feature_grid_filtered=var_complex_feature_grid_filtered,complex_hypotheses=var_complex_hypotheses,n_subsets=var_n_subsets,FDR_cutoff=var_FDR_cutoff)

getBestParameterStats_constraintFDR. Do x-fold cross-FDR estimation.
getBestParameterStats_constraintFDR(complex_feature_grid_filtered=var_complex_feature_grid_filtered,complex_hypotheses=var_complex_hypotheses,n_subsets=var_n_subsets,FDR_cutoff=var_FDR_cutoff)
}
