% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinFeatureGridSearch.R
\name{getBestProteinFeatures}
\alias{getBestProteinFeatures}
\title{getBestProteinFeatures}
\usage{
getBestProteinFeatures(res)
}
\arguments{
\item{res}{data.table with protein feature finding results}
}
\value{
List with only one best feature per protein
}
\description{
getBestProteinFeatures.
}
