% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQvalues.R
\name{getBestQvalueParameters}
\alias{getBestQvalueParameters}
\title{getBestQvalueParameters}
\usage{
getBestQvalueParameters(stats, FDR_cutoff = 0.05)
}
\arguments{
\item{stats}{data.table output from qvaluePositivesPlotGrid function}

\item{FDR_cutoff}{numeric between 0 and 1, default = 0.05}
}
\value{
data.table with best stats
}
\description{
getBestQvalueParameters
}

