% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseHypothesisWideFeatures.R
\name{getDistance}
\alias{getDistance}
\title{getDistance}
\usage{
getDistance(A, B)
}
\arguments{
\item{A}{chracter vector}

\item{B}{character vector}
}
\value{
numeric distance of subunit overlap
}
\description{
getDistance.
}
