% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseHypothesisWideFeatures.R
\name{getDistanceMatrix}
\alias{getDistanceMatrix}
\title{getDistanceMatrix}
\usage{
getDistanceMatrix(complexFeatures)
}
\arguments{
\item{complexFeature}{data.table containing filtered complex feature results.}
}
\value{
distance matrix
}
\description{
getDistanceMatrix.
}
