% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{getGenePepList}
\alias{getGenePepList}
\title{Extract matrix of trace intensities for every protein id}
\usage{
getGenePepList(traces)
}
\arguments{
\item{traces}{Object of class traces.}
}
\value{
Gene peptide list
}
\description{
Extract matrix of trace intensities for every protein id
}
