% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGlobalPairwiseCorr.R
\name{getGlobalPairwiseCorr}
\alias{getGlobalPairwiseCorr}
\title{getGlobalPairwiseCorr}
\usage{
getGlobalPairwiseCorr(traces, correlation_cutoff = 0.9, write_csv = TRUE)
}
\arguments{
\item{traces}{An object of type \code{traces}.}

\item{correlation_cutoff}{Minimum pearson correlation value to report an edge between two proteins, default=0.9.}

\item{CSV}{Logical (TRUE/FALSE) whether an output file named GlobalPairwiseCorrelationTable.csv shall be written to the working folder.}
}
\value{
A data.table containing columns peptide/protein_A peptide/protein_B and pearson_correlation, readable e.g. by Cytoscape.
}
\description{
From peptide or protein level Traces, generate a Network table with pairwise profile
correlation score
}
