% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{getIntensityMatrix}
\alias{getIntensityMatrix}
\title{Get intensity matrix from traces object}
\usage{
getIntensityMatrix(traces)
}
\arguments{
\item{traces}{Object of class traces.}
}
\value{
Numeric matrix with intensity values.
}
\description{
Get a matrix of intensity values from a traces object.
}
\examples{
intensityMatrix <- getIntensityMatrix(examplePeptideTraces)
head(intensityMatrix)
}
