% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{getIntensityMatrix.traces}
\alias{getIntensityMatrix.traces}
\title{Get a matrix of fragment intensity values.}
\usage{
getIntensityMatrix.traces(traces.obj)
}
\arguments{
\item{traces.obj}{An object of type \code{traces.obj}.}
}
\value{
A matrix of intensity values.
}
\description{
Get a matrix of fragment intensity values.
}

