% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignTraces.R
\name{getLag}
\alias{getLag}
\title{Get the shift between 2 traces objects via global corellation}
\usage{
getLag(traces1, traces2, min_lag = -5, max_lag = 5, plot = T, ...)
}
\arguments{
\item{traces1}{First traces object to align.}

\item{traces2}{Second traces object to align.}

\item{min_lag}{numeric, the lower bound of the range of lags to evaluate.}

\item{max_lag}{numeric, the upper bound of the range of lags to evaluate.}

\item{plot}{Logical, wether to plot a boxplot of corellation at different lags.}
}
\value{
Numeric, the shift that produces the highest corellation
}
\description{
Get the shift between 2 traces objects via global corellation
}
