% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_proteogenomics.R
\name{getProteogenomicsIds.traces}
\alias{getProteogenomicsIds.traces}
\title{getProteogenomicsIds of traces}
\usage{
\method{getProteogenomicsIds}{traces}(traces)
}
\description{
getProteogenomicsIds of traces
}
