% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_proteogenomics.R
\name{getProteogenomicsIds.tracesList}
\alias{getProteogenomicsIds.tracesList}
\title{getProteogenomicsIds of tracesList object}
\usage{
\method{getProteogenomicsIds}{tracesList}(tracesList)
}
\description{
getProteogenomicsIds of tracesList object
}

