% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromOpenMSLFQ.R
\name{importFromOpenMSLFQ}
\alias{importFromOpenMSLFQ}
\title{Import peptide profiles from an OpenMS MS1-LFQ DDA data analysis result
table.}
\usage{
importFromOpenMSLFQ(file.name = "OpenMS_LFQ_peptides.csv",
  annotation.table = "OpenMS_LFQ_annotation.txt")
}
\arguments{
\item{file.name}{Quantitative MS data result file. Defaults to
'OpenMS_LFQ_peptides.csv'}

\item{annotation.table}{Annotation table containing columns 'Name'
(-> raw file name), and 'Fraction'. Fraction should contain the
fraction_number.}
}
\value{
An object of class Traces (peptide level) containing a
    'traces.wide', and 'ids' table that can be processed with the
    herein contained functions.
}
\description{
Import peptide profiles from an OpenMS MS1-LFQ DDA data analysis result
table.
}

