% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromOpenSWATH.R
\name{importFromOpenSWATH}
\alias{importFromOpenSWATH}
\title{Import peptide profiles from an OpenSWATH experiment.}
\usage{
importFromOpenSWATH(data = "OpenSwathData.tsv",
  annotation.table = "annotation.txt", remove_requantified = TRUE,
  MS1Quant = FALSE, rm.decoy = FALSE)
}
\arguments{
\item{data}{Quantitative MS data in form of OpenSWATH result file or R data.table.}

\item{annotation.table}{path to tab-separated .txt file containing columns `filename` and
`fraction_number` that map the file names (occuring in input table filename column)
`to a SEC elution fraction.}

\item{remove_requantified}{Whether requantified (noise) peak group quantities
(as indicated by m_score = 2) should be removed, defaults to TRUE.}
}
\value{
An object of class Traces containing 
    "traces", "traces_type", "traces_annotation" and "fraction_annotation" list entries
    that can be processed with the herein contained functions.
}
\description{
Import peptide profiles from an OpenSWATH experiment.
}

