% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPCPdata.R
\name{importPCPdata}
\alias{importPCPdata}
\title{Import peptide- or protein-level quantitative PCP matrix.}
\usage{
importPCPdata(input_data, fraction_annotation, rm_decoys = FALSE)
}
\arguments{
\item{input_data}{data.table containing quantitative PCP-MS data. The table can be in long or
wide format and has several required columns (also see examplePCPdataWide or examplePCPdataLong).
Required columns are: protein_id, filename and intensity for long tables and
protein_id plus one column per filename for wide tables.}

\item{fraction_annotation}{Character string specifying a filename or R data.table
containing columns \code{filename} and \code{fraction_number} that map the file names
to a SEC elution fraction (alse see exampleFractionAnnotation).}

\item{rm_decoys}{Logical, whether decoys should be removed
(decoys are found by greping "DECOY" in the protein_id column), defaults to \code{FALSE}.}
}
\value{
An object of class traces containing
    "traces", "traces_type", "traces_annotation" and "fraction_annotation" list entries
    that can be processed with the herein contained functions.
}
\description{
This function takes a quantitative matrix of peptide- or protein-level intensities
generated by a typical PCP experiment and coerces it into a \code{traces} object.
}
\examples{
pcpData <- examplePCPdataWide
## or
pcpData <- examplePCPdataLong
annotation <- exampleFractionAnnotation

traces <- importPCPdata(input_data = pcpData,
                       fraction_annotation = annotation)

summary(traces)
}
