% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeatures.R
\name{imputeMissingValues}
\alias{imputeMissingValues}
\title{imputeMissingValues}
\usage{
imputeMissingValues(intensityMatrix, perturbation_cutoff)
}
\arguments{
\item{intensityMatrix}{matrix with intensties coming from \code{getIntensityMatrix(traces)}.}

\item{perturbation_cutoff}{Numeric, the quantile to use in estimating the perturbation level, default="5%".
Intensity values that are zero are replaced with random values that are
below the specified quantile of the input values. Alternatively a
cutoff value can be specified as an upper limit for perturbation values.
This is nescessary for correlation calculation.}
}
\description{
A helper function to compute a value for all missing values 
to be able to calculate correlations.
}
