% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeOutlierFractions.R
\name{imputeOutlierFractions}
\alias{imputeOutlierFractions}
\alias{imputeOutlierFractions.traces}
\alias{imputeOutlierFractions.tracesList}
\title{Impute outlier fractions}
\usage{
imputeOutlierFractions(traces, bound_left = 2, bound_right = 2,
  consider_borders = TRUE, method = c("mean", "spline"))

\method{imputeOutlierFractions}{traces}(traces, bound_left = 2,
  bound_right = 2, consider_borders = TRUE, method = c("mean", "spline"),
  ...)

\method{imputeOutlierFractions}{tracesList}(tracesList, bound_left = 2,
  bound_right = 2, consider_borders = TRUE, method = c("mean", "spline"),
  ...)
}
\arguments{
\item{traces}{An object of type traces. These should be the raw, unprocessed traces.}

\item{bound_left}{Numeric integer, the minimum number of non-zero values to the
left of a missing value to be replaced with \code{NA}.}

\item{bound_right}{Numeric integer, the minimum number of non-zero values to the
right of a missing value to be replaced with \code{NA}.}

\item{consider_borders}{Logical, whether to find missin values in the first/last fractions
(e.g. \code{0-1-1 -> NA-1-1} and \code{1-0-1-1 -> 1-NA-1-1} if the leftmost value is the
first fraction of the traces object).}

\item{method}{Character string, which interpolation method to use.}
}
\value{
Numeric matrix with intensity values.
}
\description{
Imputes values fo fractions that were labeled as outliers by getFractionInformation.
}
\details{
Unlike with standard imputation methods, missing values on the borders
are linearly extrapolated from the neighboring 2 values. Any imputed value below 0
is set to 0.
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Impute outlier fractions in traces object

\item \code{tracesList}: Impute outlier fractions in multiple traces objects
}}

\examples{
intensityMatrix <- getIntensityMatrix(examplePeptideTraces)
head(intensityMatrix)
}
