% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateReplicates.R
\name{integrateReplicates}
\alias{integrateReplicates}
\title{Generate a consensus traces object from multiple replicate traces}
\usage{
integrateReplicates(traces, design_matrix, integrate_within = NULL,
  filter_by_RepPepCorr = TRUE, repPepCorr_cutoff = 0.5)
}
\arguments{
\item{traces}{object of class tracesList.}

\item{design_matrix}{data.table, design matrix describing the architecture of the tracesList object.}

\item{integrate_within}{Character string, Condition to collapse 
(must be a valid column of the design matrix). If \code{NULL}, everything is collapsed.}

\item{filter_by_RepPepCorr}{Logical, wether to apply a RepPepCorr filter.}

\item{repPepCorr_cutoff}{Numeric, the cutoff below which traces are removed.}
}
\value{
tracesList object containing the integrated traces
}
\description{
Generate a consensus traces object from multiple replicate traces
}
