% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateTraces.R
\name{integrateTraceIntensities}
\alias{integrateTraceIntensities}
\title{Add or average traces objects to generate consensus traces.}
\usage{
integrateTraceIntensities(traces, min_present = 1, aggr_fun = c("sum",
  "mean"))
}
\arguments{
\item{traces}{object of class tracesList.
(must be a valid column of the design matrix). If \code{NULL}, everything is collapsed.}

\item{min_present}{Numeric, the minimum number of samples in which a trace must be detected.}

\item{aggr_fun}{Character string, the method of aggregating the traces.
Must be one of c("sum", "mean").}
}
\value{
tracesList object containing the integrated traces
}
\description{
Add or average traces objects to generate consensus traces.
}

