% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{iterativeMaxCorrFilter}
\alias{iterativeMaxCorrFilter}
\title{Filter iteratively by maxCorr until all outliers are removed}
\usage{
iterativeMaxCorrFilter(traces, cutoff = 0.85, plot = FALSE, PDF = FALSE,
  name = "maxCorrHist", ...)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{cutoff}{Numeric between 0 and 1. Minimum correlation of a peptide
width any sibling peptide.}

\item{plot}{logical,wether to print SibPepCorr density plot to R console.
Deafult is \code{FALSE}.}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file.
Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if
'\code{PDF=TRUE}.PDF file is saved under name.pdf. Default is "maxCorrHist".}
}
\value{
Object of class traces filtered for peptide correlation.
}
\description{
Filter iteratively by maxCorr until all outliers are removed
}

