% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEstimation.R
\name{manualFeatureMapping}
\alias{manualFeatureMapping}
\title{manualFeatureMapping}
\usage{
manualFeatureMapping(feature, dist_cutoff, manual_annotation)
}
\arguments{
\item{feature}{data.table containing filtered complex feature results.}

\item{dist_cutoff}{numeric maximal apex distance for features to be mapped}

\item{manual_annotation}{data.table containing manual annotation}
}
\value{
data.table with mached detected and manual features
}
\description{
manualFeatureMapping
}
