% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapProtToComplexFeatures.R
\name{mapProtToComplexFeatures}
\alias{mapProtToComplexFeatures}
\title{Map protein features to complex features}
\usage{
mapProtToComplexFeatures(
  proteinFeatures,
  complexFeatures,
  prot_id = "protein_id",
  max_apex_dist = 2,
  plot = F,
  PDF = F
)
}
\arguments{
\item{proteinFeatures}{data.table with protein co-elution features in the
format of the output from \code{proteinFeatureFinding} or
\code{aggregatePeptideTests}.}

\item{complexFeatures}{data.table with complex co-elution features in the
format of the output from \code{complexFeatureFinding}.}

\item{prot_id}{chracter string specifying the column name of the id type
in the proteinFeatures data.table to match, default is "protein_id".}

\item{max_apex_dist}{numeric specifying the minimum number of fractions
between a protein and complex co-elution feature to be mapped 2, default is 2.}

\item{plot}{Logical if histogram of apex distances between protein and
complex co-elution features should be plotted, default is FALSE.}

\item{PDF}{Logical if plot should be written to PDF, default is FALSE.}
}
\value{
data.table with protein co-elution features that have a matching
complex co-elution feature.
}
\description{
Map co-elution features from the protein centric feature finding
to co-elution features from the complex-centric workflow.
}
\examples{
# Load some example data:
proteinFeatures <- exampleProteinFeatures
complexFeatures <- exampleComplexFeatures
# Run the mapping:
mappedFeatures <- mapProtToComplexFeatures(proteinFeatures,
                                           complexFeatures,
                                           prot_id = "protein_id",
                                           max_apex_dist = 2
                                           )
# Inspect mapping result:
mappedFeatures
}
