% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeByCyclicLoess.R
\name{normalizeByCyclicLoess}
\alias{normalizeByCyclicLoess}
\title{#' #' Cyclic loess normalization}
\usage{
normalizeByCyclicLoess(traces_list, window = 3, step = 1,
  plot = TRUE, PDF = TRUE, name = "normalizeByCyclicLoess")
}
\arguments{
\item{traces_list}{traces list object of type peptide}

\item{window}{integer size of slding window, default = 3}

\item{step}{integer steps for sliding window, default = 1}

\item{plot}{logical, default = TRUE}

\item{PDF}{logical, defalt = FALSE}

\item{name}{character strimg specifying pdf name, default = "normalizeByCyclicLoess"}
}
\description{
Normalize peptide traces across replicates, conditions and fractions
}
