% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeTraces.R
\name{normalizeToStandard}
\alias{normalizeToStandard}
\alias{normalizeToStandard.traces}
\alias{normalizeToStandard.tracesList}
\title{Normalize trace intensities to reference traces}
\usage{
normalizeToStandard(
  traces,
  standard_ids,
  method = "median",
  filter_complete = TRUE,
  top_n = NULL
)

\method{normalizeToStandard}{traces}(
  traces,
  standard_ids,
  method = "median",
  filter_complete = TRUE,
  top_n = NULL,
  ...
)

\method{normalizeToStandard}{tracesList}(
  tracesList,
  standard_ids,
  method = "median",
  filter_complete = TRUE,
  top_n = NULL,
  ...
)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{standard_ids}{Character vector, peptide or protein ids of reference peptides}

\item{method}{Character string, whether to normalize to the median or the mean intensity 
of the standard.}
}
\value{
Object of class traces with normalized intensities.
}
\description{
Normalize trace intensities to reference traces
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Normalize to reference traces in multiple traces objects

\item \code{tracesList}: Normalize to reference traces in multiple traces objects
}}

