% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R
\name{performComplexGridSearch}
\alias{performComplexGridSearch}
\title{Perform complex feature grid search}
\usage{
performComplexGridSearch(traces, complex_hypothesis, calibration,
  corrs = c(0.5, 0.75, 0.9, 0.95), windows = c(8, 10, 12),
  smoothing = c(7, 9, 11), rt_heights = c(3, 4, 5), parallelized = TRUE,
  n_cores = 1)
}
\arguments{
\item{traces}{traces object of type protein}

\item{calibration}{list of two functions for calibration}

\item{corrs}{numeric vector}

\item{windows}{numeric vector}

\item{smoothing}{numeric vector}

\item{rt_heights}{numeric vector}

\item{parallelized}{logical default=TRUE}

\item{n_cores}{numeric number of cores to use if parallelized (default=1)}
}
\value{
List with stats
}
\description{
Perform complex feature grid search.
}
