% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.complexFeatures.R
\name{plot.complexFeatures}
\alias{plot.complexFeatures}
\title{Plot the result of the sliding window algorithm.}
\usage{
\method{plot}{complexFeatures}(res, traces.obj, complexID, calibration,
  pp = FALSE, sw = FALSE, estimate = FALSE, monomerMW = FALSE,
  log = FALSE)
}
\arguments{
\item{sw.result}{An object of type 'complexFeaturesSW'.}

\item{trace.mat}{The same matrix that was passend to
\code{findComplexFeatures}.}

\item{protein.names}{The names of proteins whose traces are contained in
the matrix \code{trace.mat}.}

\item{n.largest}{Only plot the n largest subgroups. Optional.}
}
\description{
Plot the result of the sliding window algorithm.
}

