% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexFeatureDetection.R
\name{plot.complexFeaturesSW}
\alias{plot.complexFeaturesSW}
\title{Plot the result of the slidingWindow algorithm.}
\usage{
\method{plot}{complexFeaturesSW}(sw.result, protein.traces.long,
  n.largest = NULL, log = FALSE)
}
\description{
Plot the result of the slidingWindow algorithm.
}

