% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.proteinFeatures.R
\name{plot.proteinFeatures}
\alias{plot.proteinFeatures}
\title{Plot protein features.}
\usage{
\method{plot}{proteinFeatures}(res, traces.obj, proteinID, plot_peak = TRUE,
  plot_monomer = TRUE, log = FALSE, plot_apex = TRUE,
  plot_in_complex_estimate = TRUE)
}
\arguments{
\item{res}{proteinFeatures}

\item{proteinID}{protein ID}

\item{plot_peak}{logical}

\item{plot_monomer}{logical}

\item{pepTraces}{An object of type \code{traces.obj}.}

\item{window.size}{Size of the window. Numeric.}
}
\description{
Plot protein features.
}

