% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SibPepCorrDensities.R
\name{plot.SibPepCorrDensities}
\alias{plot.SibPepCorrDensities}
\title{plot.SibPepCorrDensities}
\usage{
\method{plot}{SibPepCorrDensities}(traces, PDF = FALSE)
}
\arguments{
\item{traces}{An object of type \code{traces.obj}.}

\item{PDF}{logical TRUE or FALSE}
}
\value{
Plot.
}
\description{
Plot sibling peptide correlation in traces.object of type peptide.
}

