% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{plot.tracesList}
\alias{plot.tracesList}
\title{Plot traces}
\usage{
\method{plot}{tracesList}(
  traces,
  design_matrix = NULL,
  collapse_conditions = FALSE,
  aggregateReplicates = FALSE,
  log = FALSE,
  legend = TRUE,
  PDF = FALSE,
  name = "Traces",
  plot = TRUE,
  isoformAnnotation = FALSE,
  colour_by = "id",
  highlight = NULL,
  highlight_col = NULL,
  colorMap = NULL,
  monomer_MW = TRUE
)
}
\arguments{
\item{traces}{Object of class traces.}

\item{log}{Logical, whether the intensities should be plotted in log scale. Default is \code{FALSE}.}

\item{legend}{Logical, whether a legend of the traces should be plotted.
Should be set to \code{FALSE}
if many chromatograms are plotted. Default is \code{TRUE}.}

\item{PDF}{Logical, whether to plot to PDF. PDF file is saved in working directory.
Default is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if \code{PDF=TRUE}.
PDF file is saved under name.pdf. Default is "Traces".}

\item{plot}{Logical, wether to print or return the plot object}

\item{isoformAnnotation}{Logical, wether to colour traces by their isoform annotation}

\item{colour_by}{Character string specifying by which column to colour by. Default is id.}

\item{highlight}{Character vector, ids of the traces to highlight (can be multiple).
Default is \code{NULL}.}

\item{highlight_col}{Character string, A color to highlight traces in. Must be accepted by ggplot2.}

\item{colorMap}{named character vector containing valid color specifications for plotting.
The names of the vector must correspond to the ids of the peptides to be plotted.}

\item{monomer_MW}{Logical if monomer MWs should be indicated}
}
\description{
Plot all chromatograms in a traces object. Most generic plotting function.
}
