% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexFeatureSummary.R
\name{plotComplexCompletenessScatter}
\alias{plotComplexCompletenessScatter}
\title{plotComplexCompletenessScatter}
\usage{
plotComplexCompletenessScatter(complexFeatures, PDF = TRUE)
}
\arguments{
\item{complexFeatures}{data.table with detected complexes}

\item{PDF}{logical default = TRUE}
}
\value{
plots summarizing the completeness
}
\description{
plotComplexCompletenessScatter
}
