% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarizedMScoverage.R
\name{plotComplexMScoveragePie}
\alias{plotComplexMScoveragePie}
\title{plotComplexMScoveragePie.}
\usage{
plotComplexMScoveragePie(complexMScoverageSummary, PDF = TRUE)
}
\arguments{
\item{complexMScoverageSummary}{data.table with summary}

\item{PDF}{logical default = TRUE}
}
\value{
plots summarizing the MS coverage
}
\description{
plotComplexMScoveragePie
}
