% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarizedMScoverage.R
\name{plotComplexMScoverageScatter}
\alias{plotComplexMScoverageScatter}
\title{plotComplexMScoverageScatter.}
\usage{
plotComplexMScoverageScatter(
  complexMScoverage,
  PDF = FALSE,
  name_suffix = "complexHypotheses"
)
}
\arguments{
\item{complexMScoverage}{data.table with summary}

\item{PDF}{logical default = TRUE}
}
\value{
plots summarizing the MS coverage
}
\description{
plotComplexMScoverageScatter
}
