% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexFeatureSummary.R
\name{plotComplexSubfeatureSummary}
\alias{plotComplexSubfeatureSummary}
\title{plotComplexSubfeatureSummary}
\usage{
plotComplexSubfeatureSummary(complexFeatures, PDF = TRUE)
}
\arguments{
\item{complexFeatures}{data.table with detected complexes}

\item{PDF}{logical default = TRUE}
}
\value{
plots summarizing the complex subfeatures
}
\description{
plotComplexSubfeatureSummary
}
