% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatures.R
\name{plotFeatures}
\alias{plotFeatures}
\title{#' Plot the result of the sliding window algorithm.}
\usage{
plotFeatures(feature_table, traces, feature_id, calibration = NULL,
  annotation_label = "protein_id", onlyBest = TRUE, apex = TRUE,
  peak_area = FALSE, sliding_window_area = FALSE,
  estimated_complex_MW = FALSE, highlight = NULL, highlight_col = NULL,
  monomer_MW = FALSE, log = FALSE, legend = TRUE, PDF = FALSE,
  name = "Traces", colorMap = NULL)
}
\arguments{
\item{feature_table}{data.table}

\item{feature_id}{Character string specifying complex_id}

\item{annotation_label}{Character string specifying column name in trace annotation to use for trace labeling, default is protein_id (uniprot id).}

\item{peak_area}{Logical if selected peak area should be highlighted.}

\item{estimated_complex_MW}{Logical if estimated complex MW should be indicated}

\item{monomer_MW}{Logical if monomer MWs should be indicated}

\item{log}{Logical if intensities should be log transformed, default is \code{fFALSE}.}

\item{colorMap}{named character vector containing valid color specifications for plotting.
The names of the vector must correspond to the ids of the peptides to be plotted.}

\item{proteinTraces}{traces object of type protein}

\item{sliding_window_are}{Logical if original sliding_window area should be highlighted}
}
\description{
Chomatographic visualization
}
\examples{

#------------------------
## Complex level plotting
#------------------------

## Load example data
featureTable = exampleComplexFeatures
proteinTraces = exampleProteinTraces
complexId = "2174-1" # Complex of interest

## Plot all features found for this Protein
plotFeatures(feature_table = featureTable,
                    traces = proteinTraces,
                    feature_id = complexId,
                    peak_area = TRUE,
                    onlyBest = FALSE)

#------------------------
## Protein level plotting
#------------------------

## Load example data
 featureTable = exampleProteinFeatures
 peptideTraces = examplePeptideTraces
 proteinId = "P61201" # Protein of interest

 ## Plot all features found for this Protein
 plotFeatures(feature_table = featureTable,
                     traces = peptideTraces,
                     feature_id = proteinId,
                     peak_area = TRUE,
                     onlyBest = FALSE,
                     legend = FALSE)

}
