% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatures.R
\name{plotFeatures.tracesList}
\alias{plotFeatures.tracesList}
\title{plotFeatures.tracesList}
\usage{
\method{plotFeatures}{tracesList}(feature_table, traces, feature_id,
  design_matrix = NULL, calibration = NULL,
  annotation_label = "protein_id", onlyBest = TRUE, apex = TRUE,
  peak_area = FALSE, sliding_window_area = FALSE,
  estimated_complex_MW = FALSE, highlight = NULL, highlight_col = NULL,
  colour_by = "id", monomer_MW = FALSE, log = FALSE, legend = TRUE,
  PDF = FALSE, name = "Traces", colorMap = NULL)
}
\arguments{
\item{feature_table}{data.table}

\item{traces}{traces object of type peptide or protein.}

\item{feature_id}{Character string specifying complex_id}

\item{design_matrix}{data.table with design matrix}

\item{calibration}{calibration function}

\item{annotation_label}{Character string specifying column name in trace annotation to use for trace labeling, default is protein_id (uniprot id).}

\item{onlyBest}{logical true or false}

\item{apex}{logical true or false}

\item{peak_area}{Logical if selected peak area should be highlighted.}

\item{sliding_window_area}{Logical if original sliding_window area should be highlighted}

\item{estimated_complex_MW}{Logical if estimated complex MW should be indicated}

\item{monomer_MW}{Logical if monomer MWs should be indicated}

\item{log}{Logical if intensities should be log transformed, default is \code{fFALSE}.}

\item{colorMap}{named character vector containing valid color specifications for plotting.
The names of the vector must correspond to the ids of the peptides to be plotted.}

\item{peak_area}{logical true or false}
}
\description{
plotFeatures.tracesList
}
