% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProteogenomics.R
\name{plotGeneModel}
\alias{plotGeneModel}
\title{Plot the known transcripts for a gene of interest}
\usage{
plotGeneModel(ensdb, geneId, filter = ~gene_id == geneId & tx_biotype ==
  "protein_coding", highlight = NULL)
}
\arguments{
\item{ensdb}{EnsDb object or path to EnsDb database. If not NULL, all transcripts
for the specified geneId will be plotted.}

\item{geneId}{character, the geneId of the protein to be plotted}

\item{filter}{formula, Filtering expression to be passed to AnnotationFilter}
}
\description{
Plot the known transcripts for a gene of interest
}
