% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTraces.R
\name{plotGlobalIntensities}
\alias{plotGlobalIntensities}
\title{Plot a comparison of the intensities in different samples}
\usage{
plotGlobalIntensities(traces, plot = T, PDF = F,
  name = "IntensitySummary.pdf")
}
\arguments{
\item{traces}{Object of class tracesList.}

\item{plot}{Logical, wether to print or return the plot object}

\item{PDF}{Logical, whether to plot to PDF. PDF file is saved in working directory.
Default is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if \code{PDF=TRUE}.
PDF file is saved under name.pdf. Default is "Traces".}
}
\value{
Violin plot of trace intensities per sample
}
\description{
Plot a comparison of the intensities in different samples
}

