% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R,
%   R/getBestFeatureParameters.R
\name{plotIdFDRspace}
\alias{plotIdFDRspace}
\alias{plotIdFDRspace}
\title{Plot FDR gridsearch}
\usage{
plotIdFDRspace(grid_search_stats, level = "complex", id_level = "TP",
  FDR_cutoff = 0.1, colour_parameter = "completeness_cutoff", PDF = TRUE,
  name = "ID_FDR_plot.pdf")

plotIdFDRspace(grid_search_stats, level = "complex", id_level = "TP",
  FDR_cutoff = 0.1, colour_parameter = "completeness_cutoff", PDF = TRUE,
  name = "ID_FDR_plot.pdf")
}
\arguments{
\item{level}{character sting either complex or protein}

\item{complex_features}{data.table containing filtered complex feature results.}

\item{complex_features}{data.table containing filtered complex feature results.}

\item{level}{character sting either complex or protein}
}
\value{
List with stats

List with stats
}
\description{
Perform complex feature grid search.

Perform complex feature grid search.
}
