% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissingValues.R
\name{plotImputationSummary}
\alias{plotImputationSummary}
\title{Plot a summary of the trace imputation}
\usage{
plotImputationSummary(traces, tracesImp, PDF = FALSE,
  plot_traces = TRUE, max_n_traces = 30,
  name = "Missing_value_imputation_summary.pdf")
}
\arguments{
\item{traces}{Object of class traces with missing values indicated by \code{NA}.}

\item{tracesImp}{Object of class traces with imputed missing values.}

\item{PDF}{Logical, whether to plot to PDF. PDF file is saved in working directory.
Default is \code{FALSE}.}

\item{plot_traces}{Logical, whether to plot some traces, visualizing the imputed values.
Default is \code{TRUE}.}

\item{max_n_traces}{Numeric integer, how many traces to plot. Default is 30.}

\item{name}{Character string with name of the plot, only used if \code{PDF=TRUE}.
Default is "Missing_value_imputation_summary.pdf".}
}
\value{
Summary plots of the missing value imputation.
}
\description{
Plot the distribution of missing values per trace as well as
missing values per fraction and some imputed traces.
}
\examples{
tracesMiss <- findMissingValues(examplePeptideTraces)
tracesImp <- imputeMissingVals(tracesMiss)
View(tracesImp$traces)

plotImputationSummary(tracesMiss, tracesImp, PDF = FALSE)
}
