% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotManualFeatureComparison.R
\name{plotManualFeatureComparison}
\alias{plotManualFeatureComparison}
\title{plotManualFeatureComparison}
\usage{
plotManualFeatureComparison(mapped_features, traces, manual_stats, calibration,
  proteinID, plot_peak = TRUE, plot_monomer = TRUE, log = FALSE,
  plot_apex = TRUE, plot_in_complex_estimate = TRUE, TP = FALSE)
}
\arguments{
\item{mapped_features}{proteinFeatures}

\item{traces}{An object of type \code{traces}.}

\item{manual_stats}{Stats of manual annotation.}

\item{calibration}{list of calibraion functions form calibrateSECMW}

\item{proteinID}{protein ID}

\item{plot_peak}{TRUE,}

\item{plot_monomer}{TRUE,}

\item{log}{FALSE,}

\item{plot_apex}{TRUE,}

\item{plot_in_complex_estimate}{TRUE,}

\item{TP}{FALSE}
}
\description{
plotManualFeatureComparison
}
