% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMassDistribution.R
\name{plotMassAssemblyChange}
\alias{plotMassAssemblyChange}
\title{plotMassAssemblyChange between 2 conditions}
\usage{
plotMassAssemblyChange(
  assamblyTest,
  change_cutoff = 0.2,
  name = "massAssemblyChange",
  PDF = FALSE
)
}
\arguments{
\item{assamblyTest}{data.table, a data.table with test statistics.
An assamblyTest can be produced with \code{getMassAssemblyChange}.}

\item{change_cutoff}{Numeric fold change cutoff, default is 0.2.}

\item{name}{character string specifying the name of output if PDF=TRUE, default is "massAssemblyChange".}

\item{PDF}{logical if PDF should be created, default is FALSE.}
}
\value{
plot
}
\description{
plotMassAssemblyChange between 2 conditions
}
