% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot networks}
\usage{
plotNetwork(complex_hypothesis, dt, col_a = "darkblue", col_b = "darkred",
  all_b = FALSE)
}
\arguments{
\item{complex_hypothesis}{data.table with complex hypotheses}

\item{dt}{data.table with binary interactions between a and b}

\item{col_a}{character string specifying the color of verteces in a}

\item{col_b}{character string specifying the color of verteces in b}

\item{all_b}{logical TRUE if all vertices in b should be colored, FALSE to only color those connected to a's with the complex_hypothesis. Example: a = bait and b= prays, if FALSE, only preys that were found by pulling on the baits within the complex hypothesis are colored.}
}
\description{
Plot networks
}
