% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProteogenomics.R
\name{plotPeptidesInGenome}
\alias{plotPeptidesInGenome}
\title{Plot peptides in their genomic context}
\usage{
plotPeptidesInGenome(traces, geneId, ensdb = NULL, colorMap = NULL,
  highlight = NULL, highligt_detected_transcripts = TRUE, ...)
}
\arguments{
\item{traces}{Object of class traces with an additional item =genomic_coord=}

\item{geneId}{character, the geneId of the protein to be plotted}

\item{ensdb}{EnsDb object or path to EnsDb database. If not NULL, all transcripts
for the specified geneId will be plotted.}

\item{colorMap}{named character vector containing valid color specifications for plotting.
The names of the vector must correspond to the ids of the peptides to be plotted.}

\item{...}{Additional arguments to be passed to plot.traces}

\item{plot}{boolean, whether to plot or to return the plot.}
}
\description{
Plot peptides in their genomic context
}

