% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinSequenceCoverage.R
\name{plotPeptidesPerProtein}
\alias{plotPeptidesPerProtein}
\title{Plot a distribution of the prptides per Protein in your peptideTraces object(s)}
\usage{
plotPeptidesPerProtein(pepTraces_list, scaled = F, log = T, cutoff = 50)
}
\arguments{
\item{pepTraces_list}{A list of one or more peptideTraces objects to plot distributions for}

\item{scaled}{\code{boolean}, Wether to draw a scaled density plot or a histogram}

\item{log}{\code{boolean}, Wether to plot a logarithmic x-axis}

\item{cutoff}{\code{numeric}, The limit for the x-axis (peptides per protein)}
}
\value{
A density plot or histogram of the peptides per protein in peptideTraces objects.
}
\description{
Plot a distribution of the prptides per Protein in your peptideTraces object(s)
}
\details{
Can be used to compare peptide coverage before and after filtering, or from different searches.
Provide a list wit all peptideTraces objects to be compared.
}

