% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarizedMScoverage.R
\name{plotProteinMScoveragePie}
\alias{plotProteinMScoveragePie}
\title{plotProteinMScoveragePie.}
\usage{
plotProteinMScoveragePie(proteinMScoverageSummary, PDF = TRUE)
}
\arguments{
\item{proteinMScoverageSummary}{data.table with summary}

\item{PDF}{logical default = TRUE}
}
\value{
plots summarizing the MS coverage
}
\description{
plotProteinMScoveragePie
}
