% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{plotProteoformPvalHist}
\alias{plotProteoformPvalHist}
\title{Plots p-value histograms for the proteoform p-values.}
\usage{
plotProteoformPvalHist(traces, name = "proteoform_pval_histograms", PDF = TRUE)
}
\arguments{
\item{traces}{Object of class traces with proteoform scores and p-values.}

\item{name}{Name of the generated PDF.}

\item{PDF}{logical, wether to print plot to a PDF file.}
}
\value{
plot
}
\description{
Plots p-value histograms for the proteoform p-values.
}
