% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{plotProteoformVolcano}
\alias{plotProteoformVolcano}
\title{Plots pseudo volcano for the proteoform scores and p-values.}
\usage{
plotProteoformVolcano(traces, name = "proteoform_pseudo_volcano", PDF = TRUE)
}
\arguments{
\item{traces}{Object of class traces with proteoform scores and p-values.}

\item{name}{Name of the generated PDF.}

\item{PDF}{logical, wether to print plot to a PDF file.}
}
\value{
plot
}
\description{
Plots pseudo volcano for the proteoform scores and p-values.
}
