% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeTraces.R
\name{plotReference}
\alias{plotReference}
\title{Plot a comparison of the intensities in different samples}
\usage{
plotReference(
  traces,
  design_matrix,
  ref_id = "P00722",
  log = F,
  PDF = TRUE,
  legend = TRUE,
  average_int = "median",
  plot = TRUE
)
}
\arguments{
\item{traces}{Object of class tracesList.}

\item{design_matrix}{data.table, A valid design matrix describing the tracesList object.}

\item{ref_id}{Character string, the name(s) of proteins/peptides to plot.}

\item{log}{Logical, wether to plot on log scale.}

\item{PDF}{Logical, whether to plot to PDF. PDF file is saved in working directory.
Default is \code{FALSE}.}

\item{legend}{Logical, wether to plot a legend.}

\item{average_int}{Character string, One of mean or median. Which summary to plot.}

\item{plot}{Logical, wether to print or return the plot object}

\item{name}{Character string with name of the plot, only used if \code{PDF=TRUE}.
PDF file is saved under name.pdf. Default is "Traces".}
}
\value{
A plot of the reference traces with a mean/median trace of both the
reference and all traces.
}
\description{
Plot a comparison of the intensities in different samples
}
