% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRepPepCorr.R
\name{plotRepPepCorrDensities}
\alias{plotRepPepCorrDensities}
\title{plot RepPepCorrDensities}
\usage{
plotRepPepCorrDensities(traces, PDF = FALSE,
  name = "ReplicatePeptideCorrelationDensity")
}
\arguments{
\item{traces}{An object of type traces.}

\item{PDF}{logical, wether to print RepPepCorr density plot to a PDF file. Deafaults to \code{FALSE}.}
}
\value{
Plot.
}
\description{
Plot replicate peptide correlation in traces object of type peptide.
}

