% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQvalues.R
\name{plotScoreDistribution}
\alias{plotScoreDistribution}
\title{plotScoreDistribution}
\usage{
plotScoreDistribution(features, PDF = TRUE, name = "scoreDistribution")
}
\arguments{
\item{features}{data.table with complex or protein features including qvalues}

\item{PDF}{logical, defalt = FALSE}

\item{name}{character strimg specifying pdf name, default = "scoreDistribution"}
}
\description{
plotScoreDistribution
}
