% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinSequenceCoverage.R
\name{plotSequenceCoverage}
\alias{plotSequenceCoverage}
\title{Plot a distribution of Sequence Coverage in your peptideTraces object(s)}
\usage{
plotSequenceCoverage(pepTraces_list, proteinFasta, scaled = TRUE,
  log = FALSE)
}
\arguments{
\item{pepTraces_list}{A list of one or more peptideTraces objects to plot distributions for}

\item{proteinFasta}{Path to a fasta file with Protein Sequences and headers that contain
the Protein names in the protein_id column of your peptideTraces object. Usually this will be the
fasta file used for library generation in the proteomics search.}

\item{scaled}{\code{boolean}, Wether to draw a scaled density plot or a histogram}

\item{log}{\code{boolean}, Wether to plot a logarithmic x-axis}
}
\value{
A density plot or histogram of the peptides per protein in peptideTraces objects.
}
\description{
Plot a distribution of Sequence Coverage in your peptideTraces object(s)
}
\details{
Can be used to compare peptide coverage before and after filtering, or from different searches.
Provide a list wit all peptideTraces objects to be compared.
}

