% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexFeatureSummary.R
\name{plotSummarizedComplexes}
\alias{plotSummarizedComplexes}
\title{plotSummarizedComplexes}
\usage{
plotSummarizedComplexes(complexFeatures, hypotheses, protTraces, PDF = TRUE)
}
\arguments{
\item{complexFeatures}{data.table with detected complexes}

\item{hypotheses}{data.table with complex hypotheses}

\item{PDF}{logical default = TRUE}
}
\value{
plots summarizing the complex subfeatures
}
\description{
plotSummarizedComplexes
}
