% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarizedMScoverage.R
\name{plotSummarizedMScoverage}
\alias{plotSummarizedMScoverage}
\title{Plot MS coverage of the hypotheses in the PCP data}
\usage{
plotSummarizedMScoverage(hypotheses, protTraces, PDF = FALSE)
}
\arguments{
\item{hypotheses}{data.table containing complex hypotheses.
Should have the following columns:
\itemize{
\item complex_id: character strings, a unique id for every complex
\item complex_name: character strings, a unique name for every complex
\item protein_id: character strings, the protein id, e.g. Uniprot id
}}

\item{protTraces}{An object of class traces, type="protein".}

\item{PDF}{Logical whether to save plots in PDF, default = \code{FALSE}.}
}
\value{
Multiple plots summarizing the MS coverage.
}
\description{
Generate plots that describe the coverage of the complex hypotheses in the PCP data.
}

