% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDifferentialExperssion.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot for differential expression of peptides, proteins or complexes between 2 conditions}
\usage{
plotVolcano(testResults, highlight = NULL, FC_cutoff = 2,
  pBHadj_cutoff = 0.01, name = "volcanoPlot", PDF = FALSE,
  level = c("feature", "global"))
}
\arguments{
\item{testResults}{data.table, a data.table with test statistics.
A testResults can be produced with \code{testDifferentialExpression}.}

\item{highlight}{character string of feature_id that should be highlighted. Default = NULL.}

\item{FC_cutoff}{Numeric fold change cutoff, default is 2.}

\item{pBHadj_cutoff}{Numeric p-value cutoff, default is 0.01.}

\item{name}{character string specifying the name of output if PDF=TRUE, default is "volcanoPlot".}

\item{PDF}{logical if PDF should be created, default is FALSE.}
}
\value{
plot
}
\description{
Volcano plot for differential expression of peptides, proteins or complexes between 2 conditions
}
