% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein.mw.conc}
\alias{protein.mw.conc}
\title{A data.table that holds stores for each protein its molecular weight as
annotated in UniProt as well as an estimate on its absolute abundance within
the sample.
Its structure is as follows:
\itemize{
 \item \code{protein_id} The Uniprot identifier of the protein.
 \item \code{protein_mw} The molecular weight.
 \item \code{protein_concentration} The absolute concentration.
}}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 7304 rows and 3 columns.}
\usage{
protein.mw.conc
}
\description{
A data.table that holds stores for each protein its molecular weight as
annotated in UniProt as well as an estimate on its absolute abundance within
the sample.
Its structure is as follows:
\itemize{
 \item \code{protein_id} The Uniprot identifier of the protein.
 \item \code{protein_mw} The molecular weight.
 \item \code{protein_concentration} The absolute concentration.
}
}
\keyword{datasets}

