% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein.traces.obj}
\alias{protein.traces.obj}
\title{An example object of class \code{FragmentTraces}.}
\format{An object of class \code{FragmentTraces} of length 4.}
\usage{
protein.traces.obj
}
\description{
An example object of class \code{FragmentTraces}.
}
\keyword{datasets}

