% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein.traces}
\alias{protein.traces}
\title{Some protein traces of e4 that were filtered. TESTING DATA.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 5009 rows and 82 columns.}
\usage{
protein.traces
}
\description{
Some protein traces of e4 that were filtered. TESTING DATA.
}
\keyword{datasets}

