% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinQuantification.R
\name{proteinQuantification}
\alias{proteinQuantification}
\title{proteinQuantification}
\usage{
proteinQuantification(traces, topN = 2, keep_less = FALSE,
  rm_decoys = TRUE)
}
\arguments{
\item{traces}{An object of type \code{traces.obj}, trace_type is peptide}

\item{topN}{numeric specifying the number of peptides to use for protein quantification}

\item{keep_less}{logical specifying wheather peptides with less than topN peptides should be kept in the data, default is FALSE}

\item{rm_decoys}{logical specifying wheather decoys should be kept. The decoys have only limited use on the protein level, default is TRUE}
}
\value{
An object of type \code{traces.obj}, trace_type is protein
}
\description{
Calculate protein quantities basen on the topN peptide intensities.
}

