% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteogenomics.R
\name{proteinToGenomeFast}
\alias{proteinToGenomeFast}
\title{Fetch genomic coordinates of peptides more efficiently}
\usage{
proteinToGenomeFast(x, db, id = "name", idType = "protein_id")
}
\value{
A list of genomic ranges containing the genomic coordinates of each peptide.
}
\description{
Fetch genomic coordinates of peptides more efficiently
}
\details{
This is an adapted function of the ensembldb package. It speeds up
calculations by grouping peptides by proteins and only fetching each CDS once per
protein.
}

