% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{prune_tree}
\alias{prune_tree}
\title{Remove any single peptides that are on the outside of the tree}
\usage{
prune_tree(tree, clusters, pruning_cutoff = 0.3)
}
\arguments{
\item{tree}{hclust object}

\item{clusters}{vector of cluster assignments (as produced by cutree)}

\item{pruning_cutoff}{int, height up to which single peptides are called outliers}
}
\description{
Remove any single peptides that are on the outside of the tree
}
