% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQvalues.R
\name{qvaluePositivesPlot}
\alias{qvaluePositivesPlot}
\title{qvaluePositivesPlot}
\usage{
qvaluePositivesPlot(features, plot = TRUE, PDF = FALSE,
  name = "qvaluePositivesPlot")
}
\arguments{
\item{features}{data.table with complex or protein features including qvalues}

\item{plot}{logical, default = TRUE}

\item{PDF}{logical, defalt = FALSE}

\item{name}{character strimg specifying pdf name, default = "qvaluePositivesPlot"}
}
\value{
data.table with stats
}
\description{
qvaluePositivesPlot
}
