% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQvalues.R
\name{qvaluePositivesPlotGrid}
\alias{qvaluePositivesPlotGrid}
\title{qvaluePositivesPlotGrid}
\usage{
qvaluePositivesPlotGrid(featuresGrid, plot = TRUE, PDF = FALSE,
  name = "qvaluePositivesPlotGrid")
}
\arguments{
\item{featuresGrid}{list of complex or protein feature data.tables including qvalues}

\item{plot}{logical, default = TRUE}

\item{PDF}{logical, defalt = FALSE}

\item{name}{character strimg specifying pdf name, default = "qvaluePositivesPlotGrid"}
}
\value{
data.table with stats
}
\description{
qvaluePositivesPlotGrid
}
