% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveProteoformSpecificFeatures.R
\name{refineProteoformsByDetectedFeatures}
\alias{refineProteoformsByDetectedFeatures}
\title{Refine annotated proteoforms in traces and features based on detected co-elution signals}
\usage{
refineProteoformsByDetectedFeatures(traces, features)
}
\arguments{
\item{traces}{Object of class traces.}

\item{features}{Protein features as generated by \code{findProteinFeatures}.}
}
\value{
List containing 1. refined traces and 2. refinded features.
}
\description{
Refine annotated proteoforms in traces and features based on detected co-elution signals
}
