% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEstimation.R
\name{resolveDoubleAssignments}
\alias{resolveDoubleAssignments}
\title{resolveDoubleAssignments}
\usage{
resolveDoubleAssignments(id, mapped_features)
}
\arguments{
\item{id}{feature_id}

\item{mapped_features}{data.table with all mapped features}
}
\value{
data.table with mached detected and manual features
}
\description{
resolveDoubleAssignments
}
