% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveProteoformSpecificFeatures.R
\name{resolveProteoformSpecificFeatures}
\alias{resolveProteoformSpecificFeatures}
\title{Resolve protein features based on annotated proteoforms}
\usage{
resolveProteoformSpecificFeatures(
  features,
  traces,
  minProteoformIntensityRatio = 0.1,
  perturb_cutoff = "5\%"
)
}
\arguments{
\item{features}{Protein features as generated by \code{findProteinFeatures}.}

\item{traces}{Object of class traces.}

\item{minProteoformIntensityRatio}{Value between 0 and 1. Default = 0.1.}

\item{perturb_cutoff}{percentage. Default = "5%".}
}
\value{
Protein features resolved by proteoforms.
}
\description{
Resolve protein features based on annotated proteoforms
}
