% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{robust_clustering}
\alias{robust_clustering}
\title{Remove outlier peptides and take the silhouette width into account}
\usage{
robust_clustering(
  traces,
  pruning_cutoff = 0.3,
  initial_cut_height = 0.3,
  method = c("complete", "single", "average"),
  PDF = F,
  name = "robust_clustering"
)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{pruning_cutoff}{int, height up to which single peptides are called outliers}

\item{initial_cut_height}{int, height at which tree is initially cut}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file.
Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if
'\code{PDF=TRUE}.PDF file is saved under name.pdf. Default is "robust_clustering".}
}
\value{
tracesList including cluster information in trace_annotation
}
\description{
Remove outlier peptides and take the silhouette width into account
}
