% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.SibPepCorr.R
\name{ROC.SibPepCorr}
\alias{ROC.SibPepCorr}
\title{ROC.SibPepCorr}
\usage{
ROC.SibPepCorr(traces, FFT = 1, stepsize = 0.001, plot = TRUE,
  PDF = FALSE, CSV = FALSE)
}
\arguments{
\item{traces}{An object of type \code{traces.obj}.}

\item{FFT}{numeric specifying the fraction of false targets (FFT). Default is 1.}

\item{stepsize}{numeric specifying the number of steps for calculating ROC characteristics.}

\item{plot}{logical TRUE or FALSE}

\item{PDF}{logical TRUE or FALSE}

\item{CSV}{logical TRUE or FALSE}
}
\value{
An data.table containing ROC characteristics.
}
\description{
Calculate ROC characteristics based on sibling peptide correlation.
}

