% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSECexplorer.R
\name{runSECexplorer}
\alias{runSECexplorer}
\title{Run SECprofiler from SECexplorer y providing ids and their type.}
\usage{
runSECexplorer(ids, type)
}
\arguments{
\item{ids}{A character vector containing names or identifiers.}

\item{type}{A character string specifying the type of ids.}
}
\description{
Run SECprofiler from SECexplorer y providing ids and their type.
}

