% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeatures.R
\name{runSlidingWindow}
\alias{runSlidingWindow}
\title{runSlidingWindow function}
\usage{
runSlidingWindow(complex.id, complex_hypothesis, traces, traces.imputed,
  corr_cutoff, window_size, collapse_method, rt_height, smoothing_length)
}
\arguments{
\item{complex.id}{chracter string}

\item{traces}{original traces object}

\item{traces.imputed}{imputed traces object}
}
\description{
A helper function to execute the sliding window algorithm for a specific query complex.
}
