% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTraces.R
\name{scaleTraces}
\alias{scaleTraces}
\title{Normalize tracesList intensities to a common scale}
\usage{
scaleTraces(traces, design_matrix, common_scale_within = "Condition",
  scaling_method = c("max", "median", "mean"), verbose = TRUE)
}
\arguments{
\item{traces}{Object of class tracesList.}

\item{design_matrix}{data.table, A valid design matrix describing the tracesList object.}

\item{common_scale_within}{Character string, the name of a column in the design matrix.
Values within groups in that column are brought to a common scale. Defaults to 'Condition'.}

\item{verbose}{Logical, wether to print messages to the console.}
}
\value{
Object of class traces with normalized intensities.
}
\description{
Normalize tracesList intensities to a common scale
}
