% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDifferentialExperssion.R
\name{scoreDifferentiallyBehavingPeptides}
\alias{scoreDifferentiallyBehavingPeptides}
\title{Find Feature subunits that correlate well in on hypothesis feature but not in another.}
\usage{
scoreDifferentiallyBehavingPeptides(featureVals, peakCorrCutoff = 0,
  corrCutoff = 0, featureIntCutoff = 0, noiseCutoff = 0)
}
\arguments{
\item{featureVals}{data.table, a long-format table of intensity values within feature boundaries.
A featureVals table can be produced with \code{extractFeatureVals}.}

\item{peakCorrCutoff}{numeric integer, the minimum cutoff for the mean correlation of all peptides
both compared features must pass to be included in the output.}

\item{corrCutoff}{numeric integer, the minimum cutoff for the peptide compared to the sibling peptides in
the feature. Only one of the compared features must pass this cutoff (to not eliminate presence absence cases)}
}
\value{
A list of Proteins/Protein Complexes with calculated difference scores:
        \itemize{
         \item \code{intdiff} Pairwise difference of the trace-intensity between all features.
         \item \code{smallToLarge} Logical. Is the more intense trace in the smaller Feature?
        }
}
\description{
Find Feature subunits that correlate well in on hypothesis feature but not in another.
}
\details{
The mean Intensity(Area under the curve) of all detected subunits is calculated and normalized
to the most abundant peptide in the feature. Then all pairwise differences in intensity between features
are reported.
}
