% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQvalues.R
\name{scoreFeatures}
\alias{scoreFeatures}
\title{Score protein or complex features and filter by q-value/FDR}
\usage{
scoreFeatures(features, FDR = 0.05, plot = T, PDF = FALSE,
  name = "qvalueStats")
}
\arguments{
\item{features}{data.table output from findComplexFeatures or findProteinFeatures function}

\item{FDR}{numeric between 0 and 1, default = 0.05}

\item{plot}{logical, default = TRUE}

\item{PDF}{logical, defalt = FALSE}

\item{name}{character strimg specifying pdf name, default = "qvalueStats"}
}
\value{
data.table with scored and filtered features
}
\description{
Score protein or complex features and filter by q-value/FDR
}
