% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignTraces.R
\name{shiftTraces}
\alias{shiftTraces}
\title{Shift a traces object by n fractions}
\usage{
shiftTraces(traces, shift, cropOverhang = T, imputeMissing = T)
}
\arguments{
\item{traces}{Object of class traces to shift}

\item{shift}{Numeric integer, how many fractions to shift. 
(positive = right shift, negative = left shift).}

\item{cropOverhang}{Logical, wether to remove the fraction(s) that is overhanging after shifting.}

\item{imputeMissing}{Logical, wether to impute the fraction(s) that is lagging after shifting.}
}
\value{
traces object that is shifted by the specified fractions.
}
\description{
Shift a traces object by n fractions
}

