% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeTraces.R
\name{smootheTraces}
\alias{smootheTraces}
\alias{smootheTraces.traces}
\alias{smootheTraces.tracesList}
\title{Normalize trace intensities by spline smoothing over the fractions}
\usage{
smootheTraces(
  traces,
  method = c("mean", "median"),
  scale = c("none", "log"),
  span = 0.15,
  plot = FALSE,
  PDF = F,
  name = "Smoothing_summary.pdf",
  smoothe_on_targets = TRUE,
  ...
)

\method{smootheTraces}{traces}(
  traces,
  method = c("mean", "median"),
  scale = c("none", "log"),
  span = 0.15,
  plot = FALSE,
  PDF = F,
  name = "Smoothing_summary.pdf",
  smoothe_on_targets = TRUE,
  ...
)

\method{smootheTraces}{tracesList}(
  traces,
  method = c("mean", "median"),
  scale = c("none", "log"),
  span = 0.15,
  plot = FALSE,
  PDF = F,
  name = "Smoothing_summary.pdf",
  smoothe_on_targets = TRUE,
  ...
)
}
\arguments{
\item{method}{Character string, whether to normalize to the median or the mean intensity}

\item{scale}{Character string wether to transform values before smoothing}

\item{span}{Numeric, controls the degree of smoothing (see also \code{loess} function)}

\item{plot}{Logical, whether to plot a Visualization of the smoothing}

\item{PDF}{Logical, whether to produce a PDF of the summary plot}

\item{name}{Character string, the name of the PDF produced.}

\item{smoothe_on_targets}{Logical, whether to exclude decoy traces from smoothing}
}
\value{
Object of class traces with normalized intensities.
}
\description{
Fits a spline to a summary statistic of every elution fraction
and normalizes each fraction to fit the spline
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Smoothe traces of a single traces object

\item \code{tracesList}: Smoothe traces of a tracesList object
}}

