% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{subset.traces}
\alias{subset.traces}
\title{Subset traces by trace_ids or fraction_ids}
\usage{
subset.traces(traces, trace_ids = NULL, fraction_ids = NULL)
}
\arguments{
\item{traces}{Object of class traces.}

\item{trace_ids}{Character vector specifying the trace identifiers
for subsetting traces, e.g. peptide or protein ids.}

\item{fraction_ids}{Numeric vector specifying the fraction identifiers
for subsetting traces. The resulting traces object will not have
fraction ids starting from 1, co-elution feature finding might thus
be compromised. Please don't use this option for general
pre-processing purposes.}
}
\value{
traces Object of class traces.
}
\description{
Subset a taces object by trace ids and/or fraction ids.
}
\examples{
# Load some example data:
inputTraces <- examplePeptideTraces
subsetPeptides <- c("AIIDEFEQK","AIQLSGAEQLEALK","AKEALIAASETLK")
subsetFractions <- c(5:20)
# Run subsetting:
subsettedTraces <- subset(inputTraces,trace_ids=subsetPeptides,fraction_ids=subsetFractions)
# Inspect subsetting result:
subsettedTraces
summary(subsettedTraces)
}
