% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexMethods.R
\name{subsetComplexFeatures}
\alias{subsetComplexFeatures}
\title{Filter complex feature table}
\usage{
subsetComplexFeatures(res, complex_ids = NULL, min_completeness = NULL,
  min_subunits = NULL, min_peak_corr = 0.5,
  min_monomer_distance_factor = 1)
}
\arguments{
\item{res}{A data.table with the complex features.}

\item{complex_ids}{A character vector containing all desired \code{complex_id} values.}

\item{min_completeness}{Numeric between 0 and 1, specifying the required completeness of a feature reltive to the tested hypothesis.}

\item{min_subunits}{Integer specifying minimum number of subunits in a complex.}

\item{min_peak_corr}{Numeric value betwee 0 and 1 specifying minimum peak correlation, default is 0.5}

\item{min_monomer_distance_factor}{Numeric value specifying factor to multiply largest monomer MW, default is 1}
}
\value{
The same data.table format filtered according to the provided parameters.
}
\description{
Filter result data.table according to desired complex_ids or minimum correlation score.
}
