% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexMethods.R
\name{subsetProteinFeatures}
\alias{subsetProteinFeatures}
\title{Filter protein feature table}
\usage{
subsetProteinFeatures(res, protein_ids = NULL, min_completeness = NULL,
  min_subunits = NULL, min_peak_corr = 0.5)
}
\arguments{
\item{res}{A data.table with the protein features.}

\item{protein_ids}{A character vector containing all desired \code{protein_id} values.}

\item{min_completeness}{Numeric between 0 and 1, specifying the required completeness of a feature reltive to the tested hypothesis.}

\item{min_subunits}{Integer specifying minimum number of subunits in a protein.}

\item{min_peak_corr}{Numeric value betwee 0 and 1 specifying minimum peak correlation, default is 0.5}
}
\value{
The same data.table format filtered according to the provided parameters.
}
\description{
Filter result data.table according to desired protein_ids or minimum correlation score.
}
