% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_proteogenomics.R
\name{summarize_proteogenomics.traces}
\alias{summarize_proteogenomics.traces}
\title{Summarize a traces object}
\usage{
\method{summarize_proteogenomics}{traces}(traces)
}
\description{
Summarize a traces object
}
