% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_proteogenomics.R
\name{summarize_proteogenomics.tracesList}
\alias{summarize_proteogenomics.tracesList}
\title{Summarize a tracesList object}
\usage{
\method{summarize_proteogenomics}{tracesList}(traces)
}
\description{
Summarize a tracesList object
}
