% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeRedundantPeptides.R
\name{summarizeAlternativePeptideSequences}
\alias{summarizeAlternativePeptideSequences}
\title{Summary of alternative peptide sequences}
\usage{
summarizeAlternativePeptideSequences(traces, topN = 1,
  position = "PeptidePositionStart", verbose = TRUE)
}
\arguments{
\item{traces}{An object of type traces, trace_type must be peptide.}

\item{topN}{Numeric integer, specifying the number of peptides to sum for protein quantification. Default is 2.}

\item{verbose}{Logical if warning messages should be printed. Default is TRUE.}
}
\value{
An object of type traces, trace_type is peptide.
}
\description{
Select most abndant peptide for each group of peptides with same
genomic star position. These are e.g. peptides with missed cleavages.
}
