% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureMethods.R
\name{summarizeFeatures}
\alias{summarizeFeatures}
\title{Summarize co-elution features}
\usage{
summarizeFeatures(feature_table, plot = TRUE, PDF = FALSE,
  name = "feature_summary")
}
\arguments{
\item{feature_table}{data.table as reported by \code{\link[SECprofiler]{findComplexFeatures}}
or \code{\link[SECprofiler]{findProteinFeatures}}.}

\item{plot}{Logical, whether to generate a feature summary plot. Default is \code{TRUE}.}

\item{PDF}{Logical, whether to generate a PDF file with the summary plot. Default is \code{FALSE}.}

\item{name}{Character string specifying the name of the PDF file of the summary plot.
Only applicable if PDF=\code{TRUE}. Default is "feature_summary".}
}
\value{
A summary list with complex feature metrics.
}
\description{
Summarize co-elution features
This is only taking into account targets. If decoys are present they are removed using "DECOY" as a flag.
}
\examples{
## Load example complex feature finding results:
complexFeatures <- exampleComplexFeatures
## Run summary function:
summarizeFeatures(complexFeatures)
## Filter complex features:
filteredComplexFeatures <- filterFeatures(complexFeatures)
## Run summary function on filtered data:
summarizeFeatures(filteredComplexFeatures)
}
