% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{summary.traces}
\alias{summary.traces}
\title{Summarize a \code{traces.obj}}
\usage{
\method{summary}{traces}(traces.obj)
}
\arguments{
\item{traces.obj}{An object of type \code{traces.obj}.}
}
\description{
Summarize a \code{traces.obj}
}

