% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{summary.tracesList}
\alias{summary.tracesList}
\title{Summarize a tracesList object}
\usage{
summary.tracesList(traces)
}
\arguments{
\item{traces}{Object of class tracesList}
}
\value{
A summary list with following entries:
         \itemize{
          \item \code{metrics} Summary of traces counts including decoy statistics.
          \item \code{type} Type of traces: peptides or proteins.
          \item \code{annotations} Names of trace annotations.
          \item \code{fraction_count} Number of fractions in traces object.
          \item \code{SibPepCorr} Summary of sibling peptide correlation.
          Only reported if sibling peptide correlations were previously calculated.
         }
}
\description{
Summarize a tracesList object to get an overview.
}

