% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTraces.R
\name{sumTraces}
\alias{sumTraces}
\title{Produce analyte traces by summing traces of fragments.}
\usage{
sumTraces(trace.obj, by = "protein_id")
}
\arguments{
\item{trace.obj}{An object of class \code{FragmentTraces}.}

\item{by}{The annotation column that should be used when grouping fragments
for summation.}
}
\value{
An object of type \code{FragmentTraces} with summed traces.
}
\description{
Produce analyte traces by summing traces of fragments.
}

