% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDifferentialExperssion.R
\name{testDifferentialExpression}
\alias{testDifferentialExpression}
\title{Test differential expression of peptides or proteins between 2 conditions}
\usage{
testDifferentialExpression(featureVals, compare_between = "Condition",
  level = c("protein", "peptide", "complex"), measuredOnly = TRUE)
}
\arguments{
\item{featureVals}{data.table, a long-format table of intensity values within feature boundaries.
A featureVals table can be produced with \code{extractFeatureVals}.}

\item{compare_between}{Character string, The name of the column in the featureVals table
holding the Conditions between which differential expression should be tested.}

\item{level}{Character string, return tests on peptide level or aggregate to protein level.
Must be one of c("protein","peptide"). Defaults to "protein".}

\item{measuredOnly}{Logical, if only measured values should be used for differential expression testing.}
}
\value{
A data.table containing the differential testing results for every protein/peptide.
}
\description{
Test differential expression of peptides or proteins between 2 conditions
}
