% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGlobalDifferentialExperssion.R
\name{testGlobalDifferentialExpression}
\alias{testGlobalDifferentialExpression}
\title{Perform global differential expression analysis}
\usage{
testGlobalDifferentialExpression(traces, design_matrix = NULL)
}
\arguments{
\item{traces}{object of class traces}

\item{design_matrix}{design matrix}
}
\value{
list with peptide and protein level differential expression results.
If proteoform_ids are available also proteoform differential expression in
estimated.
}
\description{
Perform global differential expression analysis
}
