% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testLocalVsGlobal.R
\name{testLocalVsGlobal}
\alias{testLocalVsGlobal}
\title{Test differential association state of peptides or proteins between 2 conditions}
\usage{
testLocalVsGlobal(featureVals, compare_between = "Condition",
  plot = TRUE, PDF = TRUE, name = "local_vs_global_stats")
}
\arguments{
\item{featureVals}{data.table, a long-format table of intensity values within feature boundaries.
A featureVals table can be produced with \code{extractFeatureVals}.}

\item{compare_between}{Character string, The name of the column in the featureVals table
holding the Conditions between which differential expression should be tested.}

\item{plot}{logical if plot should be created, default is TRUE.}

\item{PDF}{logical if PDF should be created, default is FALSE.}

\item{name}{character string specifying the name of output if PDF=TRUE, default is "local_vs_global_stats".}
}
\value{
A data.table containing the differential testing results for every peptide.
}
\description{
Test differential association state of peptides or proteins between 2 conditions
}
