% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConversion.R, R/helper.R
\name{toLongFormat}
\alias{toLongFormat}
\title{Convert a data.table containing traces from wide format to long format.}
\usage{
toLongFormat(traces.dt)

toLongFormat(traces.dt)
}
\arguments{
\item{traces.dt}{A data.table with an id column \code{fragment_id} and
columns of continuously numbered fractions.}

\item{traces.dt}{A data.table with an id column \code{fragment_id} and
columns of continuously numbered fractions.}
}
\value{
A data.table with columns
         \itemize{
          \item \code{fragment_id}
          \item \code{fraction}
          \item \code{intensity}
         }

A data.table with columns
         \itemize{
          \item \code{fragment_id}
          \item \code{fraction}
          \item \code{intensity}
         }
}
\description{
Convert a data.table containing traces from wide format to long format.

Convert a data.table containing traces from wide format to long format.
}

