% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConversion.R, R/helper.R
\name{toWideFormat}
\alias{toWideFormat}
\title{Convert a data.table containing traces from long format to wide format.}
\usage{
toWideFormat(traces.dt)

toWideFormat(traces.dt)
}
\arguments{
\item{traces.dt}{A data.table with columns
\itemize{
 \item \code{fragment_id}
 \item \code{fraction}
 \item \code{intensity}
}}

\item{traces.dt}{A data.table with columns
\itemize{
 \item \code{fragment_id}
 \item \code{fraction}
 \item \code{intensity}
}}
}
\value{
The same data.table casted to wide format.

The same data.table casted to wide format.
}
\description{
Convert a data.table containing traces from long format to wide format.

Convert a data.table containing traces from long format to wide format.
}

