% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceBrowser.R
\name{traceBrowser}
\alias{traceBrowser}
\title{traceBrowser for SECprofiler}
\usage{
traceBrowser(Traces, calibration_functions = NULL)
}
\arguments{
\item{Traces}{traces object, peptide or protein level}

\item{calibration_functions}{calibration object created by function calibration, providing functions to map
protein_mw to fraction and back, allowing to indicate expected monomeric elution fraction of selected proteins}
}
\description{
basic data browser allowing to display select subsets of chromatographic elution
profiles, e.g. SECtraces
}
