% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{updateTraces}
\alias{updateTraces}
\alias{updateTraces.traces}
\alias{updateTraces.tracesList}
\title{Update trace and fraction annotation}
\usage{
updateTraces(traces)

\method{updateTraces}{traces}(traces)

\method{updateTraces}{tracesList}(traces)
}
\arguments{
\item{traces}{Object of class traces.}
}
\value{
Object of class traces.
}
\description{
Add information to trace and fraction annotation
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Update trace and fraction annotation

\item \code{tracesList}: Update trace and fraction annotation
}}

