% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteogenomics.R
\name{validateEnsDB}
\alias{validateEnsDB}
\title{Validate if the user specified the database correctly}
\usage{
validateEnsDB(ensdb)
}
\description{
Validate if the user specified the database correctly
}
