% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignTraces.R
\name{alignTraces}
\alias{alignTraces}
\title{Get the alignment shifts in tracesList object}
\usage{
alignTraces(tracesList, alignment_order = seq_along(tracesList),
  plot = TRUE, PDF = FALSE, name = "AlignmentSummary.pdf", ...)
}
\arguments{
\item{tracesList}{Object of class tracesList to align.}

\item{alignment_order}{Integer vector, determines the sequence of alignment.}

\item{plot}{Logical, wether to plot boxplots of corellation at different lags.}

\item{PDF}{Logical, wether to produce a PDF in the working directory.}

\item{name}{Character string, the name of the PDF file.}
}
\value{
Numeric vector, the shifts with the highest corellations for each alignment
}
\description{
Get the alignment shifts in tracesList object
}

