% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMassDistribution.R
\name{annotateMassDistribution}
\alias{annotateMassDistribution}
\title{annotateMassDistribution}
\usage{
annotateMassDistribution(traces)
}
\arguments{
\item{traces}{An object of type traces.}
}
\description{
Estimates fraction of assembled vs. monomeric mass per id.
}
