% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{annotateMolecularWeight}
\alias{annotateMolecularWeight}
\title{Annotate traces with molecular weight calibration.}
\usage{
annotateMolecularWeight(traces, calibration)
}
\arguments{
\item{Object}{of class traces.}
}
\value{
Object of class traces with moleclar weight annotation of fractions.
}
\description{
Annotate fractions in traces object with calibrated molecular weight.
This only applies to specific fractionation strategies e.g. SEC)
}
\examples{
# Load relevant data:
inputTraces <- examplePeptideTraces
calibrationTable <- exampleCalibrationTable
# Perform molecular weight calibration:
calibration = calibrateMW(calibrationTable)
# Perform molecular weight annotation:
mwTraces <- annotateMolecularWeight(inputTraces, calibration)
}
