% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFractions.R, R/tracesMethods.R
\name{annotateFractions.traces}
\alias{annotateFractions.traces}
\alias{annotateFractions.tracesList}
\alias{annotateMolecularWeight}
\alias{annotateMolecularWeight.traces}
\alias{annotateMolecularWeight.tracesList}
\title{Annotate traces with molecular weight calibration.}
\usage{
\method{annotateFractions}{traces}(traces)

\method{annotateFractions}{tracesList}(traces)

annotateMolecularWeight(traces, calibration)

\method{annotateMolecularWeight}{traces}(traces, calibration)

\method{annotateMolecularWeight}{tracesList}(traces, calibration)
}
\arguments{
\item{traces}{Object of class traces.}
<<<<<<< HEAD
=======

\item{calibration}{Calibration functions generated by calibrateMW().}
>>>>>>> master
}
\value{
Object of class traces with moleclar weight annotation of fractions.
}
\description{
Annotate fractions in traces object with calibrated molecular weight.
This only applies to specific fractionation strategies e.g. SEC)
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Annotate single traces object

\item \code{tracesList}: Annotate multiple traces objects

\item \code{traces}: Annotate single traces object

\item \code{tracesList}: Annotate tracesList object
}}

\examples{
# Load relevant data:
inputTraces <- examplePeptideTraces
calibrationTable <- exampleCalibrationTable
# Perform molecular weight calibration:
calibration = calibrateMW(calibrationTable)
# Perform molecular weight annotation:
mwTraces <- annotateMolecularWeight(inputTraces, calibration)
}
