% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateTraces.R
\name{annotateTraces}
\alias{annotateTraces}
\title{annotate Traces object}
\usage{
annotateTraces(traces, traces_id_column = "id",
  annotation_table = "uniprot.tab", annotation_table_id_column = "Entry",
  replace_whitespace = TRUE)
}
\arguments{
\item{traces}{Object of class traces}

\item{annotation_table}{path to tab-separated annotation table to be merged into trace_annotation}

\item{replace_whitespace}{whether spaces contained in the anntation table column names shall be replaced by "_"
defaults to TRUE
,typically FullPeptideName or Uniprot Entry names as protein_id. Same as in traces$traces$id}

\item{id_column}{name/header of the column in the reference table that contains the ids
defaults to "Entry" as expexted from uniprot.tab files}
}
\description{
Add columns to the trace_annotation table from an external annotation table
}

