% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateTraces.R
\name{annotateTraces}
\alias{annotateTraces}
\alias{annotateTraces.traces}
\alias{annotateTraces.tracesList}
\title{Annotate traces object}
\usage{
annotateTraces(
  traces,
  trace_annotation,
  traces_id_column = "protein_id",
  trace_annotation_id_column = "Entry",
  trace_annotation_mass_column = "Mass",
  uniprot_mass_format = TRUE,
  replace_whitespace = TRUE
)

\method{annotateTraces}{traces}(
  traces,
  trace_annotation,
  traces_id_column = "protein_id",
  trace_annotation_id_column = "Entry",
  trace_annotation_mass_column = "Mass",
  uniprot_mass_format = TRUE,
  replace_whitespace = TRUE
)

\method{annotateTraces}{tracesList}(
  traces,
  trace_annotation,
  traces_id_column = "protein_id",
  trace_annotation_id_column = "Entry",
  trace_annotation_mass_column = "Mass",
  uniprot_mass_format = TRUE,
  replace_whitespace = TRUE
)
}
\arguments{
\item{traces}{Object of class traces.}

\item{trace_annotation}{Tab-separated annotation table to be merged into trace_annotation, file or R data.table.}

\item{traces_id_column}{Character string with name of the id column specifying the entries to be annotated.
Defaults to "protein_id".}

\item{trace_annotation_id_column}{Character string with name of the column in the reference table that contains the ids.
Defaults to "Entry" as expexted from uniprot.tab files.}

\item{trace_annotation_mass_column}{Character string with name of the column in the reference table that contains the protein mass annotation in kDa. If \code{uniprot_mass_format = TRUE}, this column should be called "Mass" and is in Da (PCprofiler will do the conversion). Defaults to "Mass".}

\item{uniprot_mass_format}{Logical, whether mass annotation is in the format of uniprot ("Mass" column with molecular weight in Da and comma as thouthands separator).}

\item{replace_whitespace}{Logical, whether spaces contained in the anntation table column names shall be replaced by "_".
Defaults to \code{TRUE}.}
}
\value{
Traces object of class traces with extended annotation.
}
\description{
Add custom annotation columns to the trace_annotation in a traces object from an external annotation table, e.g. UniProt.
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Annotate single traces object

\item \code{tracesList}: Annotate single traces object
}}

\examples{
# Load some example data:
inputTraces <- examplePeptideTracesUnannotated
inputAnnotation <- exampleTraceAnnotation
# Run the annotation:
annotatedTraces <- annotateTraces(
traces=inputTraces,
trace_annotation=inputAnnotation,
traces_id_column = "protein_id",
trace_annotation_id_column = "Entry")
Inspect annotation result:
annotatedTraces$trace_annotation
}
