% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{calculateMinCorr}
\alias{calculateMinCorr}
\title{Calculate minimum correlation of each peptide
to all of its sibling peptides}
\usage{
calculateMinCorr(
  traces,
  FeatureFinding = NULL,
  plot = FALSE,
  PDF = FALSE,
  name = "minCorrHist",
  ...
)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{plot}{logical,wether to print SibPepCorr density plot to R console.
Deafult is \code{FALSE}.}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file.
Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if
'\code{PDF=TRUE}.PDF file is saved under name.pdf. Default is "minCorrHist".}
}
\value{
Object of class traces with calculated minimum.
}
\description{
Calculate minimum correlation of each peptide
to all of its sibling peptides
}
