% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypotheses.R
\name{calculatePathlength}
\alias{calculatePathlength}
\title{Calculate shortest paths between all members of a binary interaction network.
This function creates a table that contains all possible combinations of 2 from
all input proteins (binomial coeffitient (length(proteins) over 2)).}
\usage{
calculatePathlength(binaryInteractionTable)
}
\arguments{
\item{binaryInteractionTable}{data.table with binary interactions between "a" and "b" as returned from generateBinaryNetwork()}
}
\value{
data.table with three columns (x,y,dist) containing all possible combinations
of proteins and their shortest distance as determined by the input binary interaction table
}
\description{
Calculate shortest paths between all members of a binary interaction network.
This function creates a table that contains all possible combinations of 2 from
all input proteins (binomial coeffitient (length(proteins) over 2)).
}
