% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSibPepCorr.R
\name{calculateSibPepCorr}
\alias{calculateSibPepCorr}
\title{Calculate sibling peptide correlation in traces.object of type peptide}
\usage{
calculateSibPepCorr(Traces, plot = TRUE, PDF = FALSE)
}
\arguments{
\item{plot}{logical TRUE or FALSE}

\item{PDF}{logical TRUE or FALSE}

\item{traces.obj}{An object of type \code{traces.obj}.}

\item{FFT}{numeric fraction of false targets (pi0)}
}
\value{
traces.obj An object of type \code{traces.obj}.
}
\description{
Calculate sibling peptide correlation in traces.object of type peptide
}

