% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSibPepCorr.R, R/filterBySibPepCorr.R,
%   R/filterConsecutiveIdStretches.R
\name{calculateSibPepCorr}
\alias{calculateSibPepCorr}
\alias{calculateSibPepCorr.traces}
\alias{calculateSibPepCorr.tracesList}
\alias{filterBySibPepCorr.traces}
\alias{filterBySibPepCorr.tracesList}
\alias{filterConsecutiveIdStretches.traces}
\alias{filterConsecutiveIdStretches.tracesList}
\title{calculateSibPepCorr}
\usage{
calculateSibPepCorr(
  traces,
  plot = TRUE,
  PDF = FALSE,
  name = "SibPepCorr_densityplot"
)

\method{calculateSibPepCorr}{traces}(
  traces,
  plot = TRUE,
  PDF = FALSE,
  name = "SibPepCorr_densityplot"
)

\method{calculateSibPepCorr}{tracesList}(
  traces,
  plot = TRUE,
  PDF = FALSE,
  name = "SibPepCorr_densityplot",
  ...
)

\method{filterBySibPepCorr}{traces}(
  traces,
  fdr_cutoff = 0.01,
  fdr_type = "protein",
  FFT = 1,
  absolute_spcCutoff = NULL,
  rm_decoys = FALSE,
  plot = TRUE,
  PDF = FALSE,
  CSV = FALSE
)

\method{filterBySibPepCorr}{tracesList}(
  traces,
  fdr_cutoff = 0.01,
  fdr_type = "protein",
  FFT = 1,
  absolute_spcCutoff = NULL,
  rm_decoys = FALSE,
  plot = TRUE,
  PDF = FALSE,
  CSV = FALSE,
  ...
)

\method{filterConsecutiveIdStretches}{traces}(
  traces,
  min_stretch_length = 3,
  remove_empty = TRUE
)

\method{filterConsecutiveIdStretches}{tracesList}(
  traces,
  min_stretch_length = 3,
  remove_empty = TRUE,
  ...
)
}
\arguments{
\item{traces}{An object of type traces.}

\item{plot}{logical,wether to print SibPepCorr density plot to R console. Deafult is \code{TRUE}.}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file. Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if \code{PDF=TRUE}.
PDF file is saved under name.pdf. Default is "SibPepCorr_densityplot".}
}
\value{
An object of type traces with added SibPepCorr column.
}
\description{
Calculate sibling peptide correlation in traces object of type peptide.
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Calculate sibling peptide correlation in traces object.

\item \code{tracesList}: Calculate sibling peptide correlation in tracesList object.

\item \code{traces}: Filter by sibling peptide correlation in traces object.

\item \code{tracesList}: Filter by sibling peptide correlation in traces object.

\item \code{traces}: Filter Consecutive Id Stretches in Chromatograms of single traces object.

\item \code{tracesList}: #' @describeIn calculateSibPepCorr Filter Consecutive Id Stretches in Chromatograms of tracesList object.
}}

\examples{
## Load example data
 tracesRaw <- examplePeptideTracesUnannotated

 ## Calculate the SibPepCorr of every peptide
 tracesRawSpc <- calculateSibPepCorr(traces = tracesRaw)

 tracesRawSpc$trace_annotation


}
