% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateSECMW.R
\name{calibrateSECMW}
\alias{calibrateSECMW}
\title{Calibrate SEC - establish connectivity between SEC fraction number and apparent
Molecular Weight (in kDa)}
\usage{
calibrateSECMW(std_weights_kDa, std_elu_fractions, plot = TRUE)
}
\arguments{
\item{std_weights_kDa}{The Molecular Weights of the standard proteins}

\item{std_elu_fractions}{The fraction numbers where these standard proteins eluted}

\item{Traces}{An Object of type Traces on which the calibration is to be performed}
}
\value{
An additional column "apparentMW_kDa" in the Traces$fraction_annotation table
}
\description{
Calibrate SEC - establish connectivity between SEC fraction number and apparent
Molecular Weight (in kDa)
}

