% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateSECMW.R
\name{calibrateSECMW}
\alias{calibrateSECMW}
\title{calibrateSECMW}
\usage{
calibrateSECMW(calibration_table, plot = TRUE, PDF = FALSE)
}
\arguments{
\item{calibration_table}{tab-separated calibration table, file or R data.table
Columns: 
std_weights_kDa = The Molecular Weights of the standard proteins
std_elu_fractions = The fraction numbers where these standard proteins eluted}

\item{plot}{logical, if to plot calibration}

\item{PDF}{logical, if to produce a PDF}
}
\value{
List of two functions: MWtoSECfraction and SECfractionToMW
}
\description{
Calibrate SEC - establish connectivity between SEC fraction number and apparent
Molecular Weight (in kDa)
}
