% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseHypothesisWideFeatures.R
\name{complexClustering}
\alias{complexClustering}
\title{complexClustering}
\usage{
complexClustering(complexFeatures, dist_mat)
}
\arguments{
\item{dist_mat}{distance matrix from getDistanceMatrix}

\item{complexFeature}{data.table containing filtered complex feature results.}
}
\value{
cluster object
}
\description{
complexClustering.
}
