% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSECToMW.R
\name{convertSECToMW}
\alias{convertSECToMW}
\title{Convert a fraction number to an estimated molecular weight.}
\usage{
convertSECToMW(SECfraction)
}
\arguments{
\item{SECfraction}{The fraction number.}
}
\value{
Apparent MW at this SEC fraction in kDa.
}
\description{
Convert a fraction number to an estimated molecular weight.
}

