% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R
\name{estimateDecoyFDR}
\alias{estimateDecoyFDR}
\title{Estimate statistics based on decoy model.}
\usage{
estimateDecoyFDR(complex_features, grid_search_list = FALSE)
}
\arguments{
\item{complex_features}{data.table containing filtered complex feature results.}
}
\value{
List with stats
}
\description{
Estimate statistics based on decoy model.
}
