% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFDRestimation.R
\name{estimateDecoyFDR}
\alias{estimateDecoyFDR}
\title{Estimate decoy based featureFinding FDR.}
\usage{
estimateDecoyFDR(features, FFT = 1, grid_search_params = NULL, verbose = TRUE)
}
\arguments{
\item{features}{data.table containing filtered complex feature results.}

\item{FFT}{Numeric, fraction of false targets. Only useful in special cases, should not
be altered in a generic workflow. Defaults to \code{1}.}

\item{grid_search_params}{Character vector of columns to report with the statistics for the dataset.
Usually only useful for a grid search. Otherwhise the default should not be altered.
Default = \code{NULL}.}

\item{verbose}{Logical, wether to print messages. Default = \code{TRUE}.}
}
\value{
List with stats
}
\description{
Estimate FDR statistics based on a decoy model after feature finding.
}
\examples{

#-----------------------------
## Complex Features
#-----------------------------

## Load example data
complexFeatures <- exampleComplexFeatures
## Filter feature table 
complexFeaturesFiltered <- filterFeatures(complexFeatures,
                                          min_peak_corr= 0.5,
                                          min_feature_completeness= 0.5) 

## estimate the FDR
estimateDecoyFDR(complexFeaturesFiltered)

#-----------------------------
## Protein Features
#-----------------------------

## Load example data
proteinFeatures <-exampleProteinFeatures
## Filter feature table 
proteinFeaturesFiltered <- filterFeatures(proteinFeatures,
                                          min_peak_corr= 0.5,
                                          min_feature_completeness= 0.5) 

## estimate the FDR
estimateDecoyFDR(proteinFeaturesFiltered)

}
