% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R
\name{estimateGridSearchDecoyFDR}
\alias{estimateGridSearchDecoyFDR}
\title{Perform complex feature grid search FDR}
\usage{
estimateGridSearchDecoyFDR(complex_features_list)
}
\arguments{
\item{complex_features}{data.table containing filtered complex feature results.}
}
\value{
List with stats
}
\description{
Perform complex feature grid search.
}
