% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestFeatureParameters.R
\name{estimateGridSearchDecoyFDR}
\alias{estimateGridSearchDecoyFDR}
\title{Estimate feature grid search FDR}
\usage{
estimateGridSearchDecoyFDR(
  complex_features_list,
  grid_search_params = c("corr", "window", "rt_height", "smoothing_length",
    "min_feature_completeness", "min_hypothesis_completeness", "min_subunits",
    "min_peak_corr", "min_monomer_distance_factor")
)
}
\arguments{
\item{complex_features_list}{data.table containing filtered complex feature results.}

\item{grid_search_params}{Character vector of column names to report with the statistics for the dataset.
Should contain all parameters that are of interest in the grid search.}
}
\value{
A data.table with FDR estimations for every parameter combination specified.
}
\description{
Estimate feature grid search FDR by decoy counting.
}
\examples{
## Generate example data
complexFeaturesGrid <- list(exampleComplexFeatures)

complexFeaturesGridFiltered <- filterGridSearchResults(complexFeaturesGrid,
                                                       peak_corr_cutoffs = c(0.5,0.75,0.9),
                                                       feature_completeness_cutoffs = c(0,0.5,1),
                                                       hypothesis_completeness_cutoffs = c(0.5,1),
                                                       n_subunits_cutoffs =c(2,3,4),
                                                       monomer_distance_cutoffs = c(1,2),
                                                       remove_decoys=FALSE)

## Calculate the FDR statistic for every grid search result
# Here we only performed grid search of different filter cutoffs, so only those are
# specified as grid_search_params

gridStats <- estimateGridSearchDecoyFDR(complexFeaturesGridFiltered,
                                        grid_search_params =c("min_feature_completeness",
                                                              "min_hypothesis_completeness",
                                                              "min_subunits",
                                                              "min_peak_corr",
                                                              "min_monomer_distance_factor"))
## Inspect the oputput
gridStats
}
