% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplePeptideTraces}
\alias{examplePeptideTraces}
\title{A sample peptide level traces object before any filtering.}
\format{An object of class \code{traces} of length 4.}
\usage{
examplePeptideTraces
}
\description{
A sample peptide level traces object before any filtering.
}
\keyword{datasets}
