% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplePeptideTracesFiltered}
\alias{examplePeptideTracesFiltered}
\title{Examplary peptide level traces object after filtering}
\format{
An object of class \code{traces} of length 4.
}
\usage{
examplePeptideTracesFiltered
}
\description{
Peptide level traces object as generated by filtering for consecutive id stratches and sibling peptide correlation.
}
\examples{
consecutiveFilteredTraces <- filterConsecutiveIdStretches(examplePeptideTraces,
                                                          min_stretch_length = 3)
sibPepCorrFilteredTraces <- filterBySibPepCorr(consecutiveFilteredTraces,
                                               fdr_cutoff = NULL, 
                                               absolute_spcCutoff = 0.2)
all.equal(sibPepCorrFilteredTraces,examplePeptideTracesFiltered)
}
\keyword{datasets}
