% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleProteinTraces}
\alias{exampleProteinTraces}
\title{Examplary protein level traces object}
\format{An object of class \code{traces} of length 4.}
\usage{
exampleProteinTraces
}
\description{
Protein level traces as, for example, generated by \code{proteinQuantification}.
}
\examples{
## protein quantification
proteinTraces <- proteinQuantification(examplePeptideTracesFiltered) 
all.equal(proteinTraces,exampleProteinTraces)
}
\keyword{datasets}

