% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatureVals.R
\name{fillFeatureVals}
\alias{fillFeatureVals}
\title{Fill traceVals for all Samples}
\usage{
fillFeatureVals(featureVals, design_matrix, perturb_cutoff = "5\%")
}
\arguments{
\item{featureVals}{data.table, a long-format table of intensity values within feature boundaries.
A featureVals table can be produced with \code{extractFeatureVals}.}

\item{design_matrix}{data.table, design matrix describing the architecture of the tracesList object.}

\item{perturb_cutoff}{Character string or numeric, the quantile of values that noise is sampled from.
Noise needs to be imputed to calculate the correlations.}
}
\value{
Long format table containing filled feature values.
}
\description{
Fill in traces that are not detected in certain conditions with noise values
}
