% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByPeptideCorrelation.R
\name{filterByPeptideCorrelation}
\alias{filterByPeptideCorrelation}
\title{Import peptide profiles from an OpenSWATH experiment.}
\usage{
filterByPeptideCorrelation(traces, min.pairwise.corr = 0.8,
  average.corr.cutoff = ">= median")
}
\arguments{
\item{traces}{Object of class traces containing the quantitative MS data
table.}

\item{min.pairwise.corr}{Minimum pairwise pearson correlation necessary
to retain the data for a protein with 2 peptides.}

\item{average.corr.cutoff}{For proteins with >= 3 peptides, retain all
in the data whose average correlation with all siblings is how good,
defaults to '>= median' but can be replaces with a numeric cutoff
(e.g. 0.6).}
}
\value{
An object of class traces (peptide level) containing table.wide,
    table.long and id that can be processed with the herein contained
    functions.
}
\description{
filters peptide elution profiles based on sibling peptide
    correlation.
}

