% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBySibPepCorr.R
\name{filterBySibPepCorr}
\alias{filterBySibPepCorr}
\title{filter a traces object based on the Sibling Peptide Correlation (SibPepCorr).}
\usage{
filterBySibPepCorr(Traces, protein_fdr_cutoff = 0.01, FFT = 1,
  absolute_spc_cutoff = NULL, remove_decoys = FALSE, plot = TRUE,
  PDF = FALSE, CSV = FALSE)
}
\description{
Column, estimating FDR if decoys are contained in $trace_annotation$protein_id.
}

