% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterConsecutiveIdStretches.R
\name{filterConsecutiveIdStretches}
\alias{filterConsecutiveIdStretches}
\title{filterConsecutiveIdStretches}
\usage{
filterConsecutiveIdStretches(traces, min_stretch_length = 3,
  remove_empty = TRUE)
}
\arguments{
\item{traces}{An object of type \code{traces.obj}.}

\item{min_stretch_length}{The minimal length a stretch of contiguous
identifications has to have in order not to be removed.}

\item{remove_empty}{Logical whether the entries with rowSum == 0 after
filtering are removed, defaults to TRUE.}
}
\value{
An object of class traces containing the filtered
    chromatograms.
}
\description{
Filter protein profiling traces based on
minimal length of consecutive ID stretches.
}

