% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R,
%   R/getBestFeatureParameters.R
\name{filterGridSearchResults}
\alias{filterGridSearchResults}
\alias{filterGridSearchResults}
\title{Perform complex feature grid search filter}
\usage{
filterGridSearchResults(grid_search_results, peak_corr_cutoffs = c(0.5, 0.75,
  0.9), completeness_cutoffs = c(0, 0.5, 1), n_subunits_cutoffs = c(2, 3,
  4), remove_decoys = FALSE)

filterGridSearchResults(grid_search_results, peak_corr_cutoffs = c(0.5, 0.75,
  0.9), completeness_cutoffs = c(0, 0.5, 1), n_subunits_cutoffs = c(2, 3,
  4), remove_decoys = FALSE)
}
\arguments{
\item{grid_search_results}{data.table containing filtered complex feature results.}

\item{grid_search_results}{data.table containing filtered complex feature results.}
}
\value{
List with stats

List with stats
}
\description{
Perform complex feature grid search filter.

Perform complex feature grid search filter.
}
