% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{filterSinglePeptideHits}
\alias{filterSinglePeptideHits}
\alias{filterSinglePeptideHits.traces}
\alias{filterSinglePeptideHits.tracesList}
\title{Remove proteins with single peptides}
\usage{
filterSinglePeptideHits(traces)

\method{filterSinglePeptideHits}{traces}(traces)

\method{filterSinglePeptideHits}{tracesList}(tracesList)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}
}
\value{
Object of class traces containing only proteins
with multiple peptides
}
\description{
Remove proteins with single peptides
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Remove proteins with single peptides

\item \code{tracesList}: Remove proteins with single peptides
}}

