% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexFeatureDetection.R
\name{findComplexFeatures}
\alias{findComplexFeatures}
\title{Detect subgroups of proteins within a matrix of protein intensity traces by
sliding a window across the SEC dimension. Within each window proteins
with traces that correlate well are clustered together.}
\usage{
findComplexFeatures(trace.mat, protein.names, corr.cutoff = 0.75,
  window.size = 15, with.plot = F, noise.quantile = 0.2, min.sec = 1)
}
\arguments{
\item{trace.mat}{A numeric matrix where rows correspond to the different
traces.}

\item{protein.names}{A vector with protein identifiers. This vector has to
have the same length as the number of rows in `trace.mat`.}

\item{corr.cutoff}{The correlation value for chromatograms above which
proteins are considered to be coeluting.}

\item{window.size}{Size of the window. Numeric.}

\item{noise.quantile}{The quantile to use in estimating the noise level.
Intensity values that are zero are imputed with random noise
according to the noise estimation.}

\item{min.sec}{The lowest SEC number in the sample.}
}
\value{
An instance of class `complexFeaturesSW`.
}
\description{
Detect subgroups of proteins within a matrix of protein intensity traces by
sliding a window across the SEC dimension. Within each window proteins
with traces that correlate well are clustered together.
}

