% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeatures.R
\name{findComplexFeatures}
\alias{findComplexFeatures}
\title{Run the sliding window algorithm for a number of protein complexes.}
\usage{
findComplexFeatures(traces.obj, complex.protein.assoc,
  MWSECcalibrationFunctions, corr.cutoff = 0.95, window.size = 15,
  parallelized = FALSE, n.cores = parallel::detectCores(),
  perturb.cutoff = "5\%", collapse_method = "apex_only", rt_height = 5,
  smoothing_length = 11)
}
\arguments{
\item{traces.obj}{An object of type \code{traces.obj}.}

\item{complex.protein.assoc}{A data.table that encodes the input complexes
and their subunit compositions. The data.table should have the
format:
\itemize{
 \item \code{complex_id} A character vector that uniquely identifies
       a complex.
 \item \code{complex_name} A character vector that contains
       the complex name.
 \item \code{protein_id} A character vector that uniquely identifies
       a protein and links it to a trace contained in
       \code{traces.obj}.
}}

\item{corr.cutoff}{The correlation value for chromatograms above which
proteins are considered to be coeluting.}

\item{window.size}{Size of the window. Numeric.}

\item{parallelized}{If the computation should be done in parallel.
Optional.}

\item{n.cores}{The number of cores to use for parallel processing. Optional.}

\item{perturb.cutoff}{The quantile to use in estimating the perturbation level.
Intensity values that are zero are replaced with random values that are
below the specified quantile of the input values. Alternatively a
cutoff value can be specified as an upper limit for perturbation values.
This is nescessary for correlation calculation.}
}
\value{
A list containing various results.
        \itemize{
         \item \code{sw.results} A list of results of the function
               \code{findComplexFeatures}. One for each query complex.
         \item \code{input.complexes} A character vector of all query
                complexes.
         \item \code{corr.cutoff} The correlation cutoff used.
         \item \code{window.size} The window size used.
        }
}
\description{
Run the sliding window algorithm for a number of protein complexes.
}
\examples{
# NOT RUN:
# protein.ids <- corum.complex.protein.assoc[complex_id == 181, protein_id]
# traces <- subset(protein.traces[protein_id \%in\% protein.ids],
#                  select=-protein_id)
# sample.complexes <- c(181, 191)
# complex.protein.assoc <-
#     corum.complex.protein.assoc[complex_id \%in\% sample.complexes]
# findComplexFeaturesSWBulk(traces,
#                           protein.ids,
#                           protein.mw.conc[protein_id \%in\% protein.ids],
#                           complex.protein.assoc)
}

