% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComplexFeaturesSW.R
\name{findComplexFeaturesSW}
\alias{findComplexFeaturesSW}
\title{Detect subgroups of proteins within a matrix of protein intensity traces by
sliding a window across the SEC dimension. Within each window proteins
with traces that correlate well are clustered together.}
\usage{
findComplexFeaturesSW(traces.obj, corr.cutoff = 0.95, window.size = 15,
  with.plot = F, noise.quantile = 0.2, min.sec = 1)
}
\arguments{
\item{traces.obj}{An object of type \code{traces.obj}.}

\item{corr.cutoff}{The correlation value for chromatograms above which
proteins are considered to be coeluting.}

\item{window.size}{Size of the window. Numeric.}

\item{with.plot}{T (TRUE) or F (FALSE) whether to plot the correlation tree.}

\item{noise.quantile}{The quantile to use in estimating the noise level.
Intensity values that are zero are imputed with random noise
according to the noise estimation.}

\item{min.sec}{The lowest SEC number in the sample.}
}
\value{
An object of type \code{complexFeaturesSW} that is a list
       containing the following:
       \itemize{
         \item \code{feature} data.table containing complex feature candidates in the following format:
          \itemize{
          \item \code{subgroup} The protein_ids of the feature separated by semi-colons.
          \item \code{left_sw} The left boundary of the sliding-window feature.
          \item \code{right_sw} The right boundary of the sliding-window feature.
          \item \code{score} The intra-sliding-window-feature correlation.
          }
         \item \code{window.size} The window.size used when running this
             function.
         \item \code{corr.cutoff} The corr.cutoff used when running this
             function.
       }
}
\description{
Detect subgroups of proteins within a matrix of protein intensity traces by
sliding a window across the SEC dimension. Within each window proteins
with traces that correlate well are clustered together.
}
\examples{
# NOT RUN:
# protein.ids <- corum.complex.protein.assoc[complex_id == 181, protein_id]
# traces <- subset(protein.traces[protein_id \%in\% protein.ids],
#                  select=-protein_id)
# sw.res <- findComplexFeaturesSW(traces, protein.ids, protein.mw.conc)
}

