% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexFeatureDetection.R
\name{findComplexFeaturesTargeted}
\alias{findComplexFeaturesTargeted}
\title{Run the sliding window algorithm over a wide format data.table of protein
traces.}
\usage{
findComplexFeaturesTargeted(protein.traces.sw, corr.cutoff = 0.99,
  window.size = 15, n.cores = detectCores())
}
\description{
Run the sliding window algorithm over a wide format data.table of protein
traces.
}

