% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexFeatureDetection.R
\name{findComplexFeaturesWithinWindow}
\alias{findComplexFeaturesWithinWindow}
\title{Detect subgroups within a window. This is a helper function and should 
be called by `findComplexFeatures`.}
\usage{
findComplexFeaturesWithinWindow(tracemat, corr.cutoff, with.plot = F,
  sec = NULL)
}
\arguments{
\item{tracemat}{A matrix of intensity values where chromatograms of
individual proteins are given by the rows.}

\item{corr.cutoff}{The correlation value for chromatograms above which
proteins are considered to be coeluting.}
}
\value{
A vector of the same length as the number of proteins. Each element
        corresponds to a integer-based cluster label.
}
\description{
Detect subgroups within a window. This is a helper function and should 
be called by `findComplexFeatures`.
}

