% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findProteinFeatures.R
\name{findProteinFeatures}
\alias{findProteinFeatures}
\title{Protein feature detection}
\usage{
findProteinFeatures(traces, calibration, corr_cutoff = 0.95,
  window_size = 12, parallelized = FALSE, n_cores = 1,
  collapse_method = "apex_only", perturb_cutoff = "5\%", rt_height = 5,
  smoothing_length = 9, useRandomDecoyModel = TRUE)
}
\arguments{
\item{traces}{An object of type \code{traces}.}

\item{calibration}{A list of the calibration functions as generated by \code{calibrateSECMW}.}

\item{corr_cutoff}{The correlation value for chromatograms above which
peptides are considered to be coeluting, default=0.95.}

\item{window_size}{numeric size of the window, default=12}

\item{parallelized}{logical, if the computation should be done in parallel, default=FALSE}

\item{n_cores}{The number of cores to use for parallel processing, default=1}

\item{collapse_method}{Method for collapsing multiple features into one feature: "apex_only" or "apex_network", default="apex_only"}

\item{perturb_cutoff}{The quantile to use in estimating the perturbation level, default="5%".
Intensity values that are zero are replaced with random values that are
below the specified quantile of the input values. Alternatively a
cutoff value can be specified as an upper limit for perturbation values.
This is nescessary for correlation calculation.}

\item{rt_height}{numeric RT cutoff for collapsing features, default is 5}

\item{smoothing_length}{numeric smoothing length of Savitzky-Golay filter, default is 7}

\item{useRandomDecoyModel}{logical if random peptide protein associations should be used as decoy model, default = TRUE}
}
\value{
A data.table containing protein features.
}
\description{
Run the sliding window algorithm to find protein features.
}
