% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypotheses.R
\name{generateComplexTargets}
\alias{generateComplexTargets}
\title{Generate target hypotheses from distance table. This is done by choosing all
max_distance (e.g. 1st) degree interactors for each protein in a distance table.}
\usage{
generateComplexTargets(dist_info, max_distance = 1)
}
\arguments{
\item{dist_info}{data.table with three columns (x,y,dist) containing all possible combinations
of proteins and their shortest distance.
This is the output from calculatePathlength()}

\item{max_distance}{numeric maximum distance between to proteins to be considered as being in one complex}
}
\value{
data.table in the format of complex hypotheses
}
\description{
Generate target hypotheses from distance table. This is done by choosing all
max_distance (e.g. 1st) degree interactors for each protein in a distance table.
}
