% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypotheses.R
\name{generateDecoys}
\alias{generateDecoys}
\title{Generate list of proteins for one complex decoy hypothesis}
\usage{
generateDecoys(size, all_proteins, dist_info, min_distance)
}
\arguments{
\item{size}{numeric number of proteins to generate for complex decoy hypothesis}

\item{all_proteins}{character vector of all possible protein_ids}

\item{dist_info}{data.table with three columns (x,y,dist) containing all possible combinations
of proteins and their shortest distance.
This is the output from calculatePathlength()}

\item{min_distance}{numeric minimum distance between to proteins to be able to be in same complex decoy}
}
\value{
data.table with binary interactions between "a" and "b"
}
\description{
Generate list of proteins for one complex decoy hypothesis
}
