% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFDRestimation.R
\name{generateRandomPepTraces}
\alias{generateRandomPepTraces}
\title{generateRandomPepTraces}
\usage{
generateRandomPepTraces(traces, append = FALSE, rm_orig_decoys = TRUE)
}
\arguments{
\item{traces}{traces object of type peptide}

\item{append}{Logical, whether to append the generated decoy traces
to the original target traces}

\item{traces}{traces bject of type peptide}
}
\value{
traces object of type peptide with random peptides per protein
}
\description{
generateRandomPepTraces.
}
\examples{
## Load example data
exampleTraces <- examplePeptideTraces

## Generate the decoy traces object
decoyTraces <- generateRandomPepTraces(exampleTraces)
# Inspect the result
summary(decoyTraces)
#--------------------
## Directly append decoy traces to target taces
tracesWithDecoys <- generateRandomPepTraces(exampleTraces,
                                            append = T)
# Inspect the result
summary(tracesWithDecoys)

}
