% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestFeatureParameters.R
\name{getBestParameterData}
\alias{getBestParameterData}
\title{Get the best performing parameter data}
\usage{
getBestParameterData(complex_features_list, FDR = 0.1,
  grid_search_params = c("corr", "window", "rt_height", "smoothing_length",
  "min_feature_completeness", "min_hypothesis_completeness", "min_subunits",
  "min_peak_corr", "min_monomer_distance_factor"))
}
\arguments{
\item{complex_features_list}{data.table containing filtered complex feature results.}

\item{FDR}{Numeric, maximum FDR that should be considered, default = 0.1.}

\item{grid_search_params}{Character vector of column names to report with the statistics 
for the dataset. Should contain all parameters that are of interest in the grid search.}
}
\value{
data.table with features for best parameter set.
}
\description{
Pick the feature-finding result with the parameter set in a grid search output
that performs best in terms of identifications while staying within a specified FDR cutoff.
}
\examples{
## Generate example data
complexFeaturesGrid <- list(exampleComplexFeatures)

complexFeaturesGridFiltered <- filterGridSearchResults(complexFeaturesGrid,
                                                       peak_corr_cutoffs = c(0.5,0.75,0.9),
                                                       feature_completeness_cutoffs = c(0,0.5,1),
                                                       hypothesis_completeness_cutoffs = c(0.5,1),
                                                       n_subunits_cutoffs =c(2,3,4),
                                                       monomer_distance_cutoffs = c(1,2),
                                                       remove_decoys=FALSE)

## Extract the result table with the best parameter set
bestComplexFeatures <- getBestParameterData(complexFeaturesGridFiltered,
                                            grid_search_params = c("min_feature_completeness",
                                                                   "min_hypothesis_completeness",
                                                                   "min_subunits",
                                                                   "min_peak_corr",
                                                                   "min_monomer_distance_factor"))
head(bestComplexFeatures, n = 2)

}
