% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R,
%   R/getBestFeatureParameters.R
\name{getBestParameterStats}
\alias{getBestParameterStats}
\alias{getBestParameterStats}
\title{getBestParameterStats}
\usage{
getBestParameterStats(stats, FDR = 0.1)

getBestParameterStats(stats, FDR = 0.1)
}
\arguments{
\item{stats}{data.table with all stats from grid search}

\item{FDR}{numeric maximum FDR tat should be considered, default = 0.1}

\item{stats}{data.table with all stats from grid search}

\item{FDR}{numeric maximum FDR tat should be considered, default = 0.1}
}
\value{
data.table one row with best stats across grid search

data.table one row with best stats across grid search
}
\description{
getBestParameterStats.

getBestParameterStats.
}
