% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypothesesHelpers.R
\name{getDistance}
\alias{getDistance}
\title{getDistance}
\usage{
getDistance(A, B, all_A, all_B)
}
\arguments{
\item{A}{chracter vector}

\item{B}{character vector}

\item{all_A}{Numeric, length of chracter vector A.}

\item{all_B}{Numeric, length of chracter vector B.}
}
\value{
Numeric, distance of subunit overlap.
}
\description{
Calculate the distance between 2 complex hypotheses.
}

