% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypothesesHelpers.R
\name{getDistanceMatrix}
\alias{getDistanceMatrix}
\title{getDistanceMatrix}
\usage{
getDistanceMatrix(complexFeatures)
}
\arguments{
\item{complexFeatures}{data.table containing filtered complex feature results.
Must contain the following columns:
\itemize{
\item subunits_detected: Character, protein ids of complex subunits separated by ';'
}}
}
\value{
Numeric distance matrix. 0 = identical; 1 = one complex is a subset of the other;
2 or less = fractional overlap
}
\description{
calculate the pairwise distance between complex hypotheses.
}
\examples{
## load example data
complexHypotheses <- exampleComplexHypotheses
complexHypotheses <- complexHypotheses[, .(subunits_detected = paste0(protein_id, collapse = ";")),
                                       by = complex_id]
## Get the distance matrix
distMat <- getDistanceMatrix(complexHypotheses)

}

