% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformAnnotation.R
\name{getGenePepList}
\alias{getGenePepList}
\title{Filter peptides for having
at least one high correlating sibling peptide}
\usage{
getGenePepList(traces)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{cutoff}{Numeric between 0 and 1. Minimum correlation of a peptide
width any sibling peptide.}

\item{plot}{logical,wether to print SibPepCorr density plot to R console.
Deafult is \code{FALSE}.}

\item{PDF}{logical, wether to print SibPepCorr density plot to a PDF file.
Deafult is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if
'\code{PDF=TRUE}.PDF file is saved under name.pdf. Default is "maxCorrHist".}
}
\value{
Gene peptide list
}
\description{
Filter peptides for having
at least one high correlating sibling peptide
}
