% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{getIntensityMatrix}
\alias{getIntensityMatrix}
\title{getIntensityMatrix}
\usage{
getIntensityMatrix(traces)
}
\arguments{
\item{traces}{An object of type \code{traces}.}
}
\value{
A matrix with intensity values.
}
\description{
Get a matrix of intensity values from a traces object.
}
