% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{getIntensityMatrix}
\alias{getIntensityMatrix}
\title{Get a matrix of intensity values from a traces object.}
\usage{
getIntensityMatrix(traces.obj)
}
\arguments{
\item{traces.obj}{An object of type \code{traces.obj}.}
}
\value{
A matrix with intensity values.
}
\description{
Get a matrix of intensity values from a traces object.
}
