% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntensityMatrix.R
\name{getIntensityMatrix}
\alias{getIntensityMatrix}
\title{Get a matrix of fragment intensity values.}
\usage{
getIntensityMatrix(trace.obj)
}
\arguments{
\item{trace.obj}{An object of type \code{FragmentTraces}.}
}
\value{
A matrix of intensity values.
}
\description{
Get a matrix of fragment intensity values.
}

