% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMassDistribution.R
\name{getMassAssemblyChange}
\alias{getMassAssemblyChange}
\title{getMassAssemblyChange}
\usage{
getMassAssemblyChange(tracesList, design_matrix)
}
\arguments{
\item{tracesList}{An object of type traces.list including assembled mass 
annotation as produced by \code{annotateMassDistribution}.}

\item{design_matrix}{data.table, design matrix describing the architecture of the tracesList object.}
}
\description{
Estimates change of assembled vs. monomeric mass per id.
}
