% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromOpenSWATH.R
\name{importFromOpenSWATH}
\alias{importFromOpenSWATH}
\title{Import peptide profiles from an OpenSWATH experiment.}
\usage{
importFromOpenSWATH(data, annotation_table, rm_requantified = TRUE,
  rm_decoys = FALSE, MS1Quant = FALSE)
}
\arguments{
\item{data}{Quantitative MS data in form of OpenSWATH result file or R data.table.}

\item{annotation_table}{file or data.table containing columns `filename` and
`fraction_number` that map the file names (occuring in input table filename column)
`to a SEC elution fraction.}

\item{rm_requantified}{Whether requantified (noise) peak group quantities
(as indicated by m_score = 2) should be removed, defaults to TRUE.}

\item{rm_decoys}{Whether decoys should be removed, defaults to FALSE.}

\item{MS1Quant}{Wheather MS1 quantification should be used, defaults to FALSE.}
}
\value{
An object of class Traces containing
    "traces", "traces_type", "traces_annotation" and "fraction_annotation" list entries
    that can be processed with the herein contained functions.
}
\description{
This is a convenience function to directly import peptide profiles
 from an OpenSWATH experiment (after TRIC alignment). The peptide intensities are calculated by summing all charge
 states. Alternativley the MS1 signal can be used for quantification.
}
\examples{
  input_data <- exampleOpenSWATHinput
  annotation <- exampleFractionAnnotation
  traces <- importFromOpenSWATH(data = input_data,
                                annotation_table = annotation,
                                rm_requantified = T)
  summary(traces)
  
}
