% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexFeatureGridSearch.R,
%   R/estimateDecoyFDR.R
\name{performComplexGridSearch}
\alias{performComplexGridSearch}
\alias{performComplexGridSearch}
\title{Perform complex feature grid search}
\usage{
performComplexGridSearch(traces, complex_hypothesis, corrs = c(0.5, 0.75, 0.9,
  0.95), windows = c(8, 10, 12), smoothing = c(7, 9, 11),
  rt_heights = c(3, 4, 5), n_cores = 1)

performComplexGridSearch(traces, complex_hypothesis, corrs = c(0.5, 0.75, 0.9,
  0.95), windows = c(8, 10, 12), smoothing = c(7, 9, 11),
  rt_heights = c(3, 4, 5), n_cores = 1)
}
\arguments{
\item{traces}{traces object of type protein}

\item{corrs}{Numeric, vector of correlation_cutoff values to test}

\item{windows}{Numeric, vector of window_size values to test}

\item{smoothing}{Numeric, vector of smoothing_length values to test}

\item{rt_heights}{Numeric, vector of rt_height values to test
Default=\code{TRUE}}

\item{n_cores}{Numeric, number of cores to use
(if parallelized is \code{TRUE}) (default=1)}

\item{traces}{traces object of type protein}

\item{corrs}{Numeric, vector of correlation_cutoff values to test}

\item{windows}{Numeric, vector of window_size values to test}

\item{smoothing}{Numeric, vector of smoothing_length values to test}

\item{rt_heights}{Numeric, vector of rt_height values to test
Default=\code{TRUE}}

\item{n_cores}{Numeric, number of cores to use
(if parallelized is \code{TRUE}) (default=1)}
}
\value{
List with stats

List with stats
}
\description{
Perform complex feature grid search.

Perform complex feature grid search.
}
\examples{

# ## Load example data
proteinTraces <- exampleProteinTraces
complexHypotheses <- exampleComplexHypotheses
## Perform a small grid search for 2 parameter combinations
gridList <- performComplexGridSearch(traces = proteinTraces,
                         complex_hypothesis = complexHypotheses,
                         corrs = c(0.5, 0.9),
                         windows = 10,
                         smoothing = 7,
                         rt_heights = 4,
                         n_cores = 4)


exampleTraces <- exampleProteinTraces
complexHypotheses <- exampleComplexHypotheses
## Perform a small grid search for 2 parameter combinations
gridList <- performComplexGridSearch(traces = exampleTraces,
                         complex_hypothesis = complexHypotheses,
                         corrs = c(0.5, 0.9),
                         windows = 10,
                         smoothing = 7,
                         rt_heights = 4,
                         n_cores = 4)

}
