% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R
\name{performComplexGridSearch}
\alias{performComplexGridSearch}
\title{Perform complex feature grid search}
\usage{
performComplexGridSearch(traces, complex_hypothesis, calibration,
  corrs = c(0.5, 0.75, 0.9, 0.95), windows = c(8, 10, 12),
  smoothing = c(7, 9, 11), rt_heights = c(3, 4, 5), parallelized = TRUE,
  n_cores = 1)
}
\arguments{
\item{complex_features}{data.table containing filtered complex feature results.}
}
\value{
List with stats
}
\description{
Perform complex feature grid search.
}
