% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinFeatureGridSearch.R
\name{performProteinGridSearch}
\alias{performProteinGridSearch}
\title{Perform Protein feature grid search}
\usage{
performProteinGridSearch(traces, calibration, corrs = c(0.5, 0.75, 0.9, 0.95),
  windows = c(8, 10, 12), smoothing = c(7, 9, 11), rt_heights = c(3, 4,
  5), parallelized = TRUE, n_cores = 1)
}
\arguments{
\item{traces}{traces object of type peptide}

\item{calibration}{list of two functions for calibration}

\item{corrs}{numeric vector}

\item{windows}{numeric vector}

\item{smoothing}{numeric vector}

\item{rt_heights}{numeric vector}

\item{parallelized}{logical default=TRUE}

\item{n_cores}{numeric number of cores to use if parallelized (default=1)}
}
\value{
List with stats
}
\description{
Perform Protein feature grid search.
}
