% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{plot.traces}
\alias{plot.traces}
\title{Plot a traces.obj.}
\usage{
plot.traces(traces.obj, plot = TRUE, ledgend = TRUE)
}
\arguments{
\item{traces.obj}{An object of type \code{traces.obj}.}
}
\description{
Plot a traces.obj.
}

