% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{plot.tracesList}
\alias{plot.tracesList}
\title{Plot traces}
\usage{
\method{plot}{tracesList}(traces, design_matrix = NULL,
  collapse_conditions = FALSE, log = FALSE, legend = TRUE, PDF = FALSE,
  name = "Traces", plot = TRUE)
}
\arguments{
\item{log}{Logical, whether the intensities should be plotted in log scale. Default is \code{FALSE}.}

\item{legend}{Logical, whether a legend of the traces should be plotted. Should be set to \code{FALSE}
if many chromatograms are plotted. Default is \code{TRUE}.}

\item{PDF}{Logical, whether to plot to PDF. PDF file is saved in working directory. Default is \code{FALSE}.}

\item{name}{Character string with name of the plot, only used if \code{PDF=TRUE}.
PDF file is saved under name.pdf. Default is "Traces".}

\item{Object}{of class traces.}
}
\description{
Plot all chromatograms in a traces object. Most generic plotting function.
}

