% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexFeatureMethods.R
\name{plotComplexCompletenessScatter}
\alias{plotComplexCompletenessScatter}
\title{Scatter plot of complex completeness}
\usage{
plotComplexCompletenessScatter(
  complexFeatures,
  PDF = FALSE,
  name = "complex_completeness_scatter"
)
}
\arguments{
\item{complexFeatures}{data.table as reported by \code{\link[SECprofiler]{findComplexFeatures}}
or \code{\link[SECprofiler]{findProteinFeatures}}.}

\item{PDF}{Logical, whether to generate a PDF file with the summary plot. Default is \code{FALSE}.}

\item{name}{Character string specifying the name of the PDF file of the summary plot.
Only applicable if PDF=\code{TRUE}. Default is "feature_summary".}
}
\description{
Plot the number of observed co-eluting subunits over the number of subunits
in the hypothesis, colour-coded by the completeness. This is only taking into account targets.
If decoys are present they are removed using "DECOY" as a flag.
}
\examples{
## Load example complex feature finding results:
complexFeatures <- exampleComplexFeatures
## Run plotting function:
plotComplexCompletenessScatter(complexFeatures)
}
