% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDecoyFDR.R
\name{plotIdFDRspace}
\alias{plotIdFDRspace}
\title{Plot FDR gridsearch}
\usage{
plotIdFDRspace(grid_search_stats, FDR_cutoff = 0.1,
  colour_parameter = "completeness_cutoff", PDF = TRUE,
  name = "ID_FDR_plot.pdf")
}
\arguments{
\item{complex_features}{data.table containing filtered complex feature results.}
}
\value{
List with stats
}
\description{
Perform complex feature grid search.
}
