% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestFeatureParameters.R
\name{plotIdFDRspace}
\alias{plotIdFDRspace}
\title{Plot FDR gridsearch}
\usage{
plotIdFDRspace(grid_search_stats, level = "complex", id_level = "TP",
  FDR_cutoff = 0.1, colour_parameter = "min_feature_completeness",
  PDF = TRUE, name = "ID_FDR_plot.pdf")
}
\arguments{
\item{grid_search_stats}{Table of grid search statistics 
(obtained from \code{\link{estimateGridSearchDecoyFDR}}).}

\item{level}{Character string, either 'complex' or 'protein'.
Specifies which feature finding was performed. Defaults to 'complex'.}

\item{id_level}{Character string, either 'TP' or 'P'. 
Plot with true-positive numbers or all positives as y axis. Defaults to 'TP'}

\item{FDR_cutoff}{Numeric, the cutoff for the FDR (indicated by a vertical line in the plot).
Defaults to \code{0.1}.}

\item{colour_parameter}{Character string, Which parameter to color. Defaults to 'completeness_cutoff'}

\item{PDF}{Logical, wether to save the plot as a PDF file in working directory.
Defaults to \code{TRUE}.}

\item{name}{Character string, filename of the PDF output.}
}
\value{
Either a plot to the R console or a PDF file in the working directory.
}
\description{
Plot the result of a grid search depending on a specified parameter.
}
\examples{

## NOT RUN
gridStats # see function \\code{\\link{estimateGridSearchDecoyFDR}} to see how to generate this object.

## Plot the result of the grid search depending on the within feature correlation
plotIdFDRspace(gridStats, PDF = F, colour_parameter = "min_peak_corr")

}
