% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateExonSupport.R
\name{plotPeptideCluster}
\alias{plotPeptideCluster}
\title{Plot peptide clusters by relative sequence position}
\usage{
plotPeptideCluster(traces, protein)
}
\arguments{
\item{traces}{Object of class traces or tracesList.}

\item{protein}{Character string of protein_id to plot.}
}
\value{
plot
}
\description{
Plot peptide clusters by relative sequence position
}
