% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexFeatureMethods.R
\name{plotSummarizedComplexes}
\alias{plotSummarizedComplexes}
\title{Plot pie chart with complex hypothesis completeness}
\usage{
plotSummarizedComplexes(complexFeatures, hypotheses, protTraces, PDF = FALSE,
  name = "complex_completeness_pie")
}
\arguments{
\item{complexFeatures}{data.table as reported by \code{\link[SECprofiler]{findComplexFeatures}}
or \code{\link[SECprofiler]{findProteinFeatures}}.}

\item{hypotheses}{data.table containing complex hypotheses.
Should have the following columns:
\itemize{
\item complex_id: character strings, a unique id for every complex
\item complex_name: character strings, a unique name for every complex
\item protein_id: character strings, the protein id, e.g. Uniprot id
}}

\item{PDF}{Logical, whether to generate a PDF file with the summary plot. Default is \code{FALSE}.}

\item{name}{Character string specifying the name of the PDF file of the pie chart.
Only applicable if PDF=\code{TRUE}. Default is "complex_completeness_pie".}
}
\description{
Plot pie chart showing the number of complex hypotheses with co-elution features of certain completenes.
This is only taking into account targets. If decoys are present they are removed using "DECOY" as a flag.
}
\examples{
## Load example complex feature finding results:
complexFeatures <- exampleComplexFeatures
complexHypotheses <- exampleComplexHypotheses
proteinTraces <- exampleProteinTraces
## Run plotting function:
plotSummarizedComplexes(complexFeatures=complexFeatures,
                        hypotheses=complexHypotheses,
                        protTraces=proteinTraces)
}
