% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarizedMScoverage.R
\name{plotSummarizedMScoverage}
\alias{plotSummarizedMScoverage}
\title{plotSummarizedMScoverage.}
\usage{
plotSummarizedMScoverage(hypotheses, protTraces, PDF = TRUE)
}
\arguments{
\item{hypotheses}{data.table with complex hypotheses}

\item{protTraces}{traces object of type protein}

\item{PDF}{logical default = TRUE}
}
\value{
multiple plots summarizing the MS coverage
}
\description{
plotSummarizedMScoverage
}
