% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinQuantification.R
\name{proteinQuantification}
\alias{proteinQuantification}
\title{Quantify proteins based on topN peptides from peptide level traces.}
\usage{
proteinQuantification(Traces, topN = 2, keep_less = FALSE,
  remove.decoys = TRUE)
}
\arguments{
\item{Traces}{peptide level traces object as generated by import and upstream filtering functions.}

\item{topN}{integer number of highest intensity peptides to use for quantification.
The ID of the peptides used is maintained in the $trace_annotation table of the output
protein level traces object.}

\item{keep_less}{logical, whether proteins with less than topN peptides shall be 
kept (TRUE) or discarded (FALSE). Defaults to FALSE to enable label-free intensity comparisons
following the best-flyer hypothesis.}

\item{remove.decoys}{logical, whether to remove decoys}
}
\value{
Traces of $trace_type = "protein".
}
\description{
Quantify proteins based on topN peptides from peptide level traces
summing top N highest cumulative intensity peptides. $traces and $trace_annotation
 are collapsed to protein level. trace$id content switches from peptide_ids to protein_ids.
}

