% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinQuantification.R
\name{proteinQuantification}
\alias{proteinQuantification}
\alias{proteinQuantification.traces}
\alias{proteinQuantification.tracesList}
\title{Protein Quantification}
\usage{
proteinQuantification(traces, topN = 2, keep_less = FALSE,
  rm_decoys = TRUE, use_sibPepCorr = FALSE, use_repPepCorr = FALSE,
  full_intersect_only = FALSE)

\method{proteinQuantification}{traces}(traces, topN = 2, keep_less = FALSE,
  rm_decoys = TRUE, use_sibPepCorr = FALSE, use_repPepCorr = FALSE,
  full_intersect_only = FALSE)

\method{proteinQuantification}{tracesList}(traces, topN = 2,
  keep_less = FALSE, rm_decoys = TRUE, use_sibPepCorr = FALSE,
  use_repPepCorr = FALSE, full_intersect_only = FALSE)
}
\arguments{
\item{traces}{An object of type traces, trace_type must be peptide.}

\item{topN}{Numeric integer, specifying the number of peptides to sum for protein quantification. Default is 2.}

\item{keep_less}{Logical, specifying whether proteins with less than topN peptides
should be kept in the data (This may result in some protein intensities being calculated
as the sum of fewer peptide intensities than others. Only use withh caution.), default is \code{FALSE}.}

\item{rm_decoys}{Logical, specifying whether decoys should be kept.
The decoys have only limited use on the protein level, default is \code{TRUE}.}

\item{use_sibPepCorr}{Logical, specify whether the sibling peptide correlation should be considered for selecting
the top N peptides, default is \code{FALSE}.}

\item{use_repPepCorr}{Logical, specify whether the replicate peptide correlation should be considered for selecting
the top N peptides, default is \code{FALSE}. This argument is only used for tracesList objects.}

\item{full_intersect_only}{Logical, specify if only peptides detected across all traces in tracesList object should
should be used for quantification. Be aware that this might significantly reduce the number of quantifyable proteins.
This argument is only used for tracesList objects.}
}
\value{
An object of type traces, trace_type is protein.
}
\description{
Calculate protein quantities basen on the topN peptide intensities.
}
\section{Methods (by class)}{
\itemize{
\item \code{traces}: Protein Quantification for single traces object

\item \code{tracesList}: Protein Quantification multiple traces ojects.
Uses only peptides present in all traces objects for quantification.
}}

\examples{
## Load example data
pepTraces <- examplePeptideTracesFiltered

## Sum the intensities of the top 2 peptides to get protein intensities
protTraces <- proteinQuantification(pepTraces,
                                     topN = 2)
## Check the result
summary(pepTraces)
summary(protTraces)

# ProteinTraces annotation
head(protTraces$trace_annotation)
# The protein_id column from the peptide traces object becomes the new id column of the protein traces object.
# The last 2 columns indicate how many peptides could be observed and which were summed for quantification.

}
