% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracesMethods.R
\name{subset.traces}
\alias{subset.traces}
\title{Subset a traces.obj by trace_ids or fraction_ids.}
\usage{
\method{subset}{traces}(traces.obj, trace_ids = NULL, fraction_ids = NULL)
}
\arguments{
\item{traces.obj}{An object of type \code{traces.obj}.}

\item{trace_ids}{A character vector specifying the trace identifiers
for subsetting \code{trace_ids}.}

\item{fraction_ids}{A numeric vector specifying the fraction identifiers
for subsetting \code{fragment_ids}.}
}
\value{
traces.obj An object of type \code{traces.obj}.
}
\description{
Subset a traces.obj by trace_ids or fraction_ids.
}

